+++ paper/paper.bib
@@ -20,8 +20,7 @@
 	year = {2019},
 	month = {Aug},
 	publisher = {American Physical Society},
-	doi = {10.1103/PhysRevE.100.022204},
-	url = {https://link.aps.org/doi/10.1103/PhysRevE.100.022204}
+	doi = {10.1103/PhysRevE.100.022204}
 }
 
 @book{Parker_1989,
@@ -30,7 +29,7 @@
 	year = {1989},
 	isbn = {0-387-96689-7},
 	publisher = {Springer-Verlag New York, Inc.},
-	address = {New York, NY, USA},
+	address = {New York, NY, USA}
 } 
 
 @article{wiggins_role_2016,
@@ -41,7 +40,8 @@
 	author = {Wiggins, Stephen},
 	month = nov,
 	year = {2016},
-	pages = {621--638}
+	pages = {621--638},
+	doi = {10.1134/S1560354716060034}
 }
 
 @article{Pollak_1980,
