### 2019-08-22 #1670 @Shibabrat
Hi @whedon

I'm a collaborator of @WyLyu and have checked the article proof. The current paper sub-directory has the updated paper.md and paper.bib files. However, the two figures we have compiled in the pdf from paper.md doesn't show up in the article proof.

Can you kindly recompile and let us know?

Thanks
