### 2019-08-22 #1670 @kthyng
@kyleniemeyer Sure, I'll be the editor.

@Shibabrat It looks like the addresses to the pdfs are not correct — they are in the same directory as the paper, not in the tests directory.
