### 2019-08-22 ed75807
Copied figure files to paper
