### 2019-08-22 #1670 @Shibabrat
@kthyng Thank you for agreeing to be the editor and for pointing out the mistake. I have updated the location of figure files and the link to the archived version on Zenodo.

We are deliberating the reviewer names and will reply soon. Is it possible to request someone who is not in the list and uses related methods?
