### 2019-08-22 #1670 @kyleniemeyer
@Shibabrat please also remove the "software repository" and "archive" links at the beginning of the paper—these are included in the JOSS article template on the side
