### 2019-08-22 3a2387f
Removing links to github and zenodo
