### 2019-08-22 #1670 @Shibabrat
Hi @kyleniemeyer thanks, done!

@kthyng here are the usernames of potential reviewers: jochym, JoshKarpel, mdoucet, mikapfl, and or dpsanders.

Btw, the JOSS paper template is such a neat design. Finally, a journal that uses Tufte-handout style. Kudos to the team who decided on this layout.
