### 2019-08-22 #1670 @kthyng
@mikapfl and @jochym Are you interested in reviewing this JOSS submission? We ask for reviews within about 3 weeks and you can find more information here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
