### 2019-08-27 #1684 @whedon
[REVIEW]: UPOsHam:  A Python package for computing unstable periodic orbits in two degrees of freedom Hamiltonian systems
**Submitting author:** @WyLyu (<a href="http://orcid.org/0000-0003-2570-9879">WenYang Lyu</a>)
**Repository:** <a href="https://github.com/WyLyu/UPOsHam" target ="_blank">https://github.com/WyLyu/UPOsHam</a>
**Version:** v0.1.0
