### 2019-08-27 WyLyu/UPOsHam#1 @JoshKarpel
Add installation instructions
It is not immediately obvious how to install the package. Since it appears to mostly be a collection of scripts (and not a package per-se), I would guess that simply downloading any of the notebooks would work, but a quick look at https://github.com/WyLyu/UPOsHam/blob/master/run_comparison_coupled.py indicates that a significant number of extra dependencies are required, and that the scripts tend depend on each other.

There should be full installation instructions (probably somewhere in the README), including (minimally) how to install the Python dependencies (a `requirements.txt` or similar), and how to get the scripts themselves onto my computer in a consistent way (`git clone`? A GitHub release tarball?).

(Part of https://github.com/openjournals/joss-reviews/issues/1684)
