### 2019-08-27 WyLyu/UPOsHam#2 @JoshKarpel
Use docstrings instead of block comments for function documentation
Many of the functions in the support scripts (like https://github.com/WyLyu/UPOsHam/blob/master/coupled_diffcorr.py) have function documentation in block comments. These should be converted to docstrings (see https://docs.python-guide.org/writing/documentation/#code-documentation-advice for an opinionated-but-widely-accepted primer).

This is important because proper docstrings are integrated into Python itself and can be inspected at runtime or exposed to external documentation tools like Sphinx.

(Part of https://github.com/openjournals/joss-reviews/issues/1684)
