### 2019-08-27 WyLyu/UPOsHam#3 @JoshKarpel
Follow standard Python code style guide
Much of the code does not adhere to pep-8 (https://www.python.org/dev/peps/pep-0008/). This is not a blocking problem for the review, but it will look odd to Python programmers. Best practice is to mostly follow pep-8 unless you have good reason not to, and I would recommend that you do so.

Since there are a lot of things to change if you decide to do this, I recommend using https://pypi.org/project/black/ to quickly get the code into mostly-good shape.

(Part of https://github.com/openjournals/joss-reviews/issues/1684)
