### 2019-08-27 WyLyu/UPOsHam#4 @JoshKarpel
Correctness tests?
I see that https://github.com/WyLyu/UPOsHam/blob/master/run_alltests.ipynb provides a way to run all of the comparison scripts and look at their results. However, as a non-expert, it isn't obvious to me whether I've gotten the correct results at the end.

Are there solutions to these problems simple enough that you can write an automated test that asserts that your code is behaving correctly (i.e., solve analytically for some simple, contrived inputs, then run the code and compare)? Even if there aren't simple solutions, can you write tests that check that any invariants are indeed invariant? What I'm trying to get at is: some way to check that the scripts are behaving as expected beyond the code itself running successfully.

If you can write actual automated tests, I recommend using https://docs.pytest.org/en/latest/ to run them.

(Part of https://github.com/openjournals/joss-reviews/issues/1684)
