### 2019-08-27 WyLyu/UPOsHam#5 @JoshKarpel
Add "community guidelines"
Part of the JOSS requirements is to have clear guidelines for how the community can interact with the project. There need to be statements in the README, or pointers from the README to other places, where people can learn about how to:

- Contribute to the software
- Report issues or problems with the software
- Seek support

For a jump-start on these, if you go into the Insights > Community from the bar just under the repository name, you'll find some helpful templates created by GitHub.

(Part of https://github.com/openjournals/joss-reviews/issues/1684)
