### 2019-08-29 #1684 @Shibabrat
@kthyng I'm willing to explain the design decision of this package because it is related to the theory and applications of this method. Should I create a separate issue or briefly describe it here? Not sure about the etiquette at this point of review period.

Thanks
