### 2019-08-30 14f41b1
Adding installation instructions
