+++ requirements.txt
@@ -1,0 +1,4 @@
+scipy==1.1.0
+numpy==1.14.3
+matplotlib==2.2.3
+ipython==7.7.0
