### 2019-08-30 da1864d
Adding installation instructions
