### 2019-08-30 5751061
Adding installation instructions
