### 2019-08-30 720ec37
Adding installation instructions
