+++ README.Rmd
@@ -21,19 +21,7 @@ This Python package is a collection of three methods for computing unstable peri
 
 ## Installation
 
-Clone/download the git repository and check whether the following modules (also available in [requirements.txt](https://github.com/WyLyu/UPOsHam/tree/master/requirements.txt)) are installed using conda/pip
-
-scipy==1.1.0
-
-numpy==1.14.3
-
-matplotlib==2.2.3
-
-ipython==7.7.0
-
-
-
-
+Clone/download the git repository and check whether the following modules (also available in [requirements.txt](https://github.com/WyLyu/UPOsHam/tree/master/requirements.txt)) are installed using conda/pip.
 
 ## Usage
 
