### 2019-08-30 9740bfd
Adding installation instructions
