### 2019-08-30 7bd139e
Adding installation instructions
