### 2019-08-30 WyLyu/UPOsHam#1 @Shibabrat
We have added installation instruction and module version numbers to requirements.
