### 2019-08-30 82326c9
Adding installation instructions
