### 2019-08-30 WyLyu/UPOsHam#1 @JoshKarpel
Could you also add an example pip install command, probably something like "`cd` into the directory that the `git clone` created and run `pip install -r requirements.txt`"? Just to help people who don't know much about Python get it working.

LGTM otherwise!
