### 2019-08-30 967f6a5
Adding installation instructions
