### 2019-08-30 #1684 @kthyng
@Shibabrat Why don't you give a brief description of the design decision here just to help deciding how to move forward, for now. One design suggestion is for you to change this to a library so that a user doesn't have to copy and paste to use it.
