### 2019-08-31 #1684 @Shibabrat
Ok.

We have seen that researchers working in Hamiltonian dynamics have their favorite method to compute unstable periodic orbits. These methods are popular in chemical reactions, celestial mechanics, and engineering dynamics and always end up in appendices of the papers. Some of these papers are also in our references.

However, for anyone new or outside of the aforementioned communities, it becomes a slow tedious process to go through the code (even if available) and to adapt it for a specific problem. We think an illustration style set of transparent scripts that has minimal dependencies for installation, is easy to change and modify can help with learning the underlying methods.

Next, examples in the package are also in the form that is being used for ongoing work in our group on chemical reaction dynamics and hence gives a direct way to reproduce the results.
