### 2019-09-04 #1684 @kthyng
@Shibabrat Your argument that this software is useful because it demonstrates going through some physics problems is convincing. However, it would be much better if your software was a proper python package that can be used in a typical manner in order to meet your goals. Specifically, can you change your setup so that another researcher using it can avoid copying and pasting bits of code into a new project, and instead import the full package and use reusable functions?

@JoshKarpel We think that the submission can be substantially improved, as per my first comment here, though your concern with the one-off modifications is ok — the point of the requirement is that the software have reuse capacity, which this appears to have. Still, changing this to a package will also help to address your concern.
