### 2019-09-04 #1684 @JoshKarpel
@kthyng Sounds good to me. I think if it is not possible to make it more package-y, it needs to be very, very clear how and where to copy-paste-modify. But I would prefer a more package-y solution.

@Shibabrat  @WyLyu some semi-concrete thoughts on how to move toward a package-based design...

So, my understanding is that there are three solution methods, which you use to solve three problems, leading to nine examples. Is it possible to decouple the code for each solution method from the problem it is solving, such that the same solution method code could be applied to any of the problems with only top-level configuration? As an incidental benefit, this would reduce the amount of code duplication in the project...

As a very concrete example, think about how the differential equation solvers in SciPy work ( https://docs.scipy.org/doc/scipy/reference/generated/scipy.integrate.solve_ivp.html#scipy.integrate.solve_ivp). There's a high-level solver which can apply different methods (set via the `method` argument) to whatever the user passes in. Many different differential equations in many different forms can be solved by the solver (although the user may need to select the correct method for their given problem, perhaps equivalent to selecting a solution method in your code?).
