### 2019-09-10 83cd7ef
merge functions used for turning point methods
