+++ turningpoint_UPOsHam2dof.py
@@ -1,0 +1,776 @@
+# -*- coding: utf-8 -*-
+"""
+Created on Mon Sep  9 17:49:31 2019
+
+@author: wl16298
+"""
+
+import numpy as np
+import matplotlib.pyplot as plt
+from scipy.integrate import odeint,quad,trapz,solve_ivp, ode
+import math
+from IPython.display import Image # for Notebook
+from IPython.core.display import HTML
+from mpl_toolkits.mplot3d import Axes3D
+from numpy import linalg as LA
+import scipy.linalg as linalg
+from scipy.optimize import fsolve
+import time
+from functools import partial
+from scipy import optimize
+import matplotlib as mpl
+
+from pylab import rcParams
+mpl.rcParams['mathtext.fontset'] = 'cm'
+mpl.rcParams['mathtext.rm'] = 'serif'
+
+
+#%%
+def func_vec_field_eq_pt(model,x,par):
+    """ vecor field(same as pxdot, pydot), used to find the equilibrium points of the system.
+    """
+    if model == 'uncoupled':
+        dVdx = -par[3]*x[0]+par[4]*(x[0])**3
+        dVdy = par[5]*x[1]
+    elif model == 'coupled':
+        dVdx = (-par[3]+par[6])*x[0]+par[4]*(x[0])**3-par[6]*x[1]
+        dVdy = (par[5]+par[6])*x[1]-par[6]*x[0]
+    elif model == 'deleonberne':
+        dVdx = -2*par[3]*par[4]*math.e**(-par[4]*x[0])*(math.e**(-par[4]*x[0]) - 1) - 4*par[5]*par[4]*x[1]**2*(x[1]**2 - 1)*math.e**(-par[5]*par[4]*x[0])
+        dVdy = 8*x[1]*(2*x[1]**2 - 1)*math.e**(-par[5]*par[4]*x[0])
+    else:
+        print("The model you are chosen does not exist, enter the definition of Hamilton's equations")
+    F = [-dVdx, -dVdy]
+    return F
+    
+
+#%%
+def get_eq_pts(eqNum,model, par):
+    """
+    GET_EQ_PTS_BP solves the saddle center equilibrium point for a system with
+    KE + PE. 
+     H = 1/2*px^2+omega/2*py^2 -(1/2)*alpha*x^2+1/4*beta*x^4+ omega/2y^2 with alpha > 0
+    --------------------------------------------------------------------------
+       Uncouled potential energy surface notations:
+    
+               Well (stable, EQNUM = 2)    
+    
+                   Saddle (EQNUM=1)
+    
+               Well (stable, EQNUM = 3)    
+    
+    --------------------------------------------------------------------------
+       
+    
+    """
+    #fix the equilibrium point numbering convention here and make a
+    #starting guess at the solution
+    if 	eqNum == 1:
+        
+        x0 = [0, 0]                  # EQNUM = 1, saddle  
+    elif 	eqNum == 2: 
+        if model == 'uncoupled':
+            eqPt = [+math.sqrt(par[3]/par[4]),0]    # EQNUM = 2, stable
+            return eqPt
+        elif model == 'coupled':
+            eqPt = [+math.sqrt(par[3]-par[6]/par[4]),0] # EQNUM = 2, stable
+            return eqPt
+        elif model== 'deleonberne':
+            eqPt = [0, 1/math.sqrt(2)]    # EQNUM = 2, stable
+            return eqPt
+        else:
+            print("The model you are chosen does not exist, enter the definition of Hamilton's equations")
+    elif 	eqNum == 3:
+        if model == 'uncoupled':
[…]
