+++ deleonberne_diffcorr.py
@@ -1,909 +1,0 @@
-"""
-Created on Wed Jul 10 19:15:22 2019
-
-@author: wl16298
-"""
-import numpy as np
-import matplotlib.pyplot as plt
-from scipy.integrate import odeint,quad,trapz,solve_ivp, ode
-import math
-from IPython.display import Image # for Notebook
-from IPython.core.display import HTML
-from mpl_toolkits.mplot3d import Axes3D
-from numpy import linalg as LA
-import scipy.linalg as linalg
-from scipy.optimize import fsolve
-import time
-from functools import partial
-
-import matplotlib as mpl
-
-from pylab import rcParams
-mpl.rcParams['mathtext.fontset'] = 'cm'
-mpl.rcParams['mathtext.rm'] = 'serif'
-
-#%%
-def get_eq_pts_deleonberne(eqNum, parameters):
-    #GET_EQ_PTS_BP solves the saddle center equilibrium point for a system with
-    #KE + PE. 
-    #--------------------------------------------------------------------------
-    #   DeLeon-Berne potential energy surface notations:
-    #
-    #           Well (stable, EQNUM = 2)    
-    #
-    #               Saddle (EQNUM=1)
-    #
-    #           Well (stable, EQNUM = 3)    
-    #
-    #--------------------------------------------------------------------------
-    #   
-    
-    
-    #fix the equilibrium point numbering convention here and make a
-    #starting guess at the solution
-    if 	eqNum == 1:
-        
-        x0 = [0, 0];                  # EQNUM = 1, saddle  
-    elif 	eqNum == 2: 
-        eqPt = [0, 1/math.sqrt(2)]    # EQNUM = 2, stable
-        return eqPt
-    elif 	eqNum == 3:
-        eqPt = [0, -1/math.sqrt(2)]   # EQNUM = 3, stable
-        return eqPt
-    
-    
-    # F(xEq) = 0 at the equilibrium point, solve using in-built function
- 
-    def func_vec_field_eq_pt(x,par):
-        x1, x2 = x
-        # constant parameters for the surface
-        # enter the definition of the potential energy function
-        dVdx = - ( 2*par[3]*par[4]*(math.e**(-par[4]*(x1)))*(math.e**(-par[4]*(x1)) - 1) + 4*par[5]*par[4]*((x2)**2)*((x2)**2 - 1)*(math.e**(-par[5]*par[4]*(x1)))) 
-        
-        dVdy = 8*(x2)*(2*(x2)**2 - 1)*math.e**(-par[5]*par[4]*(x1))
-    
-        F = [-dVdx, -dVdy]
-        return F
-    
-    eqPt = fsolve(func_vec_field_eq_pt,x0, fprime=None,args=(parameters,)); # Call solver
-    return eqPt
-
-
-#%%# enter the definition of the potential energy function
-def get_potential_energy(x,y,par):        
-            
-    pot_energy = par[3]*( 1 - math.e**(-par[4]*x) )**2 + 4*y**2*(y**2 - 1)*math.e**(-par[5]*par[4]*x) + par[2]
-                
-    return pot_energy
-
-
-#%%
-def get_total_energy_deleonberne(orbit, parameters):
-
-#   get_total_energy_deleonberne computes the total energy of an input orbit
-#   (represented as M x N with M time steps and N = 4, dimension of phase
-#   space for the model) for the 2 DoF DeLeon-Berne potential.
-# 
-#   Orbit can be different initial conditions for the periodic orbit of
[…]
