+++ run_tpcd_POfam_uncoupled.py
@@ -1,231 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-Created on Tue Jul 30 10:02:48 2019
-
-@author: wl16298
-"""
-
-"""Plot peridoic orbits with 5 different energies using the new method"""
-
-# For the uncoupled problem
-
-import numpy as np
-import matplotlib.pyplot as plt
-from scipy.integrate import odeint,quad,trapz,solve_ivp
-import math
-from IPython.display import Image # for Notebook
-from IPython.core.display import HTML
-from mpl_toolkits.mplot3d import Axes3D
-from numpy import linalg as LA
-import scipy.linalg as linalg
-from scipy.optimize import fsolve
-import time
-from functools import partial
-import uncoupled_tpcd ### import module xxx where xxx is the name of the python file xxx.py 
-from mpl_toolkits.mplot3d import Axes3D
-import matplotlib as mpl
-from matplotlib import cm
-from pylab import rcParams
-mpl.rcParams['mathtext.fontset'] = 'cm'
-mpl.rcParams['mathtext.rm'] = 'serif'
-#%matplotlib
-
-#%%
-alpha = 1.0
-beta = 1.0
-omega = 1.0
-EPSILON_S = 0.0 #Energy of the saddle
-parameters = np.array([1,omega, EPSILON_S, alpha, beta,omega]);
-eqNum = 1;  
-eqPt = uncoupled_tpcd.get_eq_pts_uncoupled(eqNum, parameters)
-"""Initial Condition for the true periodic orbit
-H = T + V where T is the kinetic energy and V is the potential energy. In our example, 
-the potential energy V = -0.5*alpha*x**2+0.25*beta*x**4+0.5*omega*y**2.
-If we fix x, then y = +- math.sqrt((V +0.5*alpha*x**2-0.25*beta*x**4)/(0.5*omega) ) so that
-the point starts at the potential energy surface V.
-"""
-
-#%%
-e=0.01
-"""Trial initial Condition s.t. one initial condition is on the LHS of the UPO and the other one is on the RHS of the UPO"""
-state0_2 = [-0.1 , -math.sqrt(2*e+0.1**2-0.5*0.1**4),0.0,0.0]
-state0_3 = [0.11 , -math.sqrt(2*e+0.11**2-0.5*0.11**4),0.0,0.0]
-n=4
-n_turn=1
-deltaE = e-parameters[2]
-po_fam_file = open("1111x0_tpcd_deltaE%s_uncoupled.txt" %(deltaE),'a+')
-[x0po_1, T_1,energyPO_1] = uncoupled_tpcd.tpcd_uncoupled(state0_2,state0_3 ,parameters,e,n,n_turn,po_fam_file) ; 
-
-po_fam_file.close()
-
-#%%
-e=0.1
-"""Trial initial Condition s.t. one initial condition is on the LHS of the UPO and the other one is on the RHS of the UPO"""
-state0_2 = [-0.1 , -math.sqrt(2*e+0.1**2-0.5*0.1**4),0.0,0.0]
-state0_3 = [0.11 , -math.sqrt(2*e+0.11**2-0.5*0.11**4),0.0,0.0]
-n=4
-n_turn=1
-deltaE = e-parameters[2]
[…]
