+++ run_tpcd_POfam_deleonberne.py
@@ -1,216 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-Created on Tue Jul 30 10:02:48 2019
-
-@author: wl16298
-"""
-
-"""Plot peridoic orbits with 5 different energies using the new method"""
-
-# For deleonberne problem
-
-import numpy as np
-import matplotlib.pyplot as plt
-from scipy.integrate import odeint,quad,trapz,solve_ivp
-import math
-from IPython.display import Image # for Notebook
-from IPython.core.display import HTML
-from mpl_toolkits.mplot3d import Axes3D
-from numpy import linalg as LA
-import scipy.linalg as linalg
-from scipy.optimize import fsolve
-import time
-from functools import partial
-import deleonberne_tpcd ### import module xxx where xxx is the name of the python file xxx.py 
-from mpl_toolkits.mplot3d import Axes3D
-import matplotlib as mpl
-from matplotlib import cm
-from pylab import rcParams
-mpl.rcParams['mathtext.fontset'] = 'cm'
-mpl.rcParams['mathtext.rm'] = 'serif'
-#%matplotlib
-from scipy import optimize
-#%% Setting up parameters and global variables
-N = 4;          # dimension of phase space
-MASS_A = 8.0; MASS_B = 8.0; # De Leon, Marston (1989)
-EPSILON_S = 1.0;
-D_X = 10.0;
-ALPHA = 1.00;
-LAMBDA = 1.5;
-parameters = np.array([MASS_A, MASS_B, EPSILON_S, D_X, LAMBDA, ALPHA]);
-eqNum = 1;  
-eqPt = deleonberne_tpcd.get_eq_pts_deleonberne(eqNum, parameters)
-#%%
-# e is the total energy, n is the number of intervals we want to divide, n_turn is the nth turning point we want to choose.
-e=1.10
-n=4
-n_turn = 1
-deltaE = e-parameters[2]  # In this case deltaE is 0.10
-"""Trial initial Condition s.t. one initial condition is on the LHS of the UPO and the other one is on the RHS of the UPO"""
-f1 = partial(deleonberne_tpcd.get_x,y=0.06,V=e,par=parameters)
-x0_2 = optimize.newton(f1,-0.15)
-state0_2 = [x0_2,0.06,0.0,0.0]
-f2 = partial(deleonberne_tpcd.get_x,y=-0.05,V=e,par=parameters)
-x0_3 = optimize.newton(f2,-0.15)
-state0_3 = [x0_3, -0.05,0.0,0.0]
-
-po_fam_file = open("1111x0_tpcd_deltaE%s_deleonberne.txt" %(0.10),'a+')
-[x0po_1, T_1,energyPO_1] = deleonberne_tpcd.tpcd_deleonberne(state0_3,state0_2 ,parameters,e,n,n_turn,po_fam_file) ; 
-
-po_fam_file.close()
-#%%
-# e is the total energy, n is the number of intervals we want to divide, n_turn is the nth turning point we want to choose.
-e=2.0
-n=4
-n_turn = 1
-deltaE = e-parameters[2]
-"""Trial initial Condition s.t. one initial condition is on the LHS of the UPO and the other one is on the RHS of the UPO"""
[…]
