### 2019-09-11 8b4949c
Update tpcd method
