+++ run_tpcd_POfam_coupled.py
@@ -1,0 +1,261 @@
+# -*- coding: utf-8 -*-
+"""
+Created on Wed Sep 11 17:38:57 2019
+
+@author: wl16298
+"""
+
+import numpy as np
+import matplotlib.pyplot as plt
+from scipy.integrate import odeint,quad,trapz,solve_ivp
+import math
+from IPython.display import Image # for Notebook
+from IPython.core.display import HTML
+from mpl_toolkits.mplot3d import Axes3D
+from numpy import linalg as LA
+import scipy.linalg as linalg
+from scipy.optimize import fsolve
+from scipy import optimize
+import time
+from functools import partial
+import turningpoint_UPOsHam2dof ### import module xxx where xxx is the name of the python file xxx.py 
+from mpl_toolkits.mplot3d import Axes3D
+import matplotlib as mpl
+from matplotlib import cm
+from pylab import rcParams
+mpl.rcParams['mathtext.fontset'] = 'cm'
+mpl.rcParams['mathtext.rm'] = 'serif'
+#%matplotlib
+
+#%% Setting up parameters and global variables
+N = 4          # dimension of phase space
+omega=1.00
+EPSILON_S = 0.0 #Energy of the saddle
+alpha = 1.00
+beta = 1.00
+epsilon= 1e-1
+parameters = np.array([1,omega, EPSILON_S, alpha, beta,omega,epsilon])
+eqNum = 1  
+model = 'coupled'
+eqPt = turningpoint_UPOsHam2dof.get_eq_pts(eqNum,model, parameters)
+
+#%%
+# e is the total energy, n is the number of intervals we want to divide, n_turn is the nth turning point we want to choose.
+e=0.01
+n=4
+n_turn = 2
+deltaE = e-parameters[2]
+"""Trial initial Condition s.t. one initial condition is on the LHS of the UPO and the other one is on the RHS of the UPO"""
+x=0.0
+f2 = lambda y: turningpoint_UPOsHam2dof.get_coordinate(model,0.0,y,e,parameters)
+yanalytic = math.sqrt(2/(parameters[1]+parameters[6]))*(-math.sqrt( e +0.5*parameters[3]* x**2- 0.25*parameters[4]*x**4 -0.5*parameters[6]* x**2 + (parameters[6]*x)**2/(2*(parameters[1] +parameters[6]) )) +parameters[6]/(math.sqrt(2*(parameters[1]+parameters[6])) )*x ) #coupled
+state0_2 = [0.0,optimize.newton(f2,yanalytic),0.0,0.0]
+
+x=0.05
+f3 = lambda y: turningpoint_UPOsHam2dof.get_coordinate(model,x,y,e,parameters)
+yanalytic = math.sqrt(2/(parameters[1]+parameters[6]))*(-math.sqrt( e +0.5*parameters[3]* x**2- 0.25*parameters[4]*x**4 -0.5*parameters[6]* x**2 + (parameters[6]*x)**2/(2*(parameters[1] +parameters[6]) )) +parameters[6]/(math.sqrt(2*(parameters[1]+parameters[6])) )*x ) #coupled
+state0_3 = [x, optimize.newton(f3,yanalytic),0.0,0.0]
+
+
+po_fam_file = open("1111x0_tpcd_deltaE%s_coupled.txt" %(deltaE),'a+')
[…]
