### 2019-09-11 2b4e9d9
Update tpcd method
