+++ run_tpcd_POfam_deleonberne.py
@@ -28,7 +28,8 @@ mpl.rcParams['mathtext.rm'] = 'serif'
 from scipy import optimize
 #%% Setting up parameters and global variables
 N = 4         # dimension of phase space
-MASS_A = 8.0; MASS_B = 8.0 # De Leon, Marston (1989)
+MASS_A = 8.0
+MASS_B = 8.0 # De Leon, Marston (1989)
 EPSILON_S = 1.0
 D_X = 10.0
 ALPHA = 1.00
@@ -52,7 +53,7 @@ x0_3 = optimize.newton(f2,-0.15)
 state0_3 = [x0_3, -0.05,0.0,0.0]
 
 po_fam_file = open("1111x0_tpcd_deltaE%s_deleonberne.txt" %(0.10),'a+')
-[x0po_1, T_1,energyPO_1] = turningpoint_UPOsHam2dof.turningPoint_configdiff(model,state0_3,state0_2 ,parameters,e,n,n_turn,po_fam_file) ; 
+[x0po_1, T_1,energyPO_1] = turningpoint_UPOsHam2dof.turningPoint_configdiff(model,state0_3,state0_2 ,parameters,e,n,n_turn,po_fam_file)  
 
 po_fam_file.close()
 #%%
@@ -70,7 +71,7 @@ x0_3 = optimize.newton(f2,-0.15)
 state0_3 = [x0_3, -0.05,0.0,0.0]
 
 po_fam_file = open("1111x0_tpcd_deltaE%s_deleonberne.txt" %(deltaE),'a+')
-[x0po_2, T_2,energyPO_2] = turningpoint_UPOsHam2dof.turningPoint_configdiff(model,state0_3,state0_2 ,parameters,e,n,n_turn,po_fam_file) ; 
+[x0po_2, T_2,energyPO_2] = turningpoint_UPOsHam2dof.turningPoint_configdiff(model,state0_3,state0_2 ,parameters,e,n,n_turn,po_fam_file)  
 
 po_fam_file.close()
 #%%
@@ -88,7 +89,7 @@ x0_3 = optimize.newton(f2,-0.15)
 state0_3 = [x0_3, -0.05,0.0,0.0]
 
 po_fam_file = open("1111x0_tpcd_deltaE%s_deleonberne.txt" %(deltaE),'a+')
-[x0po_3, T_3,energyPO_3] = turningpoint_UPOsHam2dof.turningPoint_configdiff(model,state0_3,state0_2 ,parameters,e,n,n_turn,po_fam_file) ; 
+[x0po_3, T_3,energyPO_3] = turningpoint_UPOsHam2dof.turningPoint_configdiff(model,state0_3,state0_2 ,parameters,e,n,n_turn,po_fam_file)  
 
 po_fam_file.close()
 #%%
@@ -106,35 +107,35 @@ x0_3 = optimize.newton(f2,-0.15)
 state0_3 = [x0_3, -0.05,0.0,0.0]
 
 po_fam_file = open("1111x0_tpcd_deltaE%s_deleonberne.txt" %(deltaE),'a+')
-[x0po_4, T_4,energyPO_4] = turningpoint_UPOsHam2dof.turningPoint_configdiff(model,state0_3,state0_2 ,parameters,e,n,n_turn,po_fam_file) ; 
+[x0po_4, T_4,energyPO_4] = turningpoint_UPOsHam2dof.turningPoint_configdiff(model,state0_3,state0_2 ,parameters,e,n,n_turn,po_fam_file)  
 
 po_fam_file.close()
 
 #%% Load Data
 deltaE = 0.10
-po_fam_file = open("1111x0_tpcd_deltaE%s_deleonberne.txt" %(deltaE),'a+');
[…]
