+++ run_turningpoint_POfam_uncoupled.py
@@ -20,9 +20,10 @@ from mpl_toolkits.mplot3d import Axes3D
 from numpy import linalg as LA
 import scipy.linalg as linalg
 from scipy.optimize import fsolve
+from scipy import optimize
 import time
 from functools import partial
-import uncoupled_turningpoint ### import module xxx where xxx is the name of the python file xxx.py 
+import turningpoint_UPOsHam2dof ### import module xxx where xxx is the name of the python file xxx.py 
 from mpl_toolkits.mplot3d import Axes3D
 import matplotlib as mpl
 from matplotlib import cm
@@ -36,9 +37,10 @@ alpha = 1.0
 beta = 1.0
 omega = 1.0
 EPSILON_S = 0.0 #Energy of the saddle
-parameters = np.array([1,omega, EPSILON_S, alpha, beta,omega]);
-eqNum = 1;  
-eqPt = uncoupled_turningpoint.get_eq_pts_uncoupled(eqNum, parameters)
+parameters = np.array([1,omega, EPSILON_S, alpha, beta,omega])
+eqNum = 1 
+model = 'uncoupled'
+eqPt = turningpoint_UPOsHam2dof.get_eq_pts(eqNum,model, parameters)
 """Initial Condition for the true periodic orbit
 H = T + V where T is the kinetic energy and V is the potential energy. In our example, 
 the potential energy V = -0.5*alpha*x**2+0.25*beta*x**4+0.5*omega*y**2.
@@ -55,7 +57,7 @@ n=12
 n_turn=1
 deltaE = e-parameters[2]
 po_fam_file = open("1111x0_turningpoint_deltaE%s_uncoupled.txt" %(deltaE),'a+')
-[x0po_1, T_1,energyPO_1] = uncoupled_turningpoint.TurningPoint_uncoupled(state0_2,state0_3 ,parameters,e,n,n_turn,po_fam_file) ; 
+[x0po_1, T_1,energyPO_1] = turningpoint_UPOsHam2dof.turningPoint(model,state0_2,state0_3 ,parameters,e,n,n_turn,po_fam_file)  
 
 po_fam_file.close()
 
@@ -68,7 +70,7 @@ n=12
 n_turn=1
 deltaE = e-parameters[2]
 po_fam_file = open("1111x0_turningpoint_deltaE%s_uncoupled.txt" %(deltaE),'a+')
-[x0po_2, T_2,energyPO_2] = uncoupled_turningpoint.TurningPoint_uncoupled(state0_2,state0_3 ,parameters,e,n,n_turn,po_fam_file) ; 
+[x0po_2, T_2,energyPO_2] = turningpoint_UPOsHam2dof.turningPoint(model,state0_2,state0_3 ,parameters,e,n,n_turn,po_fam_file)  
 
 po_fam_file.close()
  
@@ -81,7 +83,7 @@ n=12
 n_turn=1
 deltaE = e-parameters[2]
 po_fam_file = open("1111x0_turningpoint_deltaE%s_uncoupled.txt" %(deltaE),'a+')
-[x0po_3, T_3,energyPO_3] = uncoupled_turningpoint.TurningPoint_uncoupled(state0_2,state0_3 ,parameters,e,n,n_turn,po_fam_file) ; 
+[x0po_3, T_3,energyPO_3] = turningpoint_UPOsHam2dof.turningPoint(model,state0_2,state0_3 ,parameters,e,n,n_turn,po_fam_file)  
 
 po_fam_file.close()
 
[…]

[…]
