+++ run_comparison_uncoupled.py
@@ -17,9 +17,8 @@ import scipy.linalg as linalg
 from scipy.optimize import fsolve
 import time
 from functools import partial
-import uncoupled_tpcd  ### import module xxx where xxx is the name of the python file xxx.py 
-import uncoupled_turningpoint
-import uncoupled_diffcorr
+import turningpoint_UPOsHam2dof ### import module xxx where xxx is the name of the python file xxx.py 
+import diffcorr_UPOsHam2dof
 
 from mpl_toolkits.mplot3d import Axes3D
 import matplotlib as mpl
@@ -32,44 +31,44 @@ from scipy import optimize
 
 
 #%% Setting up parameters and global variables(Using differential correction notation)
-N = 4;          # dimension of phase space
+N = 4          # dimension of phase space
 omega=1 # Uncoupled
-EPSILON_S = 0.0; #Energy of the saddle
+EPSILON_S = 0.0 #Energy of the saddle
 alpha = 1.00
 beta = 1.00
-parameters = np.array([1,omega, EPSILON_S, alpha, beta,omega]);
-eqNum = 1;  
-eqPt = uncoupled_diffcorr.get_eq_pts_uncoupled(eqNum, parameters)
+parameters = np.array([1,omega, EPSILON_S, alpha, beta,omega])
+eqNum = 1  
+model = 'uncoupled'
+eqPt = diffcorr_UPOsHam2dof.get_eq_pts(eqNum,model, parameters)
 
-
-eSaddle = uncoupled_diffcorr.get_total_energy_uncoupled([eqPt[0],eqPt[1],0,0], parameters) #energy of the saddle eq pt
+eSaddle = diffcorr_UPOsHam2dof.get_total_energy(model,[eqPt[0],eqPt[1],0,0], parameters) #energy of the saddle eq pt
 #%% Load Data
 deltaE = 0.10
 eSaddle = 0.0 # energy of the saddle
 
 data_path = "./data/"
 
-#po_fam_file = open("1111x0_tpcd_deltaE%s_uncoupled.txt" %(deltaE),'a+');
+#po_fam_file = open("1111x0_tpcd_deltaE%s_uncoupled.txt" %(deltaE),'a+')
 po_fam_file = "1111x0_tpcd_deltaE%s_uncoupled.txt" %(deltaE)
-print('Loading the periodic orbit family from data file',po_fam_file,'\n'); 
+print('Loading the periodic orbit family from data file',po_fam_file,'\n') 
 #x0podata = np.loadtxt(po_fam_file.name)
 #po_fam_file.close()
 x0podata = np.loadtxt(data_path + po_fam_file)
 x0po_1_tpcd = x0podata
 
 
-#po_fam_file = open("1111x0_turningpoint_deltaE%s_uncoupled.txt" %(deltaE),'a+');
+#po_fam_file = open("1111x0_turningpoint_deltaE%s_uncoupled.txt" %(deltaE),'a+')
 po_fam_file = "1111x0_turningpoint_deltaE%s_uncoupled.txt" %(deltaE)
-print('Loading the periodic orbit family from data file',po_fam_file,'\n'); 
+print('Loading the periodic orbit family from data file',po_fam_file,'\n') 
 #x0podata = np.loadtxt(po_fam_file.name)
 #po_fam_file.close()
[…]

[…]
