### 2019-09-12 2d2c811
Update diffcorr method
