+++ run_diffcorr_POfam_deleonberne.py
@@ -21,7 +21,7 @@ import scipy.linalg as linalg
 from scipy.optimize import fsolve
 import time
 from functools import partial
-import deleonberne_diffcorr ### import module xxx where xxx is the name of the python file xxx.py 
+import diffcorr_UPOsHam2dof ### import module xxx where xxx is the name of the python file xxx.py 
 from mpl_toolkits.mplot3d import Axes3D
 import matplotlib as mpl
 from matplotlib import cm
@@ -31,20 +31,20 @@ mpl.rcParams['mathtext.rm'] = 'serif'
 #%matplotlib
 from scipy import optimize
 #%% Setting up parameters and global variables
-N = 4;          # dimension of phase space
-MASS_A = 8.0; MASS_B = 8.0; # De Leon, Marston (1989)
-EPSILON_S = 1.0;
-D_X = 10.0;
-ALPHA = 1.00;
-LAMBDA = 1.5;
-parameters = np.array([MASS_A, MASS_B, EPSILON_S, D_X, LAMBDA, ALPHA]);
-eqNum = 1;  
+N = 4          # dimension of phase space
+MASS_A = 8.0 
+MASS_B = 8.0 # De Leon, Marston (1989)
+EPSILON_S = 1.0
+D_X = 10.0
+ALPHA = 1.00
+LAMBDA = 1.5
+parameters = np.array([MASS_A, MASS_B, EPSILON_S, D_X, LAMBDA, ALPHA])
+eqNum = 1  
+model = 'deleonberne'
+eqPt = diffcorr_UPOsHam2dof.get_eq_pts(eqNum,model, parameters)
 
-eqPt = deleonberne_diffcorr.get_eq_pts_deleonberne(eqNum, parameters)
 
-
-
-eSaddle = deleonberne_diffcorr.get_total_energy_deleonberne([eqPt[0],eqPt[1],0,0], parameters) #energy of the saddle eq pt
+eSaddle = diffcorr_UPOsHam2dof.get_total_energy(model,[eqPt[0],eqPt[1],0,0], parameters) #energy of the saddle eq pt
 #If orbit is an n-array, e.g. orbit = [orbit_0, orbit_1, ..., orbit_n]
 # 
 #e = np.zeros(n,1)    
@@ -70,7 +70,7 @@ t = time.time()
 
 
 po_fam_file = open("1111x0_tp_fam_eqPt%s_deleonberne.txt" %eqNum,'a+')
-[po_x0Fam,po_tpFam] = deleonberne_diffcorr.get_POFam_deleonberne(eqNum, Ax1, Ax2,nFam, po_fam_file, parameters) ; 
+[po_x0Fam,po_tpFam] = diffcorr_UPOsHam2dof.get_POFam(eqNum, Ax1, Ax2,nFam, po_fam_file, parameters,model)  
 
 poFamRuntime = time.time()-t
 x0podata = np.concatenate((po_x0Fam, po_tpFam),axis=1)
@@ -81,34 +81,34 @@ po_fam_file.close()
 #%%
 # begins with a family of periodic orbits and steps until crossing the
 # initial condition with target energy 
-# fileName = 'x0po_T_energy_case1_L41.txt';
-# fileName = 'x0po_T.txt';
+# fileName = 'x0po_T_energy_case1_L41.txt'
+# fileName = 'x0po_T.txt'
 
 deltaE = 0.10
-po_fam_file = open("1111x0_tp_fam_eqPt%s_deleonberne.txt" %eqNum ,'a+');
-eTarget = eSaddle + deltaE; 
-print('Loading the periodic orbit family from data file',po_fam_file.name,'\n'); 
[…]
