### 2019-09-12 07e0d3f
split turningpoint methods
