+++ tp_UPOsHam2dof.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 """
-Created on Mon Sep  9 17:49:31 2019
+Created on Thu Sep 12 17:34:06 2019
 
 @author: wl16298
 """
@@ -559,218 +559,5 @@ def turningPoint(model,begin1,begin2,par,e,n,n_turn,po_fam_file):
     dum =np.concatenate((x0po,T, energyPO),axis=1)
     np.savetxt(po_fam_file.name,dum,fmt='%1.16e')
     return x0po, T,energyPO
-
-
-#%%
-def configdiff(model,guess1, guess2,n_turn,par):
-    """
-    configdiff(model,guess1, guess2,n_turn,par) returns the difference of x(or y) coordinates between the guess initial conditions and the ith turning points
-    n_turn is the nth turning point we want to choose as our 'turning point' for defining the dot product
-    either difference in x coordintes(x_diff1, x_diff2) or difference in y coordinates(y_diff1, y_diff2) is returned as the result.
-    """
-    TSPAN = [0,40]
-    RelTol = 3.e-10
-    AbsTol = 1.e-10 
-    f1 = lambda t,x: Ham2dof(model,t,x,par) 
-    soln1 = solve_ivp(f1, TSPAN, guess1,method='RK45',dense_output=True, events = lambda t,x: half_period(t,x,model),rtol=RelTol, atol=AbsTol)
-    te1 = soln1.t_events[0]
-    t1 = [0,te1[n_turn]]#[0,te1[1]]
-    turn1 = soln1.sol(t1)
-    x_turn1 = turn1[0,-1] 
-    y_turn1 = turn1[1,-1]
-    x_diff1 = guess1[0] - x_turn1
-    y_diff1 = guess1[1] - y_turn1
-    f2 = lambda t,x: Ham2dof(model,t,x,par) 
-    soln2 = solve_ivp(f2, TSPAN, guess2,method='RK45',dense_output=True, events = lambda t,x: half_period(t,x,model),rtol=RelTol, atol=AbsTol)
-    te2 = soln2.t_events[0]
-    t2 = [0,te2[n_turn]]#[0,te2[1]]
-    turn2 = soln2.sol(t2)
-    x_turn2 = turn2[0,-1] 
-    y_turn2 = turn2[1,-1] 
-    x_diff2 = guess2[0] - x_turn2
-    y_diff2 = guess2[1] - y_turn2
-    if model == 'uncoupled':
-        print("Initial guess1%s, intial guess2%s, x_diff1 is %s, x_diff2 is%s " %(guess1,guess2,x_diff1, x_diff2))
-        return x_diff1, x_diff2
-    elif model == 'coupled':
-        print("Initial guess1%s, intial guess2%s, x_diff1 is %s, x_diff2 is%s " %(guess1,guess2,x_diff1, x_diff2))
-        return x_diff1, x_diff2
-    elif model == 'deleonberne':
-        print("Initial guess1%s, intial guess2%s, y_diff1 is %s, y_diff2 is%s " %(guess1,guess2,y_diff1, y_diff2))
-        return y_diff1, y_diff2
-    else: 
-        print("Need to decide to use either x_diff or y_diff")
-        print("Initial guess1%s, intial guess2%s, x_diff1 is %s, x_diff2 is%s, y_diff1 is %s, y_diff2 is%s " %(guess1,guess2,x_diff1, x_diff2,y_diff1, y_diff2))
-        return x_diff1, x_diff2,y_diff1, y_diff2
-
-
-#%%
-def turningPoint_configdiff(model,begin1,begin2,par,e,n,n_turn,po_fam_file):
-    """
[…]
