### 2019-09-12 bf5de0e
Update turningpoint methods
