### 2019-09-12 6d52238
Update turning point methods
