+++ run_turningpoint_POfam_coupled.py
@@ -22,7 +22,7 @@ from scipy.optimize import fsolve
 import time
 from functools import partial
 from scipy import optimize
-import turningpoint_UPOsHam2dof ### import module xxx where xxx is the name of the python file xxx.py 
+import tp_UPOsHam2dof ### import module xxx where xxx is the name of the python file xxx.py 
 from mpl_toolkits.mplot3d import Axes3D
 import matplotlib as mpl
 from matplotlib import cm
@@ -41,7 +41,7 @@ epsilon= 1e-1
 parameters = np.array([1,omega, EPSILON_S, alpha, beta,omega,epsilon]);
 eqNum = 1;  
 model = 'coupled'
-eqPt = turningpoint_UPOsHam2dof.get_eq_pts(eqNum,model, parameters)
+eqPt = tp_UPOsHam2dof.get_eq_pts(eqNum,model, parameters)
 
 #%%
 # e is the total energy, n is the number of intervals we want to divide, n_turn is the nth turning point we want to choose.
@@ -51,18 +51,18 @@ n_turn = 1
 deltaE = e-parameters[2]
 """Trial initial Condition s.t. one initial condition is on the LHS of the UPO and the other one is on the RHS of the UPO"""
 x=0.0
-f2 = lambda y: turningpoint_UPOsHam2dof.get_coordinate(model,0.0,y,e,parameters)
+f2 = lambda y: tp_UPOsHam2dof.get_coordinate(model,0.0,y,e,parameters)
 yanalytic = math.sqrt(2/(parameters[1]+parameters[6]))*(-math.sqrt( e +0.5*parameters[3]* x**2- 0.25*parameters[4]*x**4 -0.5*parameters[6]* x**2 + (parameters[6]*x)**2/(2*(parameters[1] +parameters[6]) )) +parameters[6]/(math.sqrt(2*(parameters[1]+parameters[6])) )*x ) #coupled
 state0_2 = [0.0,optimize.newton(f2,yanalytic),0.0,0.0]
 
 x=0.05
-f3 = lambda y: turningpoint_UPOsHam2dof.get_coordinate(model,x,y,e,parameters)
+f3 = lambda y: tp_UPOsHam2dof.get_coordinate(model,x,y,e,parameters)
 yanalytic = math.sqrt(2/(parameters[1]+parameters[6]))*(-math.sqrt( e +0.5*parameters[3]* x**2- 0.25*parameters[4]*x**4 -0.5*parameters[6]* x**2 + (parameters[6]*x)**2/(2*(parameters[1] +parameters[6]) )) +parameters[6]/(math.sqrt(2*(parameters[1]+parameters[6])) )*x ) #coupled
 state0_3 = [x, optimize.newton(f3,yanalytic),0.0,0.0]
 
 
 po_fam_file = open("1111x0_turningpoint_deltaE%s_coupled.txt" %(deltaE),'a+')
-[x0po_1, T_1,energyPO_1] = turningpoint_UPOsHam2dof.turningPoint(model,state0_2,state0_3 ,parameters,e,n,n_turn,po_fam_file) ; 
+[x0po_1, T_1,energyPO_1] = tp_UPOsHam2dof.turningPoint(model,state0_2,state0_3 ,parameters,e,n,n_turn,po_fam_file) ; 
 po_fam_file.close()
 
 #%%
@@ -72,18 +72,18 @@ n_turn = 1
[…]
