+++ run_turningpoint_POfam_uncoupled.py
@@ -23,7 +23,7 @@ from scipy.optimize import fsolve
 from scipy import optimize
 import time
 from functools import partial
-import turningpoint_UPOsHam2dof ### import module xxx where xxx is the name of the python file xxx.py 
+import tp_UPOsHam2dof ### import module xxx where xxx is the name of the python file xxx.py 
 from mpl_toolkits.mplot3d import Axes3D
 import matplotlib as mpl
 from matplotlib import cm
@@ -40,7 +40,7 @@ EPSILON_S = 0.0 #Energy of the saddle
 parameters = np.array([1,omega, EPSILON_S, alpha, beta,omega])
 eqNum = 1 
 model = 'uncoupled'
-eqPt = turningpoint_UPOsHam2dof.get_eq_pts(eqNum,model, parameters)
+eqPt = tp_UPOsHam2dof.get_eq_pts(eqNum,model, parameters)
 """Initial Condition for the true periodic orbit
 H = T + V where T is the kinetic energy and V is the potential energy. In our example, 
 the potential energy V = -0.5*alpha*x**2+0.25*beta*x**4+0.5*omega*y**2.
@@ -57,7 +57,7 @@ n=12
 n_turn=1
 deltaE = e-parameters[2]
 po_fam_file = open("1111x0_turningpoint_deltaE%s_uncoupled.txt" %(deltaE),'a+')
-[x0po_1, T_1,energyPO_1] = turningpoint_UPOsHam2dof.turningPoint(model,state0_2,state0_3 ,parameters,e,n,n_turn,po_fam_file)  
+[x0po_1, T_1,energyPO_1] = tp_UPOsHam2dof.turningPoint(model,state0_2,state0_3 ,parameters,e,n,n_turn,po_fam_file)  
 
 po_fam_file.close()
 
@@ -70,7 +70,7 @@ n=12
 n_turn=1
 deltaE = e-parameters[2]
 po_fam_file = open("1111x0_turningpoint_deltaE%s_uncoupled.txt" %(deltaE),'a+')
-[x0po_2, T_2,energyPO_2] = turningpoint_UPOsHam2dof.turningPoint(model,state0_2,state0_3 ,parameters,e,n,n_turn,po_fam_file)  
+[x0po_2, T_2,energyPO_2] = tp_UPOsHam2dof.turningPoint(model,state0_2,state0_3 ,parameters,e,n,n_turn,po_fam_file)  
 
 po_fam_file.close()
  
@@ -83,7 +83,7 @@ n=12
 n_turn=1
 deltaE = e-parameters[2]
 po_fam_file = open("1111x0_turningpoint_deltaE%s_uncoupled.txt" %(deltaE),'a+')
-[x0po_3, T_3,energyPO_3] = turningpoint_UPOsHam2dof.turningPoint(model,state0_2,state0_3 ,parameters,e,n,n_turn,po_fam_file)  
+[x0po_3, T_3,energyPO_3] = tp_UPOsHam2dof.turningPoint(model,state0_2,state0_3 ,parameters,e,n,n_turn,po_fam_file)  
 
 po_fam_file.close()
 
@@ -96,7 +96,7 @@ n=12
 n_turn=1
 deltaE = e-parameters[2]
 po_fam_file = open("1111x0_turningpoint_deltaE%s_uncoupled.txt" %(deltaE),'a+')
[…]
