+++ run_turningpoint_POfam_deleonberne.py
@@ -21,7 +21,7 @@ import scipy.linalg as linalg
 from scipy.optimize import fsolve
 import time
 from functools import partial
-import turningpoint_UPOsHam2dof ### import module xxx where xxx is the name of the python file xxx.py 
+import tp_UPOsHam2dof ### import module xxx where xxx is the name of the python file xxx.py 
 from mpl_toolkits.mplot3d import Axes3D
 import matplotlib as mpl
 from matplotlib import cm
@@ -41,7 +41,7 @@ LAMBDA = 1.5
 parameters = np.array([MASS_A, MASS_B, EPSILON_S, D_X, LAMBDA, ALPHA])
 eqNum = 1  
 model = 'deleonberne'
-eqPt = turningpoint_UPOsHam2dof.get_eq_pts(eqNum, model,parameters)
+eqPt = tp_UPOsHam2dof.get_eq_pts(eqNum, model,parameters)
 #%%
 # e is the total energy, n is the number of intervals we want to divide, n_turn is the nth turning point we want to choose.
 e=1.1
@@ -49,15 +49,15 @@ n=12
 n_turn = 1
 deltaE = e-parameters[2] #In this case deltaE = 0.1
 """Trial initial Condition s.t. one initial condition is on the top of the UPO and the other one is on the bottom of the UPO"""
-f1 = lambda x: turningpoint_UPOsHam2dof.get_coordinate(model,x,0.06,e,parameters)
+f1 = lambda x: tp_UPOsHam2dof.get_coordinate(model,x,0.06,e,parameters)
 x0_2 = optimize.newton(f1,-0.15)
 state0_2 = [x0_2,0.06,0.0,0.0]
-f2 = lambda x: turningpoint_UPOsHam2dof.get_coordinate(model,x,-0.05,e,parameters)
+f2 = lambda x: tp_UPOsHam2dof.get_coordinate(model,x,-0.05,e,parameters)
 x0_3 = optimize.newton(f2,-0.15)
 state0_3 = [x0_3, -0.05,0.0,0.0]
 
 po_fam_file = open("1111x0_turningpoint_deltaE%s_deleonberne.txt" %(0.1),'a+')
-[x0po_1, T_1,energyPO_1] = turningpoint_UPOsHam2dof.turningPoint(model,state0_3,state0_2 ,parameters,e,n,n_turn,po_fam_file)  
+[x0po_1, T_1,energyPO_1] = tp_UPOsHam2dof.turningPoint(model,state0_3,state0_2 ,parameters,e,n,n_turn,po_fam_file)  
 
 po_fam_file.close()
 #%%
@@ -67,15 +67,15 @@ n=12
 n_turn = 1
 deltaE = e-parameters[2]
 """Trial initial Condition s.t. one initial condition is on the LHS of the UPO and the other one is on the RHS of the UPO"""
-f1 = lambda x: turningpoint_UPOsHam2dof.get_coordinate(model,x,0.06,e,parameters)
+f1 = lambda x: tp_UPOsHam2dof.get_coordinate(model,x,0.06,e,parameters)
 x0_2 = optimize.newton(f1,-0.15)
 state0_2 = [x0_2,0.06,0.0,0.0]
-f2 = lambda x: turningpoint_UPOsHam2dof.get_coordinate(model,x,-0.05,e,parameters)
[…]
