### 2019-09-13 #1684 @Shibabrat
@JoshKarpel Yes, you are correct. Some (I point to those below) details of the methods can not be decoupled from the specific features of the problem. In particular, one such feature is a detailed knowledge of the bottleneck region of a potential energy surface (PES). For a given molecular system, the PES is the output of a quantum chemistry calculation and the results are highly dependent on the molecule.

The first-order correction based on the monodromy matrix for the differential correction needs derivation based on how the bottleneck is oriented in the configuration space. The correction term is different in https://github.com/WyLyu/UPOsHam/blob/master/diffcorr_UPOsHam2dof.py#L744
from https://github.com/WyLyu/UPOsHam/blob/master/diffcorr_UPOsHam2dof.py#L759, and it can also be same as in https://github.com/WyLyu/UPOsHam/blob/master/diffcorr_UPOsHam2dof.py#L752. This correction is applied to the phase space coordinate that is most natural for the bottleneck of the PES. Fixing this coordinate also determines the coordinate that is used to check convergence (https://github.com/WyLyu/UPOsHam/blob/master/diffcorr_UPOsHam2dof.py#L651) and the one that is used to terminate the integration using event function (https://github.com/WyLyu/UPOsHam/blob/master/diffcorr_UPOsHam2dof.py#L303).

The turning point based on coordinate difference relies on the orientation of the PES to choose whether to use x or y or both coordinates to check which direction the trajectory is turning. In this case, the orientation of the bottleneck of the PES determines how the turning direction is checked (https://github.com/WyLyu/UPOsHam/blob/master/tpcd_UPOsHam2dof.py#L362).

The turning point method probably can be generalized but then one needs to manually choose two initial conditions near the final/true UPO. This is probably the one that comes closest to being written without problem-specific details as long as the user starts with two assumptions: first, the two starting initial conditions for the method should be close (whatever that means for the problem) to the final UPO and second the accuracy can't be improved beyond 10e-3 (which is fine for rough estimates or to restart the TPCD method but not for using the UPO in other quantitative calculations such as transition fraction or reaction flux)

Having said that, I do think some portions of the package can be separated as user-supplied functions with specific signatures, as per your suggestion.
