### 2019-09-15 #1684 @JoshKarpel
@Shibabrat  When I see a block like https://github.com/WyLyu/UPOsHam/blob/master/diffcorr_UPOsHam2dof.py#L738 , it seems clear to me that this calculation is part of specifying the problem to be solved, and should be passed into the solver by the user. In that case it looks like the input parameters are `par`, `x0`, `x1`, and `phi_t1`, and the function should return the corrected `x0`.

If that means you have to pass ten different functions (or parameters, or whatever) into the solver, so be it: **in the current version the user would have to do that anyway, but by copy-paste-editing long functions instead of writing their own small, focused functions**. I think that passing configuration to the solver, even if there is a lot of it, is a far friendlier than being asked to make specific edits in hundreds of lines of code.

The examples would then largely be the definitions of the functions and configuration parameters for the appropriate problems, which would be much easier for people to take and re-purpose for their own problems.
