### 2019-09-17 905721a
Updating function signatures
