### 2019-09-17 b59d5a3
Updating function signatures
