+++ run_diffcorr_POfam_deleonberne.py
@@ -2,35 +2,236 @@
 """
 Created on Tue Jul 30 10:02:48 2019
 
-@author: wl16298
+@author: Wenyang Lyu and Shibabrat Naik
+
+Script to compute unstable periodic orbits at specified energies for the coupled quartic 
+Hamiltonian using differential correction
 """
-
-"""Plot peridoic orbits with 5 different energies using the new method"""
-
-# For deleonberne problem
 
 import numpy as np
 import matplotlib.pyplot as plt
-from scipy.integrate import odeint,quad,trapz,solve_ivp
-import math
-from IPython.display import Image # for Notebook
-from IPython.core.display import HTML
-from mpl_toolkits.mplot3d import Axes3D
-from numpy import linalg as LA
-import scipy.linalg as linalg
-from scipy.optimize import fsolve
+from scipy.integrate import solve_ivp
 import time
-from functools import partial
 import diffcorr_UPOsHam2dof ### import module xxx where xxx is the name of the python file xxx.py 
-from mpl_toolkits.mplot3d import Axes3D
 import matplotlib as mpl
 from matplotlib import cm
-from pylab import rcParams
 mpl.rcParams['mathtext.fontset'] = 'cm'
 mpl.rcParams['mathtext.rm'] = 'serif'
-#%matplotlib
-from scipy import optimize
+
+
+#% Begin problem specific functions
+def init_guess_eqpt_deleonberne(eqNum, par):
+    
+    if eqNum == 1:
+        x0 = [0, 0]
+    if eqNum == 2:
+        x0 = [0, 1/np.sqrt(2)]  # EQNUM = 2, center-center
+    elif eqNum == 3:
+        x0 = [0, -1/np.sqrt(2)] # EQNUM = 3, center-center
+    
+    return x0
+
+
+def grad_pot_deleonberne(x, par):
+    
+    dVdx = -2*par[3]*par[4]*np.exp(-par[4]*x[0])*(np.exp(-par[4]*x[0]) - 1) - \
+        4*par[5]*par[4]*x[1]**2*(x[1]**2 - 1)*np.exp(-par[5]*par[4]*x[0])
+    dVdy = 8*x[1]*(2*x[1]**2 - 1)*np.exp(-par[5]*par[4]*x[0])
+    
+    F = [-dVdx, -dVdy]
+    
+    return F
+
+def pot_energy_deleonberne(x, y, par):
+    return par[3]*( 1 - np.exp(-par[4]*x) )**2 + 4*y**2*(y**2 - 1)*np.exp(-par[5]*par[4]*x) + par[2]
+
+
+def eigvector_deleonberne(par):
+    """
+    eigenvectors and eigenvalues of the Jacobian evaluated at the equilibrium point, 
+    which is the correction of the initial condition.
+    check the result obtained from the Jacobian matches the analytic result.
+    """
+
+    correcx = 1
+    correcy = 0
+    
+    return correcx, correcy
+
+
+def guess_lin_deleonberne(eqPt, Ax, par):
+    
+    correcx, correcy = eigvector_deleonberne(par)
+    
+
+    return [eqPt[0] - Ax*correcx,eqPt[1] + Ax*correcy,0,0]
+
+
+def jacobian_deleonberne(eqPt, par):
+    
+    x,y,px,py = eqPt[0:4]
+    # The first order derivative of the Hamiltonian.
+    dVdx = - 2*par[3]*par[4]*np.exp(-par[4]*x)*(np.exp(-par[4]*x) - 1) - 4*par[5]*par[4]*y**2*(y**2 - 1)*np.exp(-par[5]*par[4]*x) 
+    dVdy = 8*y*(2*y**2 - 1)*np.exp(-par[5]*par[4]*x)
+
[…]
