+++ diffcorr_UPOsHam2dof.py
@@ -2,7 +2,7 @@
 """
 Created on Tue Sep 10 18:32:28 2019
 
-@author: wl16298
+@author: Wenyang Lyu and Shibabrat Naik
 """
 
 import numpy as np
@@ -128,166 +128,6 @@ def stateTransitMat(tf,x0,par,varEqns_model,fixed_step=0):
 
     
     return t,x,phi_tf,PHI
-
-
-#%%
-#def varEqns(t,PHI,par,model):
-#    
-#    """
-#    PHIdot = varEqns_bp(t,PHI) 
-#    
-#    This here is a preliminary state transition, PHI(t,t0),
-#    matrix equation attempt for a ball rolling on the surface, based on...
-#    
-#    d PHI(t, t0)
-#    ------------ =  Df(t) * PHI(t, t0)
-#        dt
-#    
-#    """
-#    
-#    phi = PHI[0:16]
-#    phimatrix  = np.reshape(PHI[0:16],(4,4))
-#    x,y,px,py = PHI[16:20]
-#    
-#    
-#    if model == 'uncoupled':
-#        # The first order derivative of the potential energy.
-#        dVdx = -par[3]*x+par[4]*x**3
-#        dVdy = par[5]*y
-#    
-#        # The second order derivative of the potential energy.  
-#        d2Vdx2 = -par[3]+par[4]*3*x**2
-#            
-#        d2Vdy2 = par[5]
-#    
-#        d2Vdydx = 0
-#    elif model == 'coupled':
-#        # The first order derivative of the potential energy.
-#        dVdx = (-par[3]+par[6])*x+par[4]*x**3-par[6]*y
-#        dVdy = (par[5]+par[6])*y-par[6]*x
-#
-#        # The second order derivative of the potential energy. 
-#        d2Vdx2 = -par[3]+par[6]+par[4]*3*x**2
-#            
-#        d2Vdy2 = par[5]+par[6]
-#    
-#        d2Vdydx = -par[6]
-#    elif model == 'deleonberne':
-#        # The first order derivative of the potential energy.
-#        dVdx = - 2*par[3]*par[4]*math.e**(-par[4]*x)*(math.e**(-par[4]*x) - 1) - 4*par[5]*par[4]*y**2*(y**2 - 1)*math.e**(-par[5]*par[4]*x) 
-#        dVdy = 8*y*(2*y**2 - 1)*math.e**(-par[5]*par[4]*x)
-#    
-#        # The second order derivative of the potential energy. 
-#        d2Vdx2 = - ( 2*par[3]*par[4]**2*( math.e**(-par[4]*x) - 2.0*math.e**(-2*par[4]*x) ) - 4*(par[5]*par[4])**2*x**2*(y**2 - 1)*math.e**(-par[5]*par[4]*x) )
-#            
-#        d2Vdy2 = 8*(6*y**2 - 1)*math.e**( -par[4]*par[5]*x )
-#    
-#        d2Vdydx = -8*y*par[4]*par[5]*math.e**( -par[4]*par[5]*x )*(2*y**2 - 1)
-#        
-#    else:
-#        print("The model you are chosen does not exist")
-#    
-#    
-#    d2Vdxdy = d2Vdydx    
-#
-#    Df    = np.array([[  0,     0,    par[0],    0],
-#              [0,     0,    0,    par[1]],
-#              [-d2Vdx2,  -d2Vdydx,   0,    0],
-#              [-d2Vdxdy, -d2Vdy2,    0,    0]])
-#
-#    
-#    phidot = np.matmul(Df, phimatrix) # variational equation
-#
-#    PHIdot        = np.zeros(20)
-#    PHIdot[0:16]  = np.reshape(phidot,(1,16)) 
-#    PHIdot[16]    = px/par[0]
[…]

[…]
