### 2019-09-25 032c3df
Removed unused function
