### 2019-09-26 794fc77
Updated turning point based configuration difference method
