### 2019-09-26 f6beab4
Problem specifics separated from the turning point based configuration difference method
