+++ run_tpcd_POfam_uncoupled.py
@@ -2,29 +2,200 @@
 """
 Created on Wed Sep 11 13:16:47 2019
 
-@author: wl16298
+@author: Wenyang Lyu and Shibabrat Naik
 """
 
 import numpy as np
 import matplotlib.pyplot as plt
-from scipy.integrate import odeint,quad,trapz,solve_ivp
+from scipy.integrate import solve_ivp
 import math
-from IPython.display import Image # for Notebook
-from IPython.core.display import HTML
-from mpl_toolkits.mplot3d import Axes3D
-from numpy import linalg as LA
-import scipy.linalg as linalg
-from scipy.optimize import fsolve
-import time
-from functools import partial
 import tpcd_UPOsHam2dof ### import module xxx where xxx is the name of the python file xxx.py 
-from mpl_toolkits.mplot3d import Axes3D
+#from mpl_toolkits.mplot3d import Axes3D
 import matplotlib as mpl
 from matplotlib import cm
-from pylab import rcParams
+
 mpl.rcParams['mathtext.fontset'] = 'cm'
 mpl.rcParams['mathtext.rm'] = 'serif'
-#%matplotlib
+
+#% Begin problem specific functions
+def init_guess_eqpt_uncoupled(eqNum, par):
+    
+    if eqNum == 1:
+        x0 = [0, 0]
+    if eqNum == 2:
+        x0 = [np.sqrt(par[3]/par[4]),0] 
+    elif eqNum == 3:
+        x0 = [-np.sqrt(par[3]/par[4]),0]  
+    
+    return x0
+
+
+def grad_pot_uncoupled(x,par):
+    
+    dVdx = -par[3]*x[0]+par[4]*(x[0])**3
+    dVdy = par[5]*x[1]
+    
+    F = [-dVdx, -dVdy]
+    
+    return F
+
+
+def pot_energy_uncoupled(x, y, par):
+    return -0.5*par[3]*x**2+0.25*par[4]*x**4 +0.5*par[5]*y**2
+
+
+def get_coord_uncoupled(x,y, E, par):
+    """ this function returns the initial position of x/y-coordinate on the potential energy surface(PES) for a specific energy V.
+    
+    """
+    return -0.5*par[3]*x**2+0.25*par[4]*x**4 +0.5*par[5]*y**2 - E
+#    elif model== 'deleonberne':
+#        return par[3]*( 1 - math.e**(-par[4]*x) )**2 + 4*y**2*(y**2 - 1)*math.e**(-par[5]*par[4]*x) + par[2]-V
+#    else:
+#        print("The model you are chosen does not exist, enter the function for finding coordinates on the PES for given x or y and V")
+
+
+def varEqns_uncoupled(t,PHI,par):
+    
+    """
+    PHIdot = varEqns_uncoupled(t,PHI) 
+    
+    This here is a preliminary state transition, PHI(t,t0),
+    matrix equation attempt for a ball rolling on the surface, based on...
+    
+    d PHI(t, t0)
+    ------------ =  Df(t) * PHI(t, t0)
+        dt
+    
+    """
+    
+    phi = PHI[0:16]
+    phimatrix  = np.reshape(PHI[0:16],(4,4))
+    x,y,px,py = PHI[16:20]
+    
+    
+    # The first order derivative of the potential energy.
+    dVdx = -par[3]*x+par[4]*x**3
+    dVdy = par[5]*y
+
+    # The second order derivative of the potential energy.  
+    d2Vdx2 = -par[3]+par[4]*3*x**2
[…]
