### 2019-09-27 2e65e30
Problem specifics separated from the method
