### 2019-09-30 7076f39
Problem specifics separated from the method
