### 2019-09-30 af7c3bd
Separated problem specifics from the turning point method
