+++ tp_UPOsHam2dof.py
@@ -7,172 +7,87 @@ Created on Thu Sep 12 17:34:06 2019
 
 import numpy as np
 import matplotlib.pyplot as plt
-from scipy.integrate import odeint,quad,trapz,solve_ivp, ode
+from scipy.integrate import solve_ivp
 import math
-from IPython.display import Image # for Notebook
-from IPython.core.display import HTML
+#from IPython.display import Image # for Notebook
+#from IPython.core.display import HTML
 from mpl_toolkits.mplot3d import Axes3D
-from numpy import linalg as LA
-import scipy.linalg as linalg
+#from numpy import linalg as LA
+#import scipy.linalg as linalg
 from scipy.optimize import fsolve
-import time
-from functools import partial
-from scipy import optimize
+#import time
+#from functools import partial
+#from scipy import optimize
 import matplotlib as mpl
 
-from pylab import rcParams
+#from pylab import rcParams
 mpl.rcParams['mathtext.fontset'] = 'cm'
 mpl.rcParams['mathtext.rm'] = 'serif'
 
 
 #%%
-def func_vec_field_eq_pt(model,x,par):
-    """ vecor field(same as pxdot, pydot), used to find the equilibrium points of the system.
-    """
-    if model == 'uncoupled':
-        dVdx = -par[3]*x[0]+par[4]*(x[0])**3
-        dVdy = par[5]*x[1]
-    elif model == 'coupled':
-        dVdx = (-par[3]+par[6])*x[0]+par[4]*(x[0])**3-par[6]*x[1]
-        dVdy = (par[5]+par[6])*x[1]-par[6]*x[0]
-    elif model == 'deleonberne':
-        dVdx = -2*par[3]*par[4]*math.e**(-par[4]*x[0])*(math.e**(-par[4]*x[0]) - 1) - 4*par[5]*par[4]*x[1]**2*(x[1]**2 - 1)*math.e**(-par[5]*par[4]*x[0])
-        dVdy = 8*x[1]*(2*x[1]**2 - 1)*math.e**(-par[5]*par[4]*x[0])
-    else:
-        print("The model you are chosen does not exist, enter the definition of Hamilton's equations")
-    F = [-dVdx, -dVdy]
-    return F
-    
 
-#%%
-def get_eq_pts(eqNum,model, par):
+def get_eq_pts(eqNum, init_guess_eqpt_model, grad_pot_model, par):
     """
-    GET_EQ_PTS_BP solves the saddle center equilibrium point for a system with
-    KE + PE. 
-     H = 1/2*px^2+omega/2*py^2 -(1/2)*alpha*x^2+1/4*beta*x^4+ omega/2y^2 with alpha > 0
-    --------------------------------------------------------------------------
-       Uncouled potential energy surface notations:
+    GET_EQ_PTS solves the coordinates of the equilibrium point for a Hamiltonian 
+    system of the form KE + PE. 
     
-               Well (stable, EQNUM = 2)    
+    eqNum = 1 is the saddle-center equilibrium point
+    eqNum = 2 or 3 is the center-center equilibrium point
     
-                   Saddle (EQNUM=1)
+    init_guess_eqpt_model: function name that defines the initial guess for the critical point
     
-               Well (stable, EQNUM = 3)    
+    grad_pot_model: function name that defines the vector of potential gradient
     
-    --------------------------------------------------------------------------
-       
-    
+    par: list of parameters for the problem
     """
[…]
