### 2019-09-30 f9b3261
Separated problem specifics from the turning point method
