### 2019-09-30 1658056
Separated problem specifics from the turning point method
