+++ run_turningpoint_POfam_coupled.py
@@ -2,36 +2,191 @@
 """
 Created on Tue Jul 30 10:02:48 2019
 
-@author: wl16298
+@author: Wenyang Lyu and Shibabrat Naik
+
+Compute unstable peridoic orbits at different energies using turning point method
 """
 
-"""Plot peridoic orbits with 5 different energies using turning point method"""
-
 # For the coupled problem
-
 import numpy as np
 import matplotlib.pyplot as plt
-from scipy.integrate import odeint,quad,trapz,solve_ivp
+from scipy.integrate import solve_ivp
 import math
-from IPython.display import Image # for Notebook
-from IPython.core.display import HTML
-from mpl_toolkits.mplot3d import Axes3D
-from numpy import linalg as LA
-import scipy.linalg as linalg
-from scipy.optimize import fsolve
-import time
-from functools import partial
 from scipy import optimize
 import tp_UPOsHam2dof ### import module xxx where xxx is the name of the python file xxx.py 
-from mpl_toolkits.mplot3d import Axes3D
 import matplotlib as mpl
 from matplotlib import cm
-from pylab import rcParams
 mpl.rcParams['mathtext.fontset'] = 'cm'
 mpl.rcParams['mathtext.rm'] = 'serif'
-#%matplotlib
+
+
+#% Begin problem specific functions
+def init_guess_eqpt_coupled(eqNum, par):
+    
+    if eqNum == 1:
+        x0 = [0, 0]
+    if eqNum == 2:
+        x0 = [np.sqrt(par[3]-par[6]/par[4]),0] 
+    elif eqNum == 3:
+        x0 = [-np.sqrt(par[3]-par[6]/par[4]),0] 
+    
+    return x0
+
+def grad_pot_coupled(x, par):
+    
+    dVdx = (-par[3]+par[6])*x[0]+par[4]*(x[0])**3-par[6]*x[1]
+    dVdy = (par[5]+par[6])*x[1]-par[6]*x[0]
+            
+    F = [-dVdx, -dVdy]
+    
+    return F
+
+def pot_energy_coupled(x, y, par):
+    return -0.5*par[3]*x**2+0.25*par[4]*x**4 +0.5*par[5]*y**2+0.5*par[6]*(x-y)**2
+
+
+def get_coord_coupled(x, y, E, par):
+    """ 
+    this function returns the initial position of x/y-coordinate on the potential energy 
+    surface(PES) for a specific energy E.
+    """
+#    if model == 'uncoupled':
+#        return -0.5*par[3]*x**2+0.25*par[4]*x**4 +0.5*par[5]*y**2-V
+#    elif model =='coupled':
+    return -0.5*par[3]*x**2+0.25*par[4]*x**4 +0.5*par[5]*y**2+0.5*par[6]*(x-y)**2 - E
+#    elif model== 'deleonberne':
+#        return par[3]*( 1 - math.e**(-par[4]*x) )**2 + 4*y**2*(y**2 - 1)*math.e**(-par[5]*par[4]*x) + par[2]-V
+#    else:
+#        print("The model you are chosen does not exist, enter the function for finding coordinates on the PES for given x or y and V")
+
+
+def varEqns_coupled(t,PHI,par):
+    
+    """
+    PHIdot = varEqns_coupled(t,PHI) 
+    
+    This here is a preliminary state transition, PHI(t,t0),
+    matrix equation attempt for a ball rolling on the surface, based on...
[…]
