### 2019-09-30 f44779d
Separated problem specifics from the differential correction method
