### 2019-10-01 75d9f51
Updated tests for the coupled and uncoupled model problems
