+++ run_comparison_coupled.py
@@ -2,33 +2,147 @@
 """
 Created on Wed Jul 31 16:45:58 2019
 
-@author: wl16298
+@author: Wenyang Lyu and Shibabrat Naik
 """
 
 import numpy as np
 import matplotlib.pyplot as plt
-from scipy.integrate import odeint,quad,trapz,solve_ivp
-import math
-from IPython.display import Image # for Notebook
-from IPython.core.display import HTML
+#from scipy.integrate import odeint,quad,trapz,solve_ivp
+from scipy.integrate import solve_ivp
+#import math
+#from IPython.display import Image # for Notebook
+#from IPython.core.display import HTML
 from mpl_toolkits.mplot3d import Axes3D
-from numpy import linalg as LA
-import scipy.linalg as linalg
-from scipy.optimize import fsolve
-import time
-from functools import partial
+#from numpy import linalg as LA
+#import scipy.linalg as linalg
+#from scipy.optimize import fsolve
+#import time
+#from functools import partial
 import tpcd_UPOsHam2dof ### import module xxx where xxx is the name of the python file xxx.py 
 import diffcorr_UPOsHam2dof
 
 
-from mpl_toolkits.mplot3d import Axes3D
+#from mpl_toolkits.mplot3d import Axes3D
 import matplotlib as mpl
 from matplotlib import cm
-from pylab import rcParams
+#from pylab import rcParams
 mpl.rcParams['mathtext.fontset'] = 'cm'
 mpl.rcParams['mathtext.rm'] = 'serif'
-#%matplotlib
-from scipy import optimize
+
+#from scipy import optimize
+
+
+#% Begin problem specific functions
+def init_guess_eqpt_coupled(eqNum, par):
+    
+    if eqNum == 1:
+        x0 = [0, 0]
+    if eqNum == 2:
+        x0 = [np.sqrt(par[3]-par[6]/par[4]),0] 
+    elif eqNum == 3:
+        x0 = [-np.sqrt(par[3]-par[6]/par[4]),0] 
+    
+    return x0
+
+def grad_pot_coupled(x, par):
+    
+    dVdx = (-par[3]+par[6])*x[0]+par[4]*(x[0])**3-par[6]*x[1]
+    dVdy = (par[5]+par[6])*x[1]-par[6]*x[0]
+    
+    F = [-dVdx, -dVdy]
+    
+    return F
+
+def pot_energy_coupled(x, y, par):
+    return -0.5*par[3]*x**2+0.25*par[4]*x**4 +0.5*par[5]*y**2+0.5*par[6]*(x-y)**2
+
+
+def varEqns_coupled(t,PHI,par):
+    
+    """
+    PHIdot = varEqns_coupled(t,PHI) 
+    
+    This here is a preliminary state transition, PHI(t,t0),
+    matrix equation attempt for a ball rolling on the surface, based on...
+    
+    d PHI(t, t0)
+    ------------ =  Df(t) * PHI(t, t0)
+        dt
+    
+    """
+    
+    phi = PHI[0:16]
+    phimatrix  = np.reshape(PHI[0:16],(4,4))
+    x,y,px,py = PHI[16:20]
+    
+    
+    # The first order derivative of the potential energy.
+    dVdx = (-par[3]+par[6])*x+par[4]*x**3-par[6]*y
+    dVdy = (par[5]+par[6])*y-par[6]*x
+
[…]
