### 2019-10-01 0f9f829
Updated comparison tests for the uncoupled, coupled, deleon-berne model problems
