+++ run_comparison_deleonberne.py
@@ -2,36 +2,157 @@
 """
 Created on Wed Jul 31 16:53:45 2019
 
-@author: wl16298
+
+@author: Wenyang Lyu and Shibabrat Naik
 """
 
 import numpy as np
 import matplotlib.pyplot as plt
-from scipy.integrate import odeint,quad,trapz,solve_ivp
-import math
-from IPython.display import Image # for Notebook
-from IPython.core.display import HTML
+#from scipy.integrate import odeint,quad,trapz,solve_ivp
+from scipy.integrate import solve_ivp
+#import math
+#from IPython.display import Image # for Notebook
+#from IPython.core.display import HTML
 from mpl_toolkits.mplot3d import Axes3D
-from numpy import linalg as LA
-import scipy.linalg as linalg
-from scipy.optimize import fsolve
-import time
-from functools import partial
+#from numpy import linalg as LA
+#import scipy.linalg as linalg
+#from scipy.optimize import fsolve
+#import time
+#from functools import partial
 import tpcd_UPOsHam2dof ### import module xxx where xxx is the name of the python file xxx.py 
 import diffcorr_UPOsHam2dof
 
 
-from mpl_toolkits.mplot3d import Axes3D
+#from mpl_toolkits.mplot3d import Axes3D
 import matplotlib as mpl
 from matplotlib import cm
-from pylab import rcParams
+#from pylab import rcParams
 mpl.rcParams['mathtext.fontset'] = 'cm'
 mpl.rcParams['mathtext.rm'] = 'serif'
-#%matplotlib
-from scipy import optimize
+
+#from scipy import optimize
+
+
+#% Begin problem specific functions
+def init_guess_eqpt_deleonberne(eqNum, par):
+    
+    if eqNum == 1:
+        x0 = [0, 0]
+    if eqNum == 2:
+        x0 = [0, 1/np.sqrt(2)]  # EQNUM = 2, center-center
+    elif eqNum == 3:
+        x0 = [0, -1/np.sqrt(2)] # EQNUM = 3, center-center
+    
+    return x0
+
+
+def grad_pot_deleonberne(x, par):
+    
+    dVdx = -2*par[3]*par[4]*np.exp(-par[4]*x[0])*(np.exp(-par[4]*x[0]) - 1) - \
+        4*par[5]*par[4]*x[1]**2*(x[1]**2 - 1)*np.exp(-par[5]*par[4]*x[0])
+    dVdy = 8*x[1]*(2*x[1]**2 - 1)*np.exp(-par[5]*par[4]*x[0])
+    
+    F = [-dVdx, -dVdy]
+    
+    return F
+
+def pot_energy_deleonberne(x, y, par):
+    return par[3]*( 1 - np.exp(-par[4]*x) )**2 + 4*y**2*(y**2 - 1)*np.exp(-par[5]*par[4]*x) + par[2]
+
+
+def varEqns_deleonberne(t,PHI,par):
+    """
+    PHIdot = varEqns_deleonberne(t,PHI) 
+    
+    This here is a preliminary state transition, PHI(t,t0),
+    matrix equation attempt for a ball rolling on the surface, based on...
+    
+    d PHI(t, t0)
+    ------------ =  Df(t) * PHI(t, t0)
+        dt
+    
+    """
+    
+    phi = PHI[0:16]
[…]
