### 2019-10-01 f323f5b
Minor updates
