+++ run_turningpoint_POfam_deleonberne.py
@@ -249,7 +249,7 @@ for i in range(len(E_vals)):
     x0_3 = optimize.newton(f2,-0.15)
     state0_3 = [x0_3,-0.05,0.0,0.0]
     
-    po_fam_file = open("x0_tpcd_deltaE%s_deleonberne.txt"%(deltaE),'a+')
+    po_fam_file = open("x0_turningpoint_deltaE%s_deleonberne.txt"%(deltaE),'a+')
     
     [x0po_1, T_1,energyPO_1] = tp_UPOsHam2dof.turningPoint(model, state0_2, state0_3, \
                                                             get_coord_deleonberne, \
@@ -273,7 +273,7 @@ for i in range(len(E_vals)):
     e = E_vals[i]
     deltaE = e - parameters[2]
 
-    po_fam_file = open("x0_tpcd_deltaE%s_deleonberne.txt" %(deltaE),'a+')
+    po_fam_file = open("x0_turningpoint_deltaE%s_deleonberne.txt" %(deltaE),'a+')
     print('Loading the periodic orbit family from data file',po_fam_file.name,'\n') 
     x0podata = np.loadtxt(po_fam_file.name)
     po_fam_file.close()
