### 2019-10-02 c37e075
Initialize contributing.md
