+++ docs/contributing.md
@@ -1,0 +1,39 @@
+# Contributing to UPOsHam
+
+Thank you for considering and taking the time to contribute!
+
+The following is a set of guidelines ([borrowed from the Atom project by Github](https://github.com/atom/atom/blob/master/CONTRIBUTING.md)) for contributing to this package. As always use your best judgment and feel free to propose changes to this document in a pull request.
+
+#### Table Of Contents
+
+[Code of Conduct](#code-of-conduct)
+
+[What should I know before I get started?](#what-should-i-know-before-i-get-started)
+  * [Unstable periodic orbits](#unstable-periodic-orbits)
+  * [Hamiltonian dynamics](#hamiltonian-dynamics)
+
+[How Can I Contribute?](#how-can-i-contribute)
+  * [Reporting Bugs](#reporting-bugs)
+  * [Suggesting Enhancements](#suggesting-enhancements)
+  * [Your First Code Contribution](#your-first-code-contribution)
+  * [Pull Requests](#pull-requests)
+
+[Styleguides](#styleguides)
+  * [Git Commit Messages](#git-commit-messages)
+  * [JavaScript Styleguide](#javascript-styleguide)
+  * [CoffeeScript Styleguide](#coffeescript-styleguide)
+  * [Specs Styleguide](#specs-styleguide)
+  * [Documentation Styleguide](#documentation-styleguide)
+
+[Additional Notes](#additional-notes)
+  * [Issue and Pull Request Labels](#issue-and-pull-request-labels)
+
+## Code of Conduct
+
+This project and everyone participating in it is governed by the [Contributor Covenant](CODE_OF_CONDUCT.md). 
+By participating, you are expected to uphold this code. Please report unacceptable behavior to 
+[shiba@vt.edu](mailto:shiba@vt.edu) or [wl16298@bristol.ac.uk](mailto:wl16298@bristol.ac.uk).
+
+
+
+
