### 2019-10-02 ad4cf57
Updating contributing outline
