### 2019-10-02 13b6bba
Updating contributing outline
