### 2019-10-02 b595296
Updating contributing outline
