+++ docs/contributing.md
@@ -1,7 +1,8 @@
 ---
-title: Contributing to UPOsHam
 bibliography: ../paper/paper.bib
 ---
+## Contributing to UPOsHam
+
 Thank you for considering and taking the time to contribute!
 
 The following is a set of guidelines ([borrowed from the Atom project by Github](https://github.com/atom/atom/blob/master/CONTRIBUTING.md)) for contributing to this package. As always use your best judgment and feel free to propose changes to this document in a pull request.
@@ -15,15 +16,7 @@ The following is a set of guidelines ([borrowed from the Atom project by Github]
 [How Can I Contribute?](#how-can-i-contribute)
   * [Reporting Bugs](#reporting-bugs)
   * [Suggesting Enhancements](#suggesting-enhancements)
-  * [Your First Code Contribution](#your-first-code-contribution)
   * [Pull Requests](#pull-requests)
-
-[Styleguides](#styleguides)
-  * [Git Commit Messages](#git-commit-messages)
-  * [JavaScript Styleguide](#javascript-styleguide)
-  * [CoffeeScript Styleguide](#coffeescript-styleguide)
-  * [Specs Styleguide](#specs-styleguide)
-  * [Documentation Styleguide](#documentation-styleguide)
 
 [Additional Notes](#additional-notes)
   * [Issue and Pull Request Labels](#issue-and-pull-request-labels)
@@ -36,7 +29,7 @@ By participating, you are expected to uphold this code. Please report unacceptab
 
 ## What should I know before I get started?
 
-We recommend familiarity with unstable periodic orbits and their role in Hamiltonian dynamics. Good starting point on this topic are @Koon2011 and @wiggins_role_2016.
+We recommend familiarity with unstable periodic orbits and their role in Hamiltonian dynamics. Good starting point on this topic are the @Koon2011 and @wiggins_role_2016.
 
 
 ## How Can I Contribute?
