### 2019-10-02 6f95228
Updating contributing outline
