### 2019-10-02 5a4c4ac
Updating contributing outline
