### 2019-10-02 ccdb8b6
minor updates
