### 2019-10-03 7c69fdf
minor updates
