+++ run_tpcd_POfam_uncoupled.py
@@ -236,8 +236,8 @@ alpha = 1.0
 beta = 1.0
 omega = 1.0
 EPSILON_S = 0.0 #Energy of the saddle
-parameters = np.array([1,omega, EPSILON_S, alpha, beta,omega]);
-eqNum = 1;  
+parameters = np.array([1,omega, EPSILON_S, alpha, beta,omega])
+eqNum = 1  
 model = 'uncoupled'
 #eqPt = tpcd_UPOsHam2dof.get_eq_pts(eqNum,model, parameters)
 eqPt = tpcd_UPOsHam2dof.get_eq_pts(eqNum, init_guess_eqpt_uncoupled, \
