+++ run_turningpoint_POfam_coupled.py
@@ -193,8 +193,8 @@ EPSILON_S = 0.0 #Energy of the saddle
 alpha = 1.00
 beta = 1.00
 epsilon= 1e-1
-parameters = np.array([1,omega, EPSILON_S, alpha, beta,omega,epsilon]);
-eqNum = 1;  
+parameters = np.array([1,omega, EPSILON_S, alpha, beta,omega,epsilon])
+eqNum = 1  
 model = 'coupled'
 #eqPt = tp_UPOsHam2dof.get_eq_pts(eqNum,model, parameters)
 eqPt = tp_UPOsHam2dof.get_eq_pts(eqNum, init_guess_eqpt_coupled, \
@@ -258,7 +258,7 @@ for i in range(len(deltaE_vals)):
 
 
 
-#[x0po_1, T_1,energyPO_1] = tp_UPOsHam2dof.turningPoint(model,state0_2,state0_3 ,parameters,e,n,n_turn,po_fam_file) ; 
+#[x0po_1, T_1,energyPO_1] = tp_UPOsHam2dof.turningPoint(model,state0_2,state0_3 ,parameters,e,n,n_turn,po_fam_file)  
 
 #%% Load periodic orbit data from ascii files
     
@@ -291,7 +291,7 @@ for i in range(len(deltaE_vals)):
 #
 #
 #po_fam_file = open("1111x0_turningpoint_deltaE%s_coupled.txt" %(deltaE),'a+')
-#[x0po_2, T_2,energyPO_2] = tp_UPOsHam2dof.turningPoint(model,state0_2,state0_3 ,parameters,e,n,n_turn,po_fam_file) ; 
+#[x0po_2, T_2,energyPO_2] = tp_UPOsHam2dof.turningPoint(model,state0_2,state0_3 ,parameters,e,n,n_turn,po_fam_file)  
 #po_fam_file.close()
 #
 ##%%
@@ -312,7 +312,7 @@ for i in range(len(deltaE_vals)):
 #
 #
 #po_fam_file = open("1111x0_turningpoint_deltaE%s_coupled.txt" %(deltaE),'a+')
-#[x0po_3, T_3,energyPO_3] = tp_UPOsHam2dof.turningPoint(model,state0_2,state0_3 ,parameters,e,n,n_turn,po_fam_file) ; 
+#[x0po_3, T_3,energyPO_3] = tp_UPOsHam2dof.turningPoint(model,state0_2,state0_3 ,parameters,e,n,n_turn,po_fam_file)  
 #po_fam_file.close()
 #
 ##%%
@@ -333,7 +333,7 @@ for i in range(len(deltaE_vals)):
 #
 #
 #po_fam_file = open("1111x0_turningpoint_deltaE%s_coupled.txt" %(deltaE),'a+')
-#[x0po_4, T_4,energyPO_4] = tp_UPOsHam2dof.turningPoint(model,state0_2,state0_3 ,parameters,e,n,n_turn,po_fam_file) ; 
+#[x0po_4, T_4,energyPO_4] = tp_UPOsHam2dof.turningPoint(model,state0_2,state0_3 ,parameters,e,n,n_turn,po_fam_file)  
 #po_fam_file.close()
 #
 ##%%
@@ -354,42 +354,42 @@ for i in range(len(deltaE_vals)):
 #
 #
 #po_fam_file = open("1111x0_turningpoint_deltaE%s_coupled.txt" %(deltaE),'a+')
-#[x0po_5, T_5,energyPO_5] = tp_UPOsHam2dof.turningPoint(model,state0_2,state0_3 ,parameters,e,n,n_turn,po_fam_file) ; 
[…]
