+++ run_turningpoint_POfam_uncoupled.py
@@ -22,6 +22,10 @@ mpl.rcParams['mathtext.rm'] = 'serif'
 
 #% Begin problem specific functions
 def init_guess_eqpt_uncoupled(eqNum, par):
+    """This function returns the position of the equilibrium points with 
+        Saddle (EQNUM=1)
+        Centre (EQNUM=2,3)
+    """
     
     if eqNum == 1:
         x0 = [0, 0]
@@ -34,6 +38,8 @@ def init_guess_eqpt_uncoupled(eqNum, par):
 
 
 def grad_pot_uncoupled(x,par):
+    """This function returns the gradient of the potential energy function V(x,y)
+    """     
     
     dVdx = -par[3]*x[0]+par[4]*(x[0])**3
     dVdy = par[5]*x[1]
@@ -44,6 +50,9 @@ def grad_pot_uncoupled(x,par):
 
 
 def pot_energy_uncoupled(x, y, par):
+    """This function returns the potential energy function V(x,y)
+    """
+    
     return -0.5*par[3]*x**2+0.25*par[4]*x**4 +0.5*par[5]*y**2
 
 
@@ -51,6 +60,7 @@ def get_coord_uncoupled(x,y, E, par):
     """ this function returns the initial position of x/y-coordinate on the potential energy surface(PES) for a specific energy V.
     
     """
+    
     return -0.5*par[3]*x**2+0.25*par[4]*x**4 +0.5*par[5]*y**2 - E
 #    elif model== 'deleonberne':
 #        return par[3]*( 1 - math.e**(-par[4]*x) )**2 + 4*y**2*(y**2 - 1)*math.e**(-par[5]*par[4]*x) + par[2]-V
@@ -59,7 +69,6 @@ def get_coord_uncoupled(x,y, E, par):
 
 
 def varEqns_uncoupled(t,PHI,par):
-    
     """
     PHIdot = varEqns_uncoupled(t,PHI) 
     
