+++ run_comparison_coupled.py
@@ -34,6 +34,10 @@ mpl.rcParams['mathtext.rm'] = 'serif'
 
 #% Begin problem specific functions
 def init_guess_eqpt_coupled(eqNum, par):
+    """This function returns the position of the equilibrium points with 
+        Saddle (EQNUM=1)
+        Centre (EQNUM=2,3)
+    """
     
     if eqNum == 1:
         x0 = [0, 0]
@@ -45,6 +49,8 @@ def init_guess_eqpt_coupled(eqNum, par):
     return x0
 
 def grad_pot_coupled(x, par):
+    """This function returns the gradient of the potential energy function V(x,y)
+    """ 
     
     dVdx = (-par[3]+par[6])*x[0]+par[4]*(x[0])**3-par[6]*x[1]
     dVdy = (par[5]+par[6])*x[1]-par[6]*x[0]
@@ -54,11 +60,13 @@ def grad_pot_coupled(x, par):
     return F
 
 def pot_energy_coupled(x, y, par):
+    """This function returns the potential energy function V(x,y)
+    """
+    
     return -0.5*par[3]*x**2+0.25*par[4]*x**4 +0.5*par[5]*y**2+0.5*par[6]*(x-y)**2
 
 
 def varEqns_coupled(t,PHI,par):
-    
     """
     PHIdot = varEqns_coupled(t,PHI) 
     
@@ -68,7 +76,6 @@ def varEqns_coupled(t,PHI,par):
     d PHI(t, t0)
     ------------ =  Df(t) * PHI(t, t0)
         dt
-    
     """
     
     phi = PHI[0:16]
@@ -152,7 +159,7 @@ EPSILON_S = 0.0 #Energy of the saddle
 alpha = 1.00
 beta = 1.00
 epsilon= 1e-1
-parameters = np.array([1,omega, EPSILON_S, alpha, beta,omega,epsilon]);
+parameters = np.array([1,omega, EPSILON_S, alpha, beta,omega,epsilon])
 eqNum = 1 
 model = 'coupled'
 #eqPt = diffcorr_UPOsHam2dof.get_eq_pts(eqNum,model, parameters)
@@ -168,11 +175,11 @@ eSaddle = diffcorr_UPOsHam2dof.get_total_energy([eqPt[0],eqPt[1],0,0], pot_energ
 
 deltaE = 0.10
 eSaddle = 0.0 # energy of the saddle
-#po_fam_file = open("./data/1111x0_newmethod_deltaE%s_coupled.txt" %(deltaE),'a+');
-#po_fam_file = open("1111x0_tpcd_deltaE%s_coupled.txt" %(deltaE),'a+');
+#po_fam_file = open("./data/1111x0_newmethod_deltaE%s_coupled.txt" %(deltaE),'a+')
+#po_fam_file = open("1111x0_tpcd_deltaE%s_coupled.txt" %(deltaE),'a+')
 data_path = "./data/"
 po_fam_file = "x0_tpcd_deltaE%s_coupled.txt" %(deltaE)
-print('Loading the periodic orbit family from data file',po_fam_file,'\n'); 
+print('Loading the periodic orbit family from data file',po_fam_file,'\n') 
 #x0podata = np.loadtxt(po_fam_file.name)
 x0podata = np.loadtxt(data_path + po_fam_file)
 
@@ -180,9 +187,9 @@ x0podata = np.loadtxt(data_path + po_fam_file)
 x0po_1_tpcd = x0podata
 
 
[…]
