### 2019-10-03 #1684 @Shibabrat
We have updated the methods by separating the problem-specific functions. The demonstration script which loads the method as module is the starting point for a user to modify according to their problem of interest.
