### 2019-10-03 WyLyu/UPOsHam#5 @Shibabrat
We have added the community guidelines based on your suggestions.
