### 2019-10-04 #1684 @JoshKarpel
I poked around in the latest version, and I definitely like where this is heading! Now that I think we're on the right architectural path for a reusable package, I'll start making issues in the repository again to address specific issues.
