### 2019-10-06 c37150f
minor updates
