+++ run_turningpoint_POfam_coupled.py
@@ -29,7 +29,7 @@ def init_guess_eqpt_coupled(eqNum, par):
     
     if eqNum == 1:
         x0 = [0, 0]
-    if eqNum == 2:
+    elif eqNum == 2:
         x0 = [np.sqrt(par[3]-par[6]/par[4]),0] 
     elif eqNum == 3:
         x0 = [-np.sqrt(par[3]-par[6]/par[4]),0] 
