### 2019-10-07 20bf54f
Updating function documentation
