### 2019-10-07 e239655
Updating references and docs
