+++ paper/paper.bib
@@ -1,3 +1,16 @@
+@book{wiggins_introduction_2003,
+	address = {New York},
+	edition = {2nd ed},
+	series = {Texts in applied mathematics},
+	title = {Introduction to applied nonlinear dynamical systems and chaos},
+	isbn = {978-0-387-00177-7},
+	language = {en},
+	number = {2},
+	publisher = {Springer},
+	author = {Wiggins, Stephen},
+	year = {2003}}
+
+
 @book{Koon2011,
 	author = {Koon, W. S. and Lo, M. W. and Marsden, J. E. and Ross, 
 	S. D.},
@@ -9,7 +22,17 @@
 	pages = {327},
 	year = {2011}}
 
-@article{naik2019bfinding,
+
+@book{Parker_1989,
+	author = {Parker, T. S. and Chua, L. O.},
+	title = {Practical Numerical Algorithms for Chaotic Systems},
+	year = {1989},
+	isbn = {0-387-96689-7},
+	publisher = {Springer-Verlag New York, Inc.},
+	address = {New York, NY, USA}} 
+
+
+@article{naik_finding_2019b,
 	title = {Finding normally hyperbolic invariant manifolds in two and three degrees of freedom with H\'enon-Heiles-type potential},
 	author = {Naik, Shibabrat and Wiggins, Stephen},
 	journal = {Phys. Rev. E},
@@ -20,29 +43,20 @@
 	year = {2019},
 	month = {Aug},
 	publisher = {American Physical Society},
-	doi = {10.1103/PhysRevE.100.022204}
-}
+	doi = {10.1103/PhysRevE.100.022204}}
 
-@book{Parker_1989,
-	author = {Parker, T. S. and Chua, L. O.},
-	title = {Practical Numerical Algorithms for Chaotic Systems},
-	year = {1989},
-	isbn = {0-387-96689-7},
-	publisher = {Springer-Verlag New York, Inc.},
-	address = {New York, NY, USA}
-} 
 
 @article{wiggins_role_2016,
-	title = {The role of normally hyperbolic invariant manifolds ({NHIMS}) in the context of the phase space setting for chemical reaction dynamics},
+	title = {The role of normally hyperbolic invariant manifolds ({NHIMs}) in the context of the phase space setting for chemical reaction dynamics},
 	volume = {21},
 	number = {6},
 	journal = {Regular and Chaotic Dynamics},
 	author = {Wiggins, Stephen},
-	month = nov,
+	month = {Nov},
 	year = {2016},
 	pages = {621--638},
-	doi = {10.1134/S1560354716060034}
-}
+	doi = {10.1134/S1560354716060034}}
+
 
 @article{Pollak_1980,
 	title = {Classical transition state theory: {A} lower bound to the reaction probability},
@@ -52,19 +66,19 @@
 	author = {Pollak, Eli and Child, Mark S. and Pechukas, Philip},
 	year = {1980},
 	pages = {1669--1678},
-	doi = {10.1063/1.439276}
-}
+	doi = {10.1063/1.439276}}
+
 
 @article{Deleon_Berne_1981,
 	title = {Intramolecular rate process: Isomerization dynamics and the transition to chaos},
 	volume = {75},
 	number = {7},
 	journal = {The Journal of Chemical Physics},
-	author = {De Leon, Nelson  and Berne, B. J. },
+	author = {De Leon, Nelson  and Berne, B. J.},
 	year = {1981},
 	pages = {3495-3510},
-	doi = {10.1063/1.442459}
-}
+	doi = {10.1063/1.442459}}
+
 
 @article{DeLeon_Marston_1989,
[…]
