### 2019-10-08 WyLyu/UPOsHam#7 @JoshKarpel
Add documentation
One of the requirements for the review is that the package have documentation; I understand that a large part of the product is "self-documenting" Jupyter Notebooks, but I think there is great value in static documentation, especially now that parts of the code are more package-like and deserve meaningful documentation on their own, since users may not look at their source code.

Using Sphinx (https://www.sphinx-doc.org/en/master/), a popular Python documentation generator, combined with #2  (which I see is in progress!), will let you produce pretty HTML docs with minimal effort. You can mix LaTeX-style math into the docs (and the docstrings), which should make the math-to-code barrier lower.

Something like https://nbsphinx.readthedocs.io/en/0.4.3/ will let you render the example notebooks directly in the documentation, so that prospective users can look at the results without going through the entire installation procedure for Python and your package.

Hosting can be done for free through either https://readthedocs.org/ (automatic deployment!) or GitHub Pages.

(Part of https://github.com/openjournals/joss-reviews/issues/1684)
