+++ diffcorr_UPOsHam2dof.py
@@ -21,41 +21,41 @@ mpl.rcParams['mathtext.rm'] = 'serif'
 def get_eq_pts(eqNum, init_guess_eqpt_model, grad_pot_model, par):
     """
     Returns configuration space coordinates of the equilibrium points.
-    
-    get_eq_pts(eqNum, init_guess_eqpt_model, grad_pot_model, par) solves the 
-    coordinates of the equilibrium point for a Hamiltonian of the form kinetic 
-    energy (KE) + potential energy (PE). 
-    
+
+    get_eq_pts(eqNum, init_guess_eqpt_model, grad_pot_model, par) solves the
+    coordinates of the equilibrium point for a Hamiltonian of the form kinetic
+    energy (KE) + potential energy (PE).
+
     Parameters
     ----------
     eqNum : int
         1 is the saddle-center equilibrium point
         2 or 3 is the center-center equilibrium point
-    
-    init_guess_eqpt_model : str
-        function name that returns the initial guess for the equilibrium point
-    
-    grad_pot_model : str
-        function name that defines the vector of potential gradient
-    
+
+    init_guess_eqpt_model : function name
+        function that returns the initial guess for the equilibrium point
+
+    grad_pot_model : function name
+        function that defines the vector of potential gradient
+
     par : float (list)
-        model parameters 
-   
+        model parameters
+
     Returns
     -------
     float (list)
         configuration space coordinates
-        
+
     """
     # Fix the equilibrium point numbering convention here and make a starting guess at the solution
     x0 = init_guess_eqpt_model(eqNum, par)
-    
+
     # F(xEq) = 0 at the equilibrium point, solve using in-built function
     # F = lambda x: func_vec_field_eq_pt(model,x,par)
     F = lambda x: grad_pot_model(x, par)
-    
+
     eqPt = fsolve(F, x0, fprime = None) # Call solver
-    
+
     return eqPt
 
 
@@ -63,64 +63,64 @@ def get_eq_pts(eqNum, init_guess_eqpt_model, grad_pot_model, par):
 def get_total_energy(orbit, pot_energy_model, parameters):
     """
     Returns total energy (value of Hamiltonian) of a phase space point on an orbit
-    
-    get_total_energy(orbit, pot_energy_model, parameters) returns the total energy for a 
-    Hamiltonian of the form KE + PE. 
-    
+
+    get_total_energy(orbit, pot_energy_model, parameters) returns the total energy for a
+    Hamiltonian of the form KE + PE.
+
     Parameters
     ----------
     orbit : float (list)
         phase space coordinates (x,y,px,py) of a point on an orbit
-    
-    pot_energy_model : str
-        function name that returns the potential energy of Hamiltonian
-    
+
+    pot_energy_model : function name
+        function that returns the potential energy of Hamiltonian
+
     parameters : float (list)
-        model parameters 
-   
+        model parameters
+
     Returns
     -------
     scalar
         total energy (value of Hamiltonian)
-    
+
     """
 
     x  = orbit[0]
     y  = orbit[1]
     px = orbit[2]
     py = orbit[3]
-    
-      
+
+
     return (1.0/(2*parameters[0]))*(px**2.0) + (1.0/(2*parameters[1]))*(py**2.0) + \
-            pot_energy_model(x, y, parameters)   
+            pot_energy_model(x, y, parameters)
 
 
 #%%
-def get_pot_surf_proj(xVec, yVec, pot_energy_model, par):            
+def get_pot_surf_proj(xVec, yVec, pot_energy_model, par):
     """
[…]
