+++ run_diffcorr_POfam_uncoupled.py
@@ -21,9 +21,10 @@ mpl.rcParams['mathtext.rm'] = 'serif'
 
 #% Begin problem specific functions
 def init_guess_eqpt_uncoupled(eqNum, par):
-    """This function returns the position of the equilibrium points with 
-        Saddle (EQNUM=1)
-        Centre (EQNUM=2,3)
+    """
+    Returns configuration space coordinates of the equilibrium points according to the index:
+    Saddle (EQNUM=1)
+    Centre (EQNUM=2,3)
     """
     
     if eqNum == 1:
@@ -37,8 +38,7 @@ def init_guess_eqpt_uncoupled(eqNum, par):
 
 
 def grad_pot_uncoupled(x,par):
-    """This function returns the gradient of the potential energy function V(x,y)
-    """ 
+    """ Returns the gradient of the potential energy function V(x,y) """ 
     
     dVdx = -par[3]*x[0]+par[4]*(x[0])**3
     dVdy = par[5]*x[1]
@@ -49,18 +49,14 @@ def grad_pot_uncoupled(x,par):
 
 
 def pot_energy_uncoupled(x, y, par):
-    """This function returns the potential energy function V(x,y)
+    """ Returns the potential energy function V(x,y)
     """
     
     return -0.5*par[3]*x**2+0.25*par[4]*x**4 +0.5*par[5]*y**2
 
 
 def eigvector_uncoupled(par):
-    """
-    eigenvectors and eigenvalues of the Jacobian evaluated at the equilibrium point, 
-    which is the correction of the initial condition.
-    check the result obtained from the Jacobian matches the analytic result.
-    """
+    """ Returns the correction factor to the eigenvectors for the linear guess """
 
     correcx = 1
     correcy = 1
@@ -69,16 +65,15 @@ def eigvector_uncoupled(par):
 
 
 def guess_lin_uncoupled(eqPt, Ax, par):
-    """This function returns an initial guess of the UPO.
-    """ 
+    """ Returns an initial guess for the unstable periodic orbit """ 
     
     correcx, correcy = eigvector_uncoupled(par)
     
-
     return [eqPt[0] + Ax*correcx,eqPt[1] + Ax*correcy,0,0]
 
 
 def jacobian_uncoupled(eqPt, par):
+    """ Returns Jacobian of the Hamiltonian vector field """
     
     x,y,px,py = eqPt[0:4]
     
@@ -104,11 +99,9 @@ def jacobian_uncoupled(eqPt, par):
 
 
 def varEqns_uncoupled(t,PHI,par):
-    """
-    PHIdot = varEqns_uncoupled(t,PHI) 
-    
-    This here is a preliminary state transition, PHI(t,t0),
-    matrix equation attempt for a ball rolling on the surface, based on...
+    """    
+    Returns the state transition matrix , PHI(t,t0), where Df(t) is the Jacobian of the 
+    Hamiltonian vector field
     
     d PHI(t, t0)
     ------------ =  Df(t) * PHI(t, t0)
@@ -154,6 +147,11 @@ def varEqns_uncoupled(t,PHI,par):
 
 
 def diffcorr_setup_uncoupled():
+    """ Returns settings for differential correction method 
+        
+        Settings include choosing coordinates for event criteria, convergence criteria, and 
+        correction (see references for details on how to choose these coordinates).
+    """
     
     dxdot1 = 1
     correctx0= 0
@@ -170,10 +168,11 @@ def conv_coord_uncoupled(x1, y1, dxdot1, dydot1):
 
 
 def diffcorr_acc_corr_uncoupled(coords, phi_t1, x0, par):
-    """This function computes the correction terms to the initial guess of the UPO
-    where correcx0 is the correction in x coodinate.
-    Using correction in x or y coordinates depend on the problem.
[…]
