+++ docs/index.rst
@@ -5,6 +5,8 @@
 
 Welcome to UPOsHam's documentation!
 ===================================
+UPOsHam is a Python package with a collection of three methods for computing unstable periodic orbits in Hamiltonian systems that model a diverse array of problems in physical sciences and engineering
+
 
 .. toctree::
    :maxdepth: 2
