+++ tp_UPOsHam2dof.py
@@ -29,17 +29,32 @@ mpl.rcParams['mathtext.rm'] = 'serif'
 
 def get_eq_pts(eqNum, init_guess_eqpt_model, grad_pot_model, par):
     """
-    GET_EQ_PTS solves the coordinates of the equilibrium point for a Hamiltonian 
-    system of the form KE + PE. 
-    
-    eqNum = 1 is the saddle-center equilibrium point
-    eqNum = 2 or 3 is the center-center equilibrium point
-    
-    init_guess_eqpt_model: function name that defines the initial guess for the critical point
-    
-    grad_pot_model: function name that defines the vector of potential gradient
-    
-    par: list of parameters for the problem
+    Returns configuration space coordinates of the equilibrium points.
+
+    get_eq_pts(eqNum, init_guess_eqpt_model, grad_pot_model, par) solves the
+    coordinates of the equilibrium point for a Hamiltonian of the form kinetic
+    energy (KE) + potential energy (PE).
+
+    Parameters
+    ----------
+    eqNum : int
+        1 is the saddle-center equilibrium point
+        2 or 3 is the center-center equilibrium point
+
+    init_guess_eqpt_model : function name
+        function that returns the initial guess for the equilibrium point
+
+    grad_pot_model : function name
+        function that defines the vector of potential gradient
+
+    par : float (list)
+        model parameters
+
+    Returns
+    -------
+    float (list)
+        configuration space coordinates
+
     """
     #fix the equilibrium point numbering convention here and make a
     #starting guess at the solution
@@ -59,7 +74,27 @@ def get_eq_pts(eqNum, init_guess_eqpt_model, grad_pot_model, par):
 #%
 def get_total_energy(orbit, pot_energy_model, parameters):
     """
-    get_total_energy computes the total energy of an input initial conditions
+    Returns total energy (value of Hamiltonian) of a phase space point on an orbit
+
+    get_total_energy(orbit, pot_energy_model, parameters) returns the total energy for a
+    Hamiltonian of the form KE + PE.
+
+    Parameters
+    ----------
+    orbit : float (list)
+        phase space coordinates (x,y,px,py) of a point on an orbit
+
+    pot_energy_model : function name
+        function that returns the potential energy of Hamiltonian
+
+    parameters : float (list)
+        model parameters
+
+    Returns
+    -------
+    scalar
+        total energy (value of Hamiltonian)
+
     """
     x  = orbit[0]
     y  = orbit[1]
@@ -72,7 +107,27 @@ def get_total_energy(orbit, pot_energy_model, parameters):
 
 #%%
 def get_pot_surf_proj(xVec, yVec, pot_energy_model, par):            
+    """
+    Returns projection of the potential energy (PE) surface on the configuration space
 
+    Parameters
+    ----------
+    xVec, yVec : 1d numpy arrays
+        x,y-coordinates that discretizes the x, y domain of the configuration space
+
+    pot_energy_model : function name
+        function that returns the potential energy of Hamiltonian
+
+    parameters : float (list)
+        model parameters
+
+    Returns
+    -------
+    2d numpy array
+        values of the PE
+
+    """
+    
     resX = np.size(xVec)
     resY = np.size(xVec)
     surfProj = np.zeros([resX, resY])
@@ -85,15 +140,47 @@ def get_pot_surf_proj(xVec, yVec, pot_energy_model, par):
     
 #%
 def stateTransitMat(tf,x0,par,varEqns_model,fixed_step=0): 
-    
     """
-    stateTransitMat(tf,x0,par,varEqns_model,fixed_step=0)
-    
-    Gets state transition matrix, phi_tf = PHI(0,tf), and the trajectory 
-    (x,t) for a length of time, tf.  
-    
[…]
