+++ run_diffcorr_POfam_deleonberne.py
@@ -21,9 +21,10 @@ mpl.rcParams['mathtext.rm'] = 'serif'
 
 #% Begin problem specific functions
 def init_guess_eqpt_deleonberne(eqNum, par):
-    """This function returns the position of the equilibrium points with 
-        Saddle (EQNUM=1)
-        Centre (EQNUM=2,3)
+    """
+    Returns configuration space coordinates of the equilibrium points according to the index:
+    Saddle (EQNUM=1)
+    Centre (EQNUM=2,3)
     """
     
     if eqNum == 1:
@@ -37,8 +38,8 @@ def init_guess_eqpt_deleonberne(eqNum, par):
 
 
 def grad_pot_deleonberne(x, par):
-    """This function returns the gradient of the potential energy function V(x,y)
-    """     
+    """ Returns the gradient of the potential energy function V(x,y) """
+     
     dVdx = -2*par[3]*par[4]*np.exp(-par[4]*x[0])*(np.exp(-par[4]*x[0]) - 1) - \
         4*par[5]*par[4]*x[1]**2*(x[1]**2 - 1)*np.exp(-par[5]*par[4]*x[0])
     dVdy = 8*x[1]*(2*x[1]**2 - 1)*np.exp(-par[5]*par[4]*x[0])
@@ -48,18 +49,13 @@ def grad_pot_deleonberne(x, par):
     return F
 
 def pot_energy_deleonberne(x, y, par):
-    """This function returns the potential energy function V(x,y)
-    """
+    """ Returns the potential energy function V(x,y) """
     
     return par[3]*( 1 - np.exp(-par[4]*x) )**2 + 4*y**2*(y**2 - 1)*np.exp(-par[5]*par[4]*x) + par[2]
 
 
 def eigvector_deleonberne(par):
-    """
-    eigenvectors and eigenvalues of the Jacobian evaluated at the equilibrium point, 
-    which is the correction of the initial condition.
-    check the result obtained from the Jacobian matches the analytic result.
-    """
+    """ Returns the correction factor to the eigenvectors for the linear guess """
 
     correcx = 1
     correcy = 0
@@ -68,8 +64,7 @@ def eigvector_deleonberne(par):
 
 
 def guess_lin_deleonberne(eqPt, Ax, par):
-    """This function returns an initial guess of the UPO.
-    """
+    """ Returns an initial guess for the unstable periodic orbit """
     
     correcx, correcy = eigvector_deleonberne(par)
     
@@ -78,6 +73,7 @@ def guess_lin_deleonberne(eqPt, Ax, par):
 
 
 def jacobian_deleonberne(eqPt, par):
+    """ Returns Jacobian of the Hamiltonian vector field """
     
     x,y,px,py = eqPt[0:4]
     # The first order derivative of the Hamiltonian.
@@ -102,11 +98,9 @@ def jacobian_deleonberne(eqPt, par):
 
 
 def varEqns_deleonberne(t,PHI,par):
-    """
-    PHIdot = varEqns_deleonberne(t,PHI) 
-    
-    This here is a preliminary state transition, PHI(t,t0),
-    matrix equation attempt for a ball rolling on the surface, based on...
+    """    
+    Returns the state transition matrix , PHI(t,t0), where Df(t) is the Jacobian of the 
+    Hamiltonian vector field
     
     d PHI(t, t0)
     ------------ =  Df(t) * PHI(t, t0)
@@ -153,6 +147,12 @@ def varEqns_deleonberne(t,PHI,par):
 
 
 def diffcorr_setup_deleonberne():
+    """ 
+    Returns settings for differential correction method 
+        
+    Settings include choosing coordinates for event criteria, convergence criteria, and 
+    correction (see references for details on how to choose these coordinates).
+    """
     
     dydot1 = 1
     correcty0= 0
[…]
