+++ run_diffcorr_POfam_coupled.py
@@ -20,9 +20,10 @@ mpl.rcParams['mathtext.rm'] = 'serif'
 
 #% Begin problem specific functions
 def init_guess_eqpt_coupled(eqNum, par):
-    """This function returns the position of the equilibrium points with 
-        Saddle (EQNUM=1)
-        Centre (EQNUM=2,3)
+    """
+    Returns configuration space coordinates of the equilibrium points according to the index:
+    Saddle (EQNUM=1)
+    Centre (EQNUM=2,3)
     """
     
     if eqNum == 1:
@@ -35,8 +36,8 @@ def init_guess_eqpt_coupled(eqNum, par):
     return x0
 
 def grad_pot_coupled(x, par):
-    """This function returns the gradient of the potential energy function V(x,y)
-    """     
+    """ Returns the gradient of the potential energy function V(x,y) """
+     
     dVdx = (-par[3]+par[6])*x[0]+par[4]*(x[0])**3-par[6]*x[1]
     dVdy = (par[5]+par[6])*x[1]-par[6]*x[0]
     
@@ -45,19 +46,15 @@ def grad_pot_coupled(x, par):
     return F
 
 def pot_energy_coupled(x, y, par):
-    """This function returns the potential energy function V(x,y)
-    """
+    """ Returns the potential energy function V(x,y) """
     
     return -0.5*par[3]*x**2+0.25*par[4]*x**4 +0.5*par[5]*y**2+0.5*par[6]*(x-y)**2
 
 
 #%
 def eigvector_coupled(par):
-    """
-     eigenvectors and eigenvalues of the Jacobian evaluated at the equilibrium point, 
-     which is the correction of the initial condition.
-     check the result obtained from the Jacobian matches the analytic result.
-     """
+    """ Returns the correction factor to the eigenvectors for the linear guess """
+    
     evaluelamb = np.sqrt(-0.5*(par[3]-par[6]-par[1]*(par[1]+par[6]) - np.sqrt(par[1]**4 + \
                                2*par[1]**3*par[6] + par[1]**2*(par[6]**2+2*par[3]-2*par[6]) + \
                                par[1]*( 2*par[6]**2 + 2*par[3]*par[6]) +(par[3]- par[6])**2)))
@@ -78,8 +75,7 @@ def eigvector_coupled(par):
 
 
 def guess_lin_coupled(eqPt, Ax, par):
-    """This function returns an initial guess of the UPO.
-    """
+    """ Returns an initial guess for the unstable periodic orbit """ 
     
     correcx, correcy = eigvector_coupled(par)
     
@@ -88,6 +84,7 @@ def guess_lin_coupled(eqPt, Ax, par):
 
 
 def jacobian_coupled(eqPt, par):
+    """ Returns Jacobian of the Hamiltonian vector field """
     
     x,y,px,py = eqPt[0:4]
     
@@ -113,12 +110,9 @@ def jacobian_coupled(eqPt, par):
 
 
 def varEqns_coupled(t,PHI,par):
-    
-    """
-    PHIdot = varEqns_coupled(t,PHI) 
-    
-    This here is a preliminary state transition, PHI(t,t0),
-    matrix equation attempt for a ball rolling on the surface, based on...
+    """    
+    Returns the state transition matrix , PHI(t,t0), where Df(t) is the Jacobian of the 
+    Hamiltonian vector field
     
     d PHI(t, t0)
     ------------ =  Df(t) * PHI(t, t0)
@@ -164,6 +158,12 @@ def varEqns_coupled(t,PHI,par):
 
 
 def diffcorr_setup_coupled():
+    """ 
+    Returns settings for differential correction method 
+        
+    Settings include choosing coordinates for event criteria, convergence criteria, and 
+    correction (see references for details on how to choose these coordinates).
+    """
     
[…]
