### 2019-10-16 #1684 @dpsanders
I suggest that all future changes to the repo are done via Pull Requests, so that we can easily see what has been changed.
