### 2019-10-16 #1684 @dpsanders
(Each Pull Request should ideally address a single, or a few related, points from the reviewers.)
