### 2019-10-16 #1684 @dpsanders
The suggestions made by @JoshKarpel about the code structure are spot on. There is still a lot of duplication between files like `run_diffcorr_POfam_coupled.py` and `run_turningpoint_POfam_coupled.py` that start out by defining exactly the same functions (as far as I can tell).

This makes it difficult to know if any copy-paste errors have been introduced and difficult to identify the basic structure and key parts of the code.

All of this code duplication should be eliminated.
