### 2019-10-17 WyLyu/UPOsHam#15 @JoshKarpel
Use standard name for requirements file
It looks like the requirements file got renamed to `requirements.dat` instead of `requirements.txt` in https://github.com/WyLyu/UPOsHam/commit/f643659977e6255d11f9df52a3ad3808cc549663. The `.txt` extension is standard for this file, and people will find it surprising if it isn't named that.

(Part of https://github.com/openjournals/joss-reviews/issues/1684)
