+++ diffcorr_UPOs_coupled.py
@@ -18,9 +18,8 @@ from matplotlib import cm
 mpl.rcParams['mathtext.fontset'] = 'cm'
 mpl.rcParams['mathtext.rm'] = 'serif'
 
-
 #% Begin problem specific functions
-def init_guess_eqpt_deleonberne(eqNum, par):
+def init_guess_eqpt_coupled(eqNum, par):
     """
     Returns configuration space coordinates of the equilibrium points according to the index:
     Saddle (EQNUM=1)
@@ -30,63 +29,76 @@ def init_guess_eqpt_deleonberne(eqNum, par):
     if eqNum == 1:
         x0 = [0, 0]
     elif eqNum == 2:
-        x0 = [0, 1/np.sqrt(2)]  # EQNUM = 2, center-center
+        x0 = [np.sqrt(par[3]-par[6]/par[4]),0] 
     elif eqNum == 3:
-        x0 = [0, -1/np.sqrt(2)] # EQNUM = 3, center-center
+        x0 = [-np.sqrt(par[3]-par[6]/par[4]),0] 
     
     return x0
 
-
-def grad_pot_deleonberne(x, par):
+def grad_pot_coupled(x, par):
     """ Returns the gradient of the potential energy function V(x,y) """
      
-    dVdx = -2*par[3]*par[4]*np.exp(-par[4]*x[0])*(np.exp(-par[4]*x[0]) - 1) - \
-        4*par[5]*par[4]*x[1]**2*(x[1]**2 - 1)*np.exp(-par[5]*par[4]*x[0])
-    dVdy = 8*x[1]*(2*x[1]**2 - 1)*np.exp(-par[5]*par[4]*x[0])
+    dVdx = (-par[3]+par[6])*x[0]+par[4]*(x[0])**3-par[6]*x[1]
+    dVdy = (par[5]+par[6])*x[1]-par[6]*x[0]
     
     F = [-dVdx, -dVdy]
     
     return F
 
-def pot_energy_deleonberne(x, y, par):
+def pot_energy_coupled(x, y, par):
     """ Returns the potential energy function V(x,y) """
     
-    return par[3]*( 1 - np.exp(-par[4]*x) )**2 + 4*y**2*(y**2 - 1)*np.exp(-par[5]*par[4]*x) + par[2]
+    return -0.5*par[3]*x**2+0.25*par[4]*x**4 +0.5*par[5]*y**2+0.5*par[6]*(x-y)**2
 
 
-def eigvector_deleonberne(par):
+#%
+def eigvector_coupled(par):
     """ Returns the correction factor to the eigenvectors for the linear guess """
-
-    correcx = 1
-    correcy = 0
+    
+    evaluelamb = np.sqrt(-0.5*(par[3]-par[6]-par[1]*(par[1]+par[6]) - np.sqrt(par[1]**4 + \
+                               2*par[1]**3*par[6] + par[1]**2*(par[6]**2+2*par[3]-2*par[6]) + \
+                               par[1]*( 2*par[6]**2 + 2*par[3]*par[6]) +(par[3]- par[6])**2)))
+#    correcx = par[6]/(-evaluelamb**2 -par[3]+par[6])
+#    correcy = 1
+    #
+    #
+    #eqPt = 1
+    #eqPt = get_eq_pts_coupled(eqNum, par)
+    #evalue, evector = np.linalg.eig(jacobian_coupled([eqPt[0],eqPt[1],0,0],par))
+    #evector = RemoveInfinitesimals(evector[:,2])
+    #correcx = (evector[0]*1j).real
+    #correcy = (evector[1]*1j).real
+    correcx = (par[1]*par[6])/(-evaluelamb**2 - par[3] + par[6])
+    correcy = par[1]
     
     return correcx, correcy
 
 
[…]
