+++ tpcd_UPOs_uncoupled.py
@@ -20,9 +20,10 @@ mpl.rcParams['mathtext.rm'] = 'serif'
 
 #% Begin problem specific functions
 def init_guess_eqpt_uncoupled(eqNum, par):
-    """This function returns the position of the equilibrium points with 
-        Saddle (EQNUM=1)
-        Centre (EQNUM=2,3)
+    """
+    Returns configuration space coordinates of the equilibrium points according to the index:
+    Saddle (EQNUM=1)
+    Centre (EQNUM=2,3)
     """
     
     if eqNum == 1:
@@ -36,7 +37,8 @@ def init_guess_eqpt_uncoupled(eqNum, par):
 
 
 def grad_pot_uncoupled(x,par):
-    """This function returns the gradient of the potential energy function V(x,y)
+    """
+    Returns the gradient of the potential energy function V(x,y)
     """  
     
     dVdx = -par[3]*x[0]+par[4]*(x[0])**3
@@ -48,14 +50,15 @@ def grad_pot_uncoupled(x,par):
 
 
 def pot_energy_uncoupled(x, y, par):
-    """This function returns the potential energy function V(x,y)
+    """Returns the potential energy function V(x,y)
     """
     
     return -0.5*par[3]*x**2+0.25*par[4]*x**4 +0.5*par[5]*y**2
 
 
 def get_coord_uncoupled(x,y, E, par):
-    """ this function returns the initial position of x/y-coordinate on the potential energy surface(PES) for a specific energy V.
+    """
+    Returns the initial position of x/y-coordinate on the potential energy surface(PES) for a specific energy V.
     """
     
     return -0.5*par[3]*x**2+0.25*par[4]*x**4 +0.5*par[5]*y**2 - E
@@ -66,11 +69,9 @@ def get_coord_uncoupled(x,y, E, par):
 
 
 def varEqns_uncoupled(t,PHI,par):
-    """
-    PHIdot = varEqns_uncoupled(t,PHI) 
-    
-    This here is a preliminary state transition, PHI(t,t0),
-    matrix equation attempt for a ball rolling on the surface, based on...
+    """    
+    Returns the state transition matrix , PHI(t,t0), where Df(t) is the Jacobian of the 
+    Hamiltonian vector field
     
     d PHI(t, t0)
     ------------ =  Df(t) * PHI(t, t0)
@@ -117,10 +118,10 @@ def varEqns_uncoupled(t,PHI,par):
 
 def configdiff_uncoupled(guess1, guess2, ham2dof_model, half_period_model, n_turn, par):
     """
-    configdiff_uncoupled(model,guess1, guess2,n_turn,par) returns the difference of x (or y) 
-    coordinates between the guess initial conditions and the ith turning points
-    n_turn is the nth turning point we want to choose as our 'turning point' for defining the 
-    dot product either difference in x coordintes(x_diff1, x_diff2) or difference in 
+    Returns the difference of x(or y) coordinates between the guess initial conditions 
+    and the ith turning points
+
+    either difference in x coordintes(x_diff1, x_diff2) or difference in 
     y coordinates(y_diff1, y_diff2) is returned as the result.
     """
     
@@ -169,8 +170,7 @@ def configdiff_uncoupled(guess1, guess2, ham2dof_model, half_period_model, n_tur
 
 
 def ham2dof_uncoupled(t, x, par):
-    """
-    Hamilton's equations of motion
+    """ Returns the Hamiltonian vector field (Hamilton's equations of motion)
     """
     
     xDot = np.zeros(4)
@@ -219,6 +219,10 @@ def guess_coords_uncoupled(guess1, guess2, i, n, e, get_coord_model, par):
     return xguess, yguess
     
[…]
