+++ examples/diffcorr_UPOs_deleonberne.py
@@ -12,6 +12,8 @@ import numpy as np
 import matplotlib.pyplot as plt
 from scipy.integrate import solve_ivp
 import time
+import sys
+sys.path.append('./src/')
 import diffcorr_UPOsHam2dof ### import module xxx where xxx is the name of the python file xxx.py 
 import matplotlib as mpl
 from matplotlib import cm
