+++ examples/diffcorr_UPOs_coupled.py
@@ -12,11 +12,14 @@ import numpy as np
 import matplotlib.pyplot as plt
 from scipy.integrate import solve_ivp
 import time
+import sys
+sys.path.append('./src/')
 import diffcorr_UPOsHam2dof ### import module xxx where xxx is the name of the python file xxx.py 
 import matplotlib as mpl
 from matplotlib import cm
 mpl.rcParams['mathtext.fontset'] = 'cm'
 mpl.rcParams['mathtext.rm'] = 'serif'
+
 
 #% Begin problem specific functions
 def init_guess_eqpt_coupled(eqNum, par):
