+++ README.Rmd
@@ -25,27 +25,30 @@ Clone/download the git repository using
 ```git
 $ git clone git@github.com:WyLyu/UPOsHam.git
 $ cd UPOsHam
-$ pip install -r requirements.dat
+$ pip install -r requirements.txt (or pip3 install -r requirements.txt)
 ```
-and check the modules shown in [requirements.dat](https://github.com/WyLyu/UPOsHam/tree/master/requirements.dat) are installed using conda/pip. Specific problems can be imported as modules for use in further analysis.
+and check the modules shown in [requirements.txt](https://github.com/WyLyu/UPOsHam/tree/master/requirements.txt) are installed using conda/pip. Specific problems can be imported as modules for use in further analysis.
 
 ## Usage
 
 __Comparison of the three methods for the coupled quartic Hamiltonian__
+
 Comparison of the results (data is located [here](https://github.com/WyLyu/UPOsHam/tree/master/data)) obtained using the three methods for the coupled quartic Hamiltonian can be done using
 
-```python
->> run run_comparison_coupled.py
+```
+>> ipython
+>> run ./tests/compare_methods_coupled.py
 ```
 and the generated figure is located [here](tests/comparison_coupled.pdf)
 
 To obtain the unstable periodic orbits for a specific model Hamiltonian using a specific method, one uses
 
-```python
->> run run_diffcorr_POfam_coupled.py
+```
+>> ipython
+>> run ./examples/diffcorr_UPOs_coupled.py
 ```
 
-Other tests can be performed by running the cells in the run_alltests.ipynb
+Other tests can be performed by running the cells in the ./tests/tests.ipynb
 
 ## Contributing
 
