+++ examples/tp_UPOs_uncoupled.py
@@ -245,7 +245,7 @@ for i in range(len(deltaE_vals)):
     state0_2 = [-0.1 , -math.sqrt(2*e+0.1**2-0.5*0.1**4),0.0,0.0]
     state0_3 = [0.11 , -math.sqrt(2*e+0.11**2-0.5*0.11**4),0.0,0.0]
     
-    po_fam_file = open("x0_turningpoint_deltaE%s_uncoupled.dat" %(deltaE),'a+')
+    po_fam_file = open("x0_turningpoint_deltaE%s_uncoupled.txt" %(deltaE),'a+')
     [x0po_1, T_1,energyPO_1] = tp_UPOsHam2dof.turningPoint(state0_2, state0_3, \
                                                             get_coord_uncoupled, \
                                                             guess_coords_uncoupled, \
@@ -274,7 +274,7 @@ x0po = np.zeros((4,len(deltaE_vals))) #each column is a different initial condit
 for i in range(len(deltaE_vals)):
     deltaE = deltaE_vals[i]
 
-    po_fam_file = open("x0_turningpoint_deltaE%s_uncoupled.dat" %(deltaE),'a+')
+    po_fam_file = open("x0_turningpoint_deltaE%s_uncoupled.txt" %(deltaE),'a+')
     print('Loading the periodic orbit family from data file',po_fam_file.name,'\n') 
     x0podata = np.loadtxt(po_fam_file.name)
     po_fam_file.close()
@@ -348,7 +348,7 @@ plt.savefig('turningpoint_POfam_uncoupled.pdf',format='pdf',bbox_inches='tight')
 #n=12
 #n_turn=1
 #deltaE = e-parameters[2]
-#po_fam_file = open("1111x0_turningpoint_deltaE%s_uncoupled.dat" %(deltaE),'a+')
+#po_fam_file = open("1111x0_turningpoint_deltaE%s_uncoupled.txt" %(deltaE),'a+')
 #[x0po_2, T_2,energyPO_2] = tp_UPOsHam2dof.turningPoint(model,state0_2,state0_3 ,parameters,e,n,n_turn,po_fam_file)  
 #
 #po_fam_file.close()
@@ -361,7 +361,7 @@ plt.savefig('turningpoint_POfam_uncoupled.pdf',format='pdf',bbox_inches='tight')
 #n=12
 #n_turn=1
 #deltaE = e-parameters[2]
-#po_fam_file = open("1111x0_turningpoint_deltaE%s_uncoupled.dat" %(deltaE),'a+')
+#po_fam_file = open("1111x0_turningpoint_deltaE%s_uncoupled.txt" %(deltaE),'a+')
 #[x0po_3, T_3,energyPO_3] = tp_UPOsHam2dof.turningPoint(model,state0_2,state0_3 ,parameters,e,n,n_turn,po_fam_file)  
 #
 #po_fam_file.close()
@@ -374,7 +374,7 @@ plt.savefig('turningpoint_POfam_uncoupled.pdf',format='pdf',bbox_inches='tight')
 #n=12
 #n_turn=1
 #deltaE = e-parameters[2]
-#po_fam_file = open("1111x0_turningpoint_deltaE%s_uncoupled.dat" %(deltaE),'a+')
+#po_fam_file = open("1111x0_turningpoint_deltaE%s_uncoupled.txt" %(deltaE),'a+')
 #[x0po_4, T_4,energyPO_4] = tp_UPOsHam2dof.turningPoint(model,state0_2,state0_3 ,parameters,e,n,n_turn,po_fam_file)  
 #
 #po_fam_file.close()
[…]
