+++ examples/diffcorr_UPOs_coupled.py
@@ -298,7 +298,7 @@ Ax2  = 2*Ax1 # initial amplitude (2 of 2)
 t = time.time()
 
 #  get the initial conditions and periods for a family of periodic orbits
-po_fam_file = open("x0_diffcorr_fam_eqPt%s_coupled.dat" %eqNum,'a+')
+po_fam_file = open("x0_diffcorr_fam_eqPt%s_coupled.txt" %eqNum,'a+')
 [po_x0Fam,po_tpFam] = diffcorr_UPOsHam2dof.get_POFam(eqNum, Ax1, Ax2, nFam, \
                                                     po_fam_file, init_guess_eqpt_coupled, \
                                                     grad_pot_coupled, jacobian_coupled, \
@@ -318,8 +318,8 @@ po_fam_file.close()
 #%%
 # begins with a family of periodic orbits and steps until crossing the
 # initial condition with target energy 
-# fileName = 'x0po_T_energy_case1_L41.dat'
-# fileName = 'x0po_T.dat'
+# fileName = 'x0po_T_energy_case1_L41.txt'
+# fileName = 'x0po_T.txt'
 
 #deltaE_vals = [0.01, 0.1, 1.00, 2.0, 4.0]
 #linecolor = ['b','r','g','m','c']
@@ -329,7 +329,7 @@ linecolor = ['b','r']
 for i in range(len(deltaE_vals)):
     deltaE = deltaE_vals[i]
     
-    po_fam_file = open("x0_diffcorr_fam_eqPt%s_coupled.dat" %eqNum ,'a+')
+    po_fam_file = open("x0_diffcorr_fam_eqPt%s_coupled.txt" %eqNum ,'a+')
     eTarget = eSaddle + deltaE 
     print('Loading the periodic orbit family from data file',po_fam_file.name,'\n') 
     x0podata = np.loadtxt(po_fam_file.name)
@@ -337,7 +337,7 @@ for i in range(len(deltaE_vals)):
     
     
     #%
-    po_brac_file = open("x0po_T_energyPO_eqPt%s_brac%s_coupled.dat" %(eqNum,deltaE),'a+')
+    po_brac_file = open("x0po_T_energyPO_eqPt%s_brac%s_coupled.txt" %(eqNum,deltaE),'a+')
     t = time.time()
     # [x0poTarget,TTarget] = bracket_POEnergy_bp(eTarget, x0podata, po_brac_file)
     x0poTarget,TTarget = diffcorr_UPOsHam2dof.poBracketEnergy(eTarget, x0podata, po_brac_file, \
@@ -350,7 +350,7 @@ for i in range(len(deltaE_vals)):
                                                               plot_iter_orbit_coupled, \
                                                               parameters)
     poTarE_runtime = time.time()-t
-    model_parameters_file = open("model_parameters_eqPt%s_DelE%s_coupled.dat" %(eqNum,deltaE),'a+')
+    model_parameters_file = open("model_parameters_eqPt%s_DelE%s_coupled.txt" %(eqNum,deltaE),'a+')
     np.savetxt(model_parameters_file.name, parameters,fmt='%1.16e')
     model_parameters_file.close()
     po_brac_file.close()
@@ -361,7 +361,7 @@ for i in range(len(deltaE_vals)):
     # Target PO of specific energy with high precision does not work for the
     # model 
     
-    po_target_file = open("x0_diffcorr_deltaE%s_coupled.dat" %(deltaE),'a+')
+    po_target_file = open("x0_diffcorr_deltaE%s_coupled.txt" %(deltaE),'a+')
                     
     [x0po, T,energyPO] = diffcorr_UPOsHam2dof.poTargetEnergy(x0poTarget,eTarget, \
[…]
