+++ requirements.txt
@@ -1,7 +1,7 @@
-scipy==1.1
-numpy==1.14
-matplotlib==2.2
-ipython==7.7
+scipy==1
+numpy==1
+matplotlib==2
+ipython==7
 
 
 
