### 2019-10-19 c6cbe9b
minor changes
