+++ paper/paper.bib
@@ -42,7 +42,6 @@
 	numpages = {14},
 	year = {2019},
 	month = {Aug},
-	publisher = {American Physical Society},
 	doi = {10.1103/PhysRevE.100.022204}}
 
 
