+++ README.Rmd
@@ -11,15 +11,13 @@ output:
     highlight: zenburn
     toc: yes
 bibliography: ./paper/paper.bib
-biblio-style: "apalike"
-link-citations: true
 ---
 
 [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.3373396.svg)](https://doi.org/10.5281/zenodo.3373396)
 
 ## Summary
 
-This Python package is a collection of three methods for computing unstable periodic orbits [@Parker_1989] in Hamiltonian systems that model a diverse array of problems in physical sciences and engineering [@wiggins_introduction_2003]. The unstable periodic orbits exist in the bottleneck of the equipotential line $V(x,y) = E$ and project as lines on the configuration space $(x,y)$ [@wiggins_role_2016]. The three methods [@Pollak_1980,@Koon2011,@naik_finding_2019b] described below have been implemented for three Hamiltonian systems of the form kinetic plus potential energy and a brief description can be found in the [paper.pdf](https://github.com/WyLyu/UPOsHam/tree/master/paper/paper.pdf). The scripts are written as demonstration of how to modify and adapt the code for a problem of interest. 
+This Python package is a collection of three methods for computing unstable periodic orbits in Hamiltonian systems that model a diverse array of problems in physical sciences and engineering [@wiggins_introduction_2003,@Parker_1989]. The unstable periodic orbits exist in the bottleneck of the equipotential line $V(x,y) = E$ and project as lines on the configuration space $(x,y)$ [@wiggins_role_2016]. The three methods described below have been implemented for three Hamiltonian systems of the form kinetic plus potential energy [@Pollak_1980,@Koon2011,@naik_finding_2019b]. A brief description of the methods and the applications can be found in the [paper.pdf](https://github.com/WyLyu/UPOsHam/tree/master/paper/paper.pdf). The scripts are written as demonstration of how to modify and adapt the code for a problem of interest. 
 
 ## Installation
 
