### 2019-10-20 #1684 @Shibabrat
We've mostly done the last few changes that include the filename, moving code to different directories, and structuring documentation using pull requests. But maybe we didn't tag or reference the issues so that's why it's not clear what we have been changing. Apologies for the confusion. We are cutting our teeth with the branching and pull requests on different but related issues.

In summary, we are working on the automated tests and removing the duplication in the scripts. These sets of changes will also address how to run the scripts without the figures interfering with the progress of the iteration.
