### 2019-10-20 WyLyu/UPOsHam#27 @dpsanders
Make filenames clearer
Files like `diffcorr_UPOsHam2dof.py` do not need to have `UPOsHam2dof` in them, since that's understood since it's in this repo.

On the other hand, I don't know what `diffcorr` and certainly not `tp` and `tpcd` mean, so you might want to expand those abbreviations a bit.

In the examples, I would like to see files like `diffcorr_UPOs_coupled.py` be replaced by `double_pendulum.py` or whatever, that just define the functions for each model. I understand you're working on this.

Part of JOSS review @ https://github.com/openjournals/joss-reviews/issues/1684
