### 2019-10-22 adad244
add options for not generating figures
