### 2019-10-22 29ee6bb
add options for not generating figures
