### 2019-10-22 8e78c41
Revert "Check usage instructions"
