### 2019-10-24 62b4e18
Update to the differential correction for coupled quartic Hamiltonian
