### 2019-10-26 1f12e0d
Revert "update function names"
