+++ examples/diffcorr_UPOs_uncoupled.py
@@ -14,7 +14,7 @@ from scipy.integrate import solve_ivp
 import time
 import sys
 sys.path.append('./src/')
-import differential_correction ### import module xxx where xxx is the name of the python file xxx.py 
+import diffcorr_UPOsHam2dof ### import module xxx where xxx is the name of the python file xxx.py 
 import matplotlib as mpl
 from matplotlib import cm
 mpl.rcParams['mathtext.fontset'] = 'cm'
@@ -256,12 +256,12 @@ epsilon= 1e-1
 parameters = np.array([1, omega, EPSILON_S, alpha, beta, omega, epsilon])
 eqNum = 1 
 model = 'uncoupled'
-#eqPt = differential_correction.get_eq_pts(eqNum,model, parameters)
-eqPt = differential_correction.get_eq_pts(eqNum, init_guess_eqpt_uncoupled, \
+#eqPt = diffcorr_UPOsHam2dof.get_eq_pts(eqNum,model, parameters)
+eqPt = diffcorr_UPOsHam2dof.get_eq_pts(eqNum, init_guess_eqpt_uncoupled, \
                                        grad_pot_uncoupled, parameters)
 
 #energy of the saddle eq pt
-eSaddle = differential_correction.get_total_energy([eqPt[0],eqPt[1],0,0], pot_energy_uncoupled, \
+eSaddle = diffcorr_UPOsHam2dof.get_total_energy([eqPt[0],eqPt[1],0,0], pot_energy_uncoupled, \
                                                 parameters) 
 
 #If orbit is an n-array, e.g. orbit = [orbit_0, orbit_1, ..., orbit_n]
@@ -282,7 +282,7 @@ t = time.time()
 
 # get the initial conditions and periods for a family of periodic orbits
 po_fam_file = open("x0_diffcorr_fam_eqPt%s_uncoupled.txt" %eqNum,'a+')
-[po_x0Fam,po_tpFam] = differential_correction.get_POFam(eqNum, Ax1, Ax2, nFam, \
+[po_x0Fam,po_tpFam] = diffcorr_UPOsHam2dof.get_POFam(eqNum, Ax1, Ax2, nFam, \
                                                     po_fam_file, init_guess_eqpt_uncoupled, \
                                                     grad_pot_uncoupled, jacobian_uncoupled, \
                                                     guess_lin_uncoupled, diffcorr_setup_uncoupled, \
@@ -321,7 +321,7 @@ for i in range(len(deltaE_vals)):
     po_brac_file = open("x0po_T_energyPO_eqPt%s_brac%s_uncoupled.txt" %(eqNum,deltaE),'a+')
     t = time.time()
     # [x0poTarget,TTarget] = bracket_POEnergy_bp(eTarget, x0podata, po_brac_file)
-    x0poTarget,TTarget = differential_correction.poBracketEnergy(eTarget, x0podata, po_brac_file, \
+    x0poTarget,TTarget = diffcorr_UPOsHam2dof.poBracketEnergy(eTarget, x0podata, po_brac_file, \
                                                               diffcorr_setup_uncoupled, \
                                                               conv_coord_uncoupled, \
                                                               diffcorr_acc_corr_uncoupled, \
@@ -344,7 +344,7 @@ for i in range(len(deltaE_vals)):
     
     po_target_file = open("x0_diffcorr_deltaE%s_uncoupled.txt" %(deltaE),'a+')
 
-    [x0po, T,energyPO] = differential_correction.poTargetEnergy(x0poTarget,eTarget, \
+    [x0po, T,energyPO] = diffcorr_UPOsHam2dof.poTargetEnergy(x0poTarget,eTarget, \
                                                         po_target_file, \
                                                         diffcorr_setup_uncoupled, \
                                                         conv_coord_uncoupled, \
[…]
