+++ examples/tp_UPOs_coupled.py
@@ -15,7 +15,7 @@ import math
 from scipy import optimize
 import sys
 sys.path.append('./src/')
-import turning_point ### import module xxx where xxx is the name of the python file xxx.py 
+import tp_UPOsHam2dof ### import module xxx where xxx is the name of the python file xxx.py 
 import matplotlib as mpl
 from matplotlib import cm
 mpl.rcParams['mathtext.fontset'] = 'cm'
@@ -141,7 +141,7 @@ def ham2dof_coupled(t, x, par):
 
 def half_period_coupled(t, x, par):
     """
-    Returns the turning point where we want to stop the integration                           
+    Return the turning point where we want to stop the integration                           
     
     pxDot = x[0]
     pyDot = x[1]
@@ -210,8 +210,8 @@ epsilon= 1e-1
 parameters = np.array([1,omega, EPSILON_S, alpha, beta,omega,epsilon])
 eqNum = 1  
 #model = 'coupled'
-#eqPt = turning_point.get_eq_pts(eqNum,model, parameters)
-eqPt = turning_point.get_eq_pts(eqNum, init_guess_eqpt_coupled, \
+#eqPt = tp_UPOsHam2dof.get_eq_pts(eqNum,model, parameters)
+eqPt = tp_UPOsHam2dof.get_eq_pts(eqNum, init_guess_eqpt_coupled, \
                                    grad_pot_coupled, parameters)
 
 #%%
@@ -258,7 +258,7 @@ for i in range(len(deltaE_vals)):
     
     
     po_fam_file = open("x0_turningpoint_deltaE%s_coupled.txt" %(deltaE),'a+')
-    [x0po_1, T_1,energyPO_1] = turning_point.turningPoint(state0_2, state0_3, \
+    [x0po_1, T_1,energyPO_1] = tp_UPOsHam2dof.turningPoint(state0_2, state0_3, \
                                                             get_coord_coupled, \
                                                             guess_coords_coupled, \
                                                             ham2dof_coupled, \
@@ -272,7 +272,7 @@ for i in range(len(deltaE_vals)):
 
 
 
-#[x0po_1, T_1,energyPO_1] = turning_point.turningPoint(model,state0_2,state0_3 ,parameters,e,n,n_turn,po_fam_file)  
+#[x0po_1, T_1,energyPO_1] = tp_UPOsHam2dof.turningPoint(model,state0_2,state0_3 ,parameters,e,n,n_turn,po_fam_file)  
 
 #%% Load periodic orbit data from ascii files
     
@@ -294,18 +294,18 @@ for i in range(len(deltaE_vals)):
 #deltaE = e-parameters[2]
 #"""Trial initial Condition s.t. one initial condition is on the LHS of the UPO and the other one is on the RHS of the UPO"""
 #x=-0.05
-#f2 = lambda y: turning_point.get_coordinate(model,0.0,y,e,parameters)
+#f2 = lambda y: tp_UPOsHam2dof.get_coordinate(model,0.0,y,e,parameters)
[…]

[…]
