### 2019-10-26 fd83a0f
Update usage of differential correction for coupled quartic Hamiltonian
