+++ examples/coupled_quartic_hamiltonian.py
@@ -1,0 +1,274 @@
+# -*- coding: utf-8 -*-
+"""
+Created on Tue Jul 30 10:02:48 2019
+
+@author: Wenyang Lyu and Shibabrat Naik
+
+Script to define expressions for the coupled quartic Hamiltonian
+"""
+
+import numpy as np
+import matplotlib.pyplot as plt
+from scipy.integrate import solve_ivp
+import time
+ 
+
+# import matplotlib as mpl
+# from matplotlib import cm
+# mpl.rcParams['mathtext.fontset'] = 'cm'
+# mpl.rcParams['mathtext.rm'] = 'serif'
+
+
+#% Begin problem specific functions
+def init_guess_eqpt_coupled(eqNum, par):
+    """
+    Returns configuration space coordinates of the equilibrium points according to the index:
+    Saddle (EQNUM=1)
+    Centre (EQNUM=2,3)
+    """
+    
+    if eqNum == 1:
+        x0 = [0, 0]
+    elif eqNum == 2:
+        x0 = [np.sqrt(par[3]-par[6]/par[4]),0] 
+    elif eqNum == 3:
+        x0 = [-np.sqrt(par[3]-par[6]/par[4]),0] 
+    
+    return x0
+
+def grad_pot_coupled(x, par):
+    """ Returns the gradient of the potential energy function V(x,y) """
+     
+    dVdx = (-par[3]+par[6])*x[0]+par[4]*(x[0])**3-par[6]*x[1]
+    dVdy = (par[5]+par[6])*x[1]-par[6]*x[0]
+    
+    F = [-dVdx, -dVdy]
+    
+    return F
+
+def pot_energy_coupled(x, y, par):
+    """ Returns the potential energy function V(x,y) """
+    
+    return -0.5*par[3]*x**2+0.25*par[4]*x**4 +0.5*par[5]*y**2+0.5*par[6]*(x-y)**2
+
+
+#%
+def eigvector_coupled(par):
+    """ Returns the correction factor to the eigenvectors for the linear guess """
+    
+    evaluelamb = np.sqrt(-0.5*(par[3]-par[6]-par[1]*(par[1]+par[6]) - np.sqrt(par[1]**4 + \
+                               2*par[1]**3*par[6] + par[1]**2*(par[6]**2+2*par[3]-2*par[6]) + \
+                               par[1]*( 2*par[6]**2 + 2*par[3]*par[6]) +(par[3]- par[6])**2)))
+#    correcx = par[6]/(-evaluelamb**2 -par[3]+par[6])
+#    correcy = 1
+    #
+    #
+    #eqPt = 1
+    #eqPt = get_eq_pts_coupled(eqNum, par)
+    #evalue, evector = np.linalg.eig(jacobian_coupled([eqPt[0],eqPt[1],0,0],par))
+    #evector = RemoveInfinitesimals(evector[:,2])
+    #correcx = (evector[0]*1j).real
+    #correcy = (evector[1]*1j).real
+    correcx = (par[1]*par[6])/(-evaluelamb**2 - par[3] + par[6])
+    correcy = par[1]
+    
+    return correcx, correcy
+
+
+def guess_lin_coupled(eqPt, Ax, par):
+    """ Returns an initial guess for the unstable periodic orbit """ 
+    
+    correcx, correcy = eigvector_coupled(par)
+    
+
+    return [eqPt[0]-Ax*correcx,eqPt[1]-Ax*correcy,0,0]
+
+
+def jacobian_coupled(eqPt, par):
+    """ Returns Jacobian of the Hamiltonian vector field """
+    
+    x,y,px,py = eqPt[0:4]
+    
+    # The first order derivative of the Hamiltonian.
[…]
