### 2019-10-26 bbe79ef
Revert "Update usage of differential correction for coupled quartic Hamiltonian"
This reverts commit fd83a0f62e63e0130ca6567d6f7ba18babdb520a.
