### 2019-10-26 11665b3
Update usage of differential correction for coupled quartic hamiltonian
