### 2019-10-26 838989b
Updates to the turning point coordinate differnce for coupled quartic
