+++ examples/coupled_quartic_hamiltonian.py
@@ -9,6 +9,8 @@ import numpy as np
 import matplotlib.pyplot as plt
 from scipy.integrate import solve_ivp
 import time
+import math
+from scipy import optimize
  
 
 # import matplotlib as mpl
@@ -178,6 +180,21 @@ def conv_coord_coupled(x1, y1, dxdot1, dydot1):
     return dxdot1
 
 
+def get_coord_coupled(x, y, E, par):
+    """ 
+    Returns the initial position of x/y-coordinate on the potential energy 
+    surface(PES) for a specific energy E.
+    """
+#    if model == 'uncoupled':
+#        return -0.5*par[3]*x**2+0.25*par[4]*x**4 +0.5*par[5]*y**2-V
+#    elif model =='coupled':
+    return -0.5*par[3]*x**2+0.25*par[4]*x**4 +0.5*par[5]*y**2+0.5*par[6]*(x-y)**2 - E
+#    elif model== 'deleonberne':
+#        return par[3]*( 1 - math.e**(-par[4]*x) )**2 + 4*y**2*(y**2 - 1)*math.e**(-par[5]*par[4]*x) + par[2]-V
+#    else:
+#        print("The model you are chosen does not exist, enter the function for finding coordinates on the PES for given x or y and V")
+
+
 def diffcorr_acc_corr_coupled(coords, phi_t1, x0, par):
     """ 
     Returns the new guess initial condition of the unstable periodic orbit after applying 
@@ -199,6 +216,69 @@ def diffcorr_acc_corr_coupled(coords, phi_t1, x0, par):
     x0[0] = x0[0] - correctx0
     
     return x0
+
+
+def configdiff_coupled(guess1, guess2, ham2dof_model, half_period_model, n_turn, par):
+    """
+    Returns the difference of x(or y) coordinates between the guess initial conditions 
+    and the ith turning points
+
+    either difference in x coordintes(x_diff1, x_diff2) or difference in 
+    y coordinates(y_diff1, y_diff2) is returned as the result.
+    """
+    
+    TSPAN = [0,40]
+    RelTol = 3.e-10
+    AbsTol = 1.e-10 
+    
+    f1 = lambda t,x: ham2dof_model(t,x,par) 
+    soln1 = solve_ivp(f1, TSPAN, guess1, method='RK45', dense_output=True, \
+                      events = lambda t,x: half_period_model(t, x, par), rtol=RelTol, atol=AbsTol)
+    te1 = soln1.t_events[0]
+    t1 = [0,te1[n_turn]]#[0,te1[1]]
+    turn1 = soln1.sol(t1)
+    x_turn1 = turn1[0,-1] 
+    y_turn1 = turn1[1,-1]
+    x_diff1 = guess1[0] - x_turn1
+    y_diff1 = guess1[1] - y_turn1
+    
+    f2 = lambda t,x: ham2dof_model(t,x,par) 
+    soln2 = solve_ivp(f2, TSPAN, guess2,method='RK45', dense_output=True, \
+                      events = lambda t,x: half_period_model(t, x, par), rtol=RelTol, atol=AbsTol)
+    te2 = soln2.t_events[0]
+    t2 = [0,te2[n_turn]]#[0,te2[1]]
+    turn2 = soln2.sol(t2)
+    x_turn2 = turn2[0,-1] 
+    y_turn2 = turn2[1,-1] 
+    x_diff2 = guess2[0] - x_turn2
+    y_diff2 = guess2[1] - y_turn2
+    
+
+    print("Initial guess1%s, initial guess2 %s, x_diff1 is %s, x_diff2 is %s " %(guess1, \
[…]
