### 2019-10-28 39d5ba0
Update turning point for coupled quartic hamiltonian
