+++ examples/turning_pt_coupled.py
@@ -1,0 +1,398 @@
+# -*- coding: utf-8 -*-
+"""
+Created on Tue Jul 30 10:02:48 2019
+
+@author: Wenyang Lyu and Shibabrat Naik
+
+Compute unstable peridoic orbits at different energies using turning point method
+"""
+
+# For the coupled problem
+import numpy as np
+import matplotlib.pyplot as plt
+from scipy.integrate import solve_ivp
+import math
+from scipy import optimize
+
+import matplotlib as mpl
+from matplotlib import cm
+mpl.rcParams['mathtext.fontset'] = 'cm'
+mpl.rcParams['mathtext.rm'] = 'serif'
+
+# This needs testing for installation 
+import sys
+sys.path.insert(0, './src/')
+import turning_point as tp
+import coupled_quartic_hamiltonian as coupled
+# This needs testing for installation
+
+
+#%% Setting up parameters and global variables
+
+save_final_plot = True
+show_final_plot = False
+show_itrsteps_plots = False # show iteration of the UPOs in plots
+N = 4          # dimension of phase space
+omega=1.00
+EPSILON_S = 0.0 #Energy of the saddle
+alpha = 1.00
+beta = 1.00
+epsilon= 1e-1
+parameters = np.array([1,omega, EPSILON_S, alpha, beta,omega,epsilon])
+eqNum = 1  
+#model = 'coupled'
+#eqPt = tp.get_eq_pts(eqNum,model, parameters)
+eqPt = tp.get_eq_pts(eqNum, coupled.init_guess_eqpt_coupled, \
+                            coupled.grad_pot_coupled, parameters)
+
+#%%
+#deltaE_vals = [0.01, 0.1, 1.00, 2.0, 4.0]
+#xLeft = [0.0,-0.05,0.01,0.01,0.18]
+#xRight = [0.05,0.10,0.18,0.18,0.18]
+#linecolor = ['b','r','g','m','c']
+deltaE_vals = [0.1, 1.00]
+xLeft = [0.0,0.01]
+xRight = [0.05,0.11]
+linecolor = ['b','r']
+
+"""
+e is the total energy, 
+n is the number of intervals we want to divide, 
+n_turn is the nth turning point we want to choose.
+"""
+for i in range(len(deltaE_vals)):
+    
+    e = deltaE_vals[i]
+    n = 12
+    n_turn = 1
+    deltaE = e - parameters[2]
+    
+    """
+    Trial initial Condition s.t. one initial condition is on the LHS of the UPO and the 
+    other one is on the RHS of the UPO
+    """
+    x = xLeft[i]
+    f2 = lambda y: coupled.get_coord_coupled(x, y, e, parameters)
+    yanalytic = math.sqrt(2/(parameters[1]+parameters[6]))*(-math.sqrt( e + \
+                         0.5*parameters[3]* x**2- 0.25*parameters[4]*x**4 - 0.5*parameters[6]*x**2 + \
+                         (parameters[6]*x)**2/(2*(parameters[1] + parameters[6]) )) + \
+                            parameters[6]/(math.sqrt(2*(parameters[1]+parameters[6])) )*x ) #coupled
+    state0_2 = [x,optimize.newton(f2,yanalytic),0.0,0.0]
+    
+    x = xRight[i]
[…]
