### 2019-10-28 a436470
Update turning point for coupled quartic hamiltonian
