+++ examples/coupled_quartic_hamiltonian.py
@@ -344,6 +344,183 @@ def half_period_coupled(t, x, par):
 #     Returns configuration space coordinates of the equilibrium points according to the index:
 #     Saddle (EQNUM=1)
 #     Centre (EQNUM=2,3)
+#     """ 
+    
+#     if eqNum == 1:
+#         x0 = [0, 0]
+#     elif eqNum == 2:
+#         x0 = [np.sqrt(par[3]-par[6]/par[4]),0] 
+#     elif eqNum == 3:
+#         x0 = [-np.sqrt(par[3]-par[6]/par[4]),0] 
+    
+#     return x0
+
+# def grad_pot_coupled(x, par):
+#     """
+#     Returns the gradient of the potential energy function V(x,y)
+#     """ 
+    
+#     dVdx = (-par[3]+par[6])*x[0]+par[4]*(x[0])**3-par[6]*x[1]
+#     dVdy = (par[5]+par[6])*x[1]-par[6]*x[0]
+            
+#     F = [-dVdx, -dVdy]
+    
+#     return F
+
+# def pot_energy_coupled(x, y, par):
+#     """Returns the potential energy function V(x,y)
+#     """
+    
+#     return -0.5*par[3]*x**2+0.25*par[4]*x**4 +0.5*par[5]*y**2+0.5*par[6]*(x-y)**2
+
+
+# def get_coord_coupled(x, y, E, par):
+#     """ 
+#     Returns the initial position of x/y-coordinate on the potential energy 
+#     surface(PES) for a specific energy E.
+#     """
+    
+# #    if model == 'uncoupled':
+# #        return -0.5*par[3]*x**2+0.25*par[4]*x**4 +0.5*par[5]*y**2-V
+# #    elif model =='coupled':
+#     return -0.5*par[3]*x**2+0.25*par[4]*x**4 +0.5*par[5]*y**2+0.5*par[6]*(x-y)**2 - E
+# #    elif model== 'deleonberne':
+# #        return par[3]*( 1 - math.e**(-par[4]*x) )**2 + 4*y**2*(y**2 - 1)*math.e**(-par[5]*par[4]*x) + par[2]-V
+# #    else:
+# #        print("The model you are chosen does not exist, enter the function for finding coordinates on the PES for given x or y and V")
+
+
+# def varEqns_coupled(t,PHI,par):
+#     """    
+#     Returns the state transition matrix , PHI(t,t0), where Df(t) is the Jacobian of the 
+#     Hamiltonian vector field
+    
+#     d PHI(t, t0)
+#     ------------ =  Df(t) * PHI(t, t0)
+#         dt
+    
+#     """
+    
+#     phi = PHI[0:16]
+#     phimatrix  = np.reshape(PHI[0:16],(4,4))
+#     x,y,px,py = PHI[16:20]
+    
+    
+#     # The first order derivative of the potential energy.
+#     dVdx = (-par[3]+par[6])*x+par[4]*x**3-par[6]*y
+#     dVdy = (par[5]+par[6])*y-par[6]*x
+
+#     # The second order derivative of the potential energy. 
+#     d2Vdx2 = -par[3]+par[6]+par[4]*3*x**2
+        
+#     d2Vdy2 = par[5]+par[6]
+
+#     d2Vdydx = -par[6]
+
+    
+#     d2Vdxdy = d2Vdydx    
+
+#     Df    = np.array([[  0,     0,    par[0],    0],
+#               [0,     0,    0,    par[1]],
+#               [-d2Vdx2,  -d2Vdydx,   0,    0],
+#               [-d2Vdxdy, -d2Vdy2,    0,    0]])
+
+    
+#     phidot = np.matmul(Df, phimatrix) # variational equation
+
[…]
