### 2019-11-03 06fef92
Turning point for coupled quartic Hamiltonian
