### 2019-11-04 8a2de24
Update methods for uncoupled quartic Hamiltonian
