+++ examples/turning_pt_uncoupled.py
@@ -1,0 +1,365 @@
+
+# -*- coding: utf-8 -*-
+"""
+Created on Tue Jul 30 10:02:48 2019
+
+@author: Wenyang Lyu and Shibabrat Naik
+
+Compute unstable peridoic orbits at different energies using turning point method
+"""
+
+# For the coupled problem
+import numpy as np
+import matplotlib.pyplot as plt
+from scipy.integrate import solve_ivp
+import math
+from scipy import optimize
+import sys
+
+import matplotlib as mpl
+from matplotlib import cm
+mpl.rcParams['mathtext.fontset'] = 'cm'
+mpl.rcParams['mathtext.rm'] = 'serif'
+
+# This needs testing for installation 
+import sys
+import importlib
+sys.path.insert(0, './src/')
+
+import turning_point
+importlib.reload(turning_point)
+import turning_point as tp
+
+import uncoupled_quartic_hamiltonian
+importlib.reload(uncoupled_quartic_hamiltonian)
+import uncoupled_quartic_hamiltonian as uncoupled
+# This needs testing for installation
+
+
+#%% Setting up parameters and global variables
+
+save_final_plot = True
+show_final_plot = False
+show_itrsteps_plots = False # show iteration of the UPOs in plots
+N = 4          # dimension of phase space
+alpha = 1.0
+beta = 1.0
+omega = 1.0
+EPSILON_S = 0.0 #Energy of the saddle
+parameters = np.array([1,omega, EPSILON_S, alpha, beta,omega])
+eqNum = 1 
+#model = 'uncoupled'
+#eqPt = tp_UPOsHam2dof.get_eq_pts(eqNum,model, parameters)
+eqPt = tp.get_eq_pts(eqNum, uncoupled.init_guess_eqpt_uncoupled, \
+                    uncoupled.grad_pot_uncoupled, parameters)
+
+
+#%%
+
+#deltaE_vals = [0.01, 0.1, 1.00, 2.0, 4.0]
+#xLeft = [0.0,-0.05,0.01,0.01,0.18]
+#xRight = [0.05,0.10,0.18,0.18,0.18]
+#linecolor = ['b','r','g','m','c']
+deltaE_vals = [0.1, 1.00]
+xLeft = [0.0,0.01]
+xRight = [0.05,0.11]
+linecolor = ['b','r']
+
+
+for i in range(len(deltaE_vals)):
+    
+    e = deltaE_vals[i] # total energy
+    n = 12 # number of intervals we want to divide
+    n_turn = 1 # nth turning point we want to choose
+    deltaE = e - parameters[2]
+    
+    """
+    Trial initial Condition s.t. one initial condition is on the LHS of the UPO and the 
+    other one is on the RHS of the UPO
+    """
+    state0_2 = [-0.1 , -math.sqrt(2*e+0.1**2-0.5*0.1**4),0.0,0.0]
+    state0_3 = [0.11 , -math.sqrt(2*e+0.11**2-0.5*0.11**4),0.0,0.0]
+    
+    po_fam_file = open("x0_turningpoint_deltaE%s_uncoupled.dat" %(deltaE),'a+')
+    [x0po_1, T_1,energyPO_1] = tp.turningPoint(
[…]
