### 2019-11-04 afb37e7
remove old script filenames
