+++ examples/tp_UPOs_coupled.py
@@ -1,561 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-Created on Tue Jul 30 10:02:48 2019
-
-@author: Wenyang Lyu and Shibabrat Naik
-
-Compute unstable peridoic orbits at different energies using turning point method
-"""
-
-# For the coupled problem
-import numpy as np
-import matplotlib.pyplot as plt
-from scipy.integrate import solve_ivp
-import math
-from scipy import optimize
-import sys
-sys.path.append('./src/')
-import tp_UPOsHam2dof ### import module xxx where xxx is the name of the python file xxx.py 
-import matplotlib as mpl
-from matplotlib import cm
-mpl.rcParams['mathtext.fontset'] = 'cm'
-mpl.rcParams['mathtext.rm'] = 'serif'
-
-
-#% Begin problem specific functions
-def init_guess_eqpt_coupled(eqNum, par):
-    """
-    Returns configuration space coordinates of the equilibrium points according to the index:
-    Saddle (EQNUM=1)
-    Centre (EQNUM=2,3)
-    """ 
-    
-    if eqNum == 1:
-        x0 = [0, 0]
-    elif eqNum == 2:
-        x0 = [np.sqrt(par[3]-par[6]/par[4]),0] 
-    elif eqNum == 3:
-        x0 = [-np.sqrt(par[3]-par[6]/par[4]),0] 
-    
-    return x0
-
-def grad_pot_coupled(x, par):
-    """
-    Returns the gradient of the potential energy function V(x,y)
-    """ 
-    
-    dVdx = (-par[3]+par[6])*x[0]+par[4]*(x[0])**3-par[6]*x[1]
-    dVdy = (par[5]+par[6])*x[1]-par[6]*x[0]
-            
-    F = [-dVdx, -dVdy]
-    
-    return F
-
-def pot_energy_coupled(x, y, par):
-    """Returns the potential energy function V(x,y)
-    """
-    
-    return -0.5*par[3]*x**2+0.25*par[4]*x**4 +0.5*par[5]*y**2+0.5*par[6]*(x-y)**2
-
-
-def get_coord_coupled(x, y, E, par):
-    """ 
-    Returns the initial position of x/y-coordinate on the potential energy 
-    surface(PES) for a specific energy E.
-    """
-    
-#    if model == 'uncoupled':
-#        return -0.5*par[3]*x**2+0.25*par[4]*x**4 +0.5*par[5]*y**2-V
-#    elif model =='coupled':
-    return -0.5*par[3]*x**2+0.25*par[4]*x**4 +0.5*par[5]*y**2+0.5*par[6]*(x-y)**2 - E
-#    elif model== 'deleonberne':
-#        return par[3]*( 1 - math.e**(-par[4]*x) )**2 + 4*y**2*(y**2 - 1)*math.e**(-par[5]*par[4]*x) + par[2]-V
-#    else:
-#        print("The model you are chosen does not exist, enter the function for finding coordinates on the PES for given x or y and V")
-
-
-def varEqns_coupled(t,PHI,par):
-    """    
-    Returns the state transition matrix , PHI(t,t0), where Df(t) is the Jacobian of the 
-    Hamiltonian vector field
-    
-    d PHI(t, t0)
-    ------------ =  Df(t) * PHI(t, t0)
-        dt
-    
-    """
-    
-    phi = PHI[0:16]
-    phimatrix  = np.reshape(PHI[0:16],(4,4))
-    x,y,px,py = PHI[16:20]
-    
-    
[…]
