### 2019-11-04 a89ce03
Add separate script for uncoupled quartic Hamiltonian
