+++ examples/uncoupled_quartic_hamiltonian.py
@@ -1,0 +1,642 @@
+# -*- coding: utf-8 -*-
+"""
+Created on Tue Jul 30 10:02:48 2019
+@author: Wenyang Lyu and Shibabrat Naik
+Script to define expressions for the coupled quartic Hamiltonian
+"""
+
+import numpy as np
+import matplotlib.pyplot as plt
+from scipy.integrate import solve_ivp
+import time
+import math
+from scipy import optimize
+ 
+
+# import matplotlib as mpl
+# from matplotlib import cm
+# mpl.rcParams['mathtext.fontset'] = 'cm'
+# mpl.rcParams['mathtext.rm'] = 'serif'
+
+#% Begin problem specific functions
+def init_guess_eqpt_uncoupled(eqNum, par):
+    """
+    Returns configuration space coordinates of the equilibrium points according to the index:
+    Saddle (EQNUM=1)
+    Centre (EQNUM=2,3)
+    """
+    
+    if eqNum == 1:
+        x0 = [0, 0]
+    elif eqNum == 2:
+        x0 = [np.sqrt(par[3]/par[4]),0] 
+    elif eqNum == 3:
+        x0 = [-np.sqrt(par[3]/par[4]),0]  
+    
+    return x0
+
+
+def grad_pot_uncoupled(x,par):
+    """ Returns the gradient of the potential energy function V(x,y) """ 
+    
+    dVdx = -par[3]*x[0]+par[4]*(x[0])**3
+    dVdy = par[5]*x[1]
+    
+    F = [-dVdx, -dVdy]
+    
+    return F
+
+
+def pot_energy_uncoupled(x, y, par):
+    """ Returns the potential energy function V(x,y) """
+    
+    return -0.5*par[3]*x**2+0.25*par[4]*x**4 +0.5*par[5]*y**2
+
+
+def eigvector_uncoupled(par):
+    """ Returns the correction factor to the eigenvectors for the linear guess """
+
+    correcx = 1
+    correcy = 1
+    
+    return correcx, correcy
+
+
+def guess_lin_uncoupled(eqPt, Ax, par):
+    """ Returns an initial guess for the unstable periodic orbit """ 
+    
+    correcx, correcy = eigvector_uncoupled(par)
+    
+    return [eqPt[0] + Ax*correcx,eqPt[1] + Ax*correcy,0,0]
+
+
+def jacobian_uncoupled(eqPt, par):
+    """ Returns Jacobian of the Hamiltonian vector field """
+    
+    x,y,px,py = eqPt[0:4]
+    
+    # The first order derivative of the Hamiltonian.
+    dVdx = -par[3]*x+par[4]*x**3
+    dVdy = par[5]*y
+
+    # The following is the Jacobian matrix 
+    d2Vdx2 = -par[3]+par[4]*3*x**2
+        
+    d2Vdy2 = par[5]
+
+    d2Vdydx = 0
+        
+    d2Vdxdy = d2Vdydx    
+
+    Df = np.array([[  0,     0,    par[0],    0],
+                   [0,     0,    0,    par[1]],
+                   [-d2Vdx2,  -d2Vdydx,   0,    0],
+                   [-d2Vdxdy, -d2Vdy2,    0,    0]])
+    
+    return Df
+
+
+def varEqns_uncoupled(t,PHI,par):
+    """    
+    Returns the state transition matrix , PHI(t,t0), where Df(t) is the Jacobian of the 
+    Hamiltonian vector field
+    
+    d PHI(t, t0)
+    ------------ =  Df(t) * PHI(t, t0)
+        dt
+    
+    """
+    
+    phi = PHI[0:16]
+    phimatrix  = np.reshape(PHI[0:16],(4,4))
+    x,y,px,py = PHI[16:20]
+    
+    
[…]
