### 2019-11-04 f95f120
Update gitignore and add setup.py
