+++ examples/differential_correction_deleonberne.py
@@ -1,0 +1,245 @@
+# -*- coding: utf-8 -*-
+"""
+Created on Tue Jul 30 10:02:48 2019
+
+@author: Wenyang Lyu and Shibabrat Naik
+
+Script to compute unstable periodic orbits at specified energies for the coupled quartic 
+Hamiltonian using differential correction
+"""
+
+import numpy as np
+import matplotlib.pyplot as plt
+from scipy.integrate import solve_ivp
+import time
+# import sys
+# sys.path.append('../src/')
+# import deleonberne ### import module xxx where xxx is the name of the python file xxx.py 
+
+# This needs testing for installation 
+import sys
+import importlib
+sys.path.insert(0, './src/')
+
+import differential_correction
+importlib.reload(differential_correction)
+import differential_correction as diffcorr
+
+import deleonberne_hamiltonian
+importlib.reload(deleonberne_hamiltonian)
+import deleonberne_hamiltonian as deleonberne
+# This needs testing for installation
+
+import matplotlib as mpl
+from matplotlib import cm
+mpl.rcParams['mathtext.fontset'] = 'cm'
+mpl.rcParams['mathtext.rm'] = 'serif'
+
+        
+#%% Setting up parameters and global variables
+    
+save_final_plot = True
+show_final_plot = False
+N = 4          # dimension of phase space
+MASS_A = 8.0 
+MASS_B = 8.0 # De Leon, Marston (1989)
+EPSILON_S = 1.0
+D_X = 10.0
+ALPHA = 1.00
+LAMBDA = 1.5
+parameters = np.array([MASS_A, MASS_B, EPSILON_S, D_X, LAMBDA, ALPHA])
+eqNum = 1  
+# model = 'deleonberne'
+eqPt = diffcorr.get_eq_pts(eqNum, deleonberne.init_guess_eqpt_deleonberne, \
+                            deleonberne.grad_pot_deleonberne, parameters)
+
+#energy of the saddle eq pt
+eSaddle = diffcorr.get_total_energy([eqPt[0],eqPt[1],0,0], \
+                                    deleonberne.pot_energy_deleonberne, \
+                                    parameters) 
+
+#If orbit is an n-array, e.g. orbit = [orbit_0, orbit_1, ..., orbit_n]
+# 
+#e = np.zeros(n,1)    
+#for i in range(n):
+#    e[i] = get_total_energy_deleonberne(orbit[i], parameters)
+#e = np.mean(e)
+
+
+#%% 
+nFam = 100 # use nFam = 10 for low energy
+
+# first two amplitudes for continuation procedure to get p.o. family
+Ax1  = 2.e-5 # initial amplitude (1 of 2) values to use: 2.e-3
+Ax2  = 2*Ax1 # initial amplitude (2 of 2)
+
+t = time.time()
+
+#  get the initial conditions and periods for a family of periodic orbits
+
+
+po_fam_file = open("x0_diffcorr_fam_eqPt%s_deleonberne.dat" %eqNum,'a+')
+#[po_x0Fam,po_tpFam] = deleonberne.get_POFam(eqNum, Ax1, Ax2,nFam, po_fam_file, parameters,model)  
+[po_x0Fam,po_tpFam] = diffcorr.get_POFam(
+    eqNum, Ax1, Ax2, nFam, po_fam_file, \
+    deleonberne.init_guess_eqpt_deleonberne, \
+    deleonberne.grad_pot_deleonberne, deleonberne.jacobian_deleonberne, \
[…]
