### 2019-11-05 ca2188f
Minor changes
