+++ examples/diffcorr_UPOs_deleonberne.py
@@ -1,446 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-Created on Tue Jul 30 10:02:48 2019
-
-@author: Wenyang Lyu and Shibabrat Naik
-
-Script to compute unstable periodic orbits at specified energies for the coupled quartic 
-Hamiltonian using differential correction
-"""
-
-import numpy as np
-import matplotlib.pyplot as plt
-from scipy.integrate import solve_ivp
-import time
-import sys
-sys.path.append('../src/')
-import diffcorr_UPOsHam2dof ### import module xxx where xxx is the name of the python file xxx.py 
-import matplotlib as mpl
-from matplotlib import cm
-mpl.rcParams['mathtext.fontset'] = 'cm'
-mpl.rcParams['mathtext.rm'] = 'serif'
-
-
-#% Begin problem specific functions
-def init_guess_eqpt_deleonberne(eqNum, par):
-    """
-    Returns configuration space coordinates of the equilibrium points according to the index:
-    Saddle (EQNUM=1)
-    Centre (EQNUM=2,3)
-    """
-    
-    if eqNum == 1:
-        x0 = [0, 0]
-    elif eqNum == 2:
-        x0 = [0, 1/np.sqrt(2)]  # EQNUM = 2, center-center
-    elif eqNum == 3:
-        x0 = [0, -1/np.sqrt(2)] # EQNUM = 3, center-center
-    
-    return x0
-
-
-def grad_pot_deleonberne(x, par):
-    """ Returns the gradient of the potential energy function V(x,y) """
-     
-    dVdx = -2*par[3]*par[4]*np.exp(-par[4]*x[0])*(np.exp(-par[4]*x[0]) - 1) - \
-        4*par[5]*par[4]*x[1]**2*(x[1]**2 - 1)*np.exp(-par[5]*par[4]*x[0])
-    dVdy = 8*x[1]*(2*x[1]**2 - 1)*np.exp(-par[5]*par[4]*x[0])
-    
-    F = [-dVdx, -dVdy]
-    
-    return F
-
-def pot_energy_deleonberne(x, y, par):
-    """ Returns the potential energy function V(x,y) """
-    
-    return par[3]*( 1 - np.exp(-par[4]*x) )**2 + 4*y**2*(y**2 - 1)*np.exp(-par[5]*par[4]*x) + par[2]
-
-
-def eigvector_deleonberne(par):
-    """ Returns the correction factor to the eigenvectors for the linear guess """
-
-    correcx = 1
-    correcy = 0
-    
-    return correcx, correcy
-
-
-def guess_lin_deleonberne(eqPt, Ax, par):
-    """ Returns an initial guess for the unstable periodic orbit """
-    
-    correcx, correcy = eigvector_deleonberne(par)
-    
-
-    return [eqPt[0] - Ax*correcx,eqPt[1] + Ax*correcy,0,0]
-
-
-def jacobian_deleonberne(eqPt, par):
-    """ Returns Jacobian of the Hamiltonian vector field """
-    
-    x,y,px,py = eqPt[0:4]
-    # The first order derivative of the Hamiltonian.
-    dVdx = - 2*par[3]*par[4]*np.exp(-par[4]*x)*(np.exp(-par[4]*x) - 1) - 4*par[5]*par[4]*y**2*(y**2 - 1)*np.exp(-par[5]*par[4]*x) 
-    dVdy = 8*y*(2*y**2 - 1)*np.exp(-par[5]*par[4]*x)
-
-    # The following is the Jacobian matrix 
[…]
