### 2019-11-05 7452383
Separate script for turning point method for De Leon-Berne
