+++ examples/tp_UPOs_deleonberne.py
@@ -1,537 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-Created on Tue Jul 30 10:02:48 2019
-
-@author: Wenyang Lyu and Shibabrat Naik
-
-Compute unstable peridoic orbits at different energies using turning point method
-"""
-
-# For the DeLeon-Berne problem
-import numpy as np
-import matplotlib.pyplot as plt
-from scipy.integrate import solve_ivp
-import math
-from scipy import optimize
-import sys
-sys.path.append('./src/')
-import tp_UPOsHam2dof ### import module xxx where xxx is the name of the python file xxx.py 
-import matplotlib as mpl
-from matplotlib import cm
-mpl.rcParams['mathtext.fontset'] = 'cm'
-mpl.rcParams['mathtext.rm'] = 'serif'
-
-
-#% Begin problem specific functions
-def init_guess_eqpt_deleonberne(eqNum, par):
-    """
-    Returns configuration space coordinates of the equilibrium points according to the index:
-    Saddle (EQNUM=1)
-    Centre (EQNUM=2,3)
-    """ 
-    
-    if eqNum == 1:
-        x0 = [0, 0]
-    elif eqNum == 2:
-        x0 = [0, 1/np.sqrt(2)]  # EQNUM = 2, center-center
-    elif eqNum == 3:
-        x0 = [0, -1/np.sqrt(2)] # EQNUM = 3, center-center
-    
-    return x0
-
-
-def grad_pot_deleonberne(x, par):
-    """
-    Returns the gradient of the potential energy function V(x,y)
-    """    
-    
-    dVdx = -2*par[3]*par[4]*np.exp(-par[4]*x[0])*(np.exp(-par[4]*x[0]) - 1) - \
-        4*par[5]*par[4]*x[1]**2*(x[1]**2 - 1)*np.exp(-par[5]*par[4]*x[0])
-    dVdy = 8*x[1]*(2*x[1]**2 - 1)*np.exp(-par[5]*par[4]*x[0])
-    
-    F = [-dVdx, -dVdy]
-    
-    return F
-
-def pot_energy_deleonberne(x, y, par):
-    """Returns the potential energy function V(x,y)
-    """
-    
-    return par[3]*( 1 - np.exp(-par[4]*x) )**2 + 4*y**2*(y**2 - 1)*np.exp(-par[5]*par[4]*x) + par[2]
-
-
-
-def get_coord_deleonberne(x,y, E, par):
-    """ 
-    Returns the initial position of x/y-coordinate on the potential energy 
-    surface(PES) for a specific energy E.
-    """
-
-    return par[3]*( 1 - math.e**(-par[4]*x) )**2 + \
-                4*y**2*(y**2 - 1)*math.e**(-par[5]*par[4]*x) + par[2] - E
-
-
-def varEqns_deleonberne(t,PHI,par):
-    """    
-    Returns the state transition matrix , PHI(t,t0), where Df(t) is the Jacobian of the 
-    Hamiltonian vector field
-    
-    d PHI(t, t0)
-    ------------ =  Df(t) * PHI(t, t0)
-        dt
-    
-    """
-    
-    phi = PHI[0:16]
-    phimatrix  = np.reshape(PHI[0:16],(4,4))
-    x,y,px,py = PHI[16:20]
-    
-    
-    # The first order derivative of the potential energy.
[…]
