+++ examples/turning_pt_deleonberne.py
@@ -1,0 +1,361 @@
+# -*- coding: utf-8 -*-
+"""
+Created on Tue Jul 30 10:02:48 2019
+
+@author: Wenyang Lyu and Shibabrat Naik
+
+Compute unstable peridoic orbits at different energies using turning point method
+"""
+
+# For the DeLeon-Berne problem
+import numpy as np
+import matplotlib.pyplot as plt
+from scipy.integrate import solve_ivp
+import math
+from scipy import optimize
+
+import matplotlib as mpl
+from matplotlib import cm
+mpl.rcParams['mathtext.fontset'] = 'cm'
+mpl.rcParams['mathtext.rm'] = 'serif'
+
+# This needs testing for installation 
+import sys
+import importlib
+sys.path.insert(0, './src/')
+
+import turning_point
+importlib.reload(turning_point)
+import turning_point as tp
+
+import deleonberne_hamiltonian
+importlib.reload(deleonberne_hamiltonian)
+import deleonberne_hamiltonian as deleonberne
+# This needs testing for installation
+
+#%% Setting up parameters and global variables
+"""
+Initial Condition for the true periodic orbit
+H = T + V where T is the kinetic energy and V is the potential energy. In our example, 
+the potential energy V = -0.5*alpha*x**2+0.25*beta*x**4+0.5*omega*y**2.
+If we fix x, then y = +- math.sqrt((V +0.5*alpha*x**2-0.25*beta*x**4)/(0.5*omega) ) so that
+the point starts at the potential energy surface V.
+"""
+save_final_plot = True
+show_final_plot = False
+show_itrsteps_plots = False # show iteration of the UPOs in plots
+N = 4         # dimension of phase space
+MASS_A = 8.0
+MASS_B = 8.0 # De Leon, Marston (1989)
+EPSILON_S = 1.0
+D_X = 10.0
+ALPHA = 1.00
+LAMBDA = 1.5
+parameters = np.array([MASS_A, MASS_B, EPSILON_S, D_X, LAMBDA, ALPHA])
+eqNum = 1 
+eqPt = tp.get_eq_pts(eqNum, deleonberne.init_guess_eqpt_deleonberne, \
+                    deleonberne.grad_pot_deleonberne, parameters)
+
+
+#%% 
+#E_vals = [1.1, 2.00, 3.00, 5.00]
+#linecolor = ['b','r','g','m','c']
+E_vals = [1.1, 2.00]
+linecolor = ['b','r']
+
+n = 4 # number of intervals we want to divide
+n_turn = 1 # nth turning point we want to choose.
+    
+for i in range(len(E_vals)):
+    
+    e = E_vals[i] # total energy
+    deltaE = e - parameters[2]
+    
+    """
+    Trial initial Condition s.t. one initial condition is on the LHS of the UPO and the 
+    other one is on the RHS of the UPO
+    """
+    f1 = lambda x: deleonberne.get_coord_deleonberne(x,0.06,e,parameters)
+    x0_2 = optimize.newton(f1,-0.15)
+    state0_2 = [x0_2,0.06,0.0,0.0]
+
+    f2 = lambda x: deleonberne.get_coord_deleonberne(x,-0.05,e,parameters)
[…]
