+++ tests/compare_methods_deleonberne.py
@@ -143,7 +143,7 @@ def ham2dof_deleonberne(t, x, par):
 
 def half_period_deleonberne(t,x,par):
     """
-    Return the turning point where we want to stop the integration                           
+    Returns the turning point where we want to stop the integration                           
     
     pxDot = x[0]
     pyDot = x[1]
