### 2019-11-11 #1684 @Shibabrat
We have now started two PRs (https://github.com/WyLyu/UPOsHam/pull/34, https://github.com/WyLyu/UPOsHam/pull/35) that will resolve most of the issues currently pending. @JoshKarpel and @dpsanders can you please let us know if it makes sense what we have addressed? We can then go ahead and merge them or make appropriate changes.
