### 2019-11-12 2e85dcd
Remove commented-out code
