+++ examples/coupled_quartic_hamiltonian.py
@@ -11,12 +11,6 @@ from scipy.integrate import solve_ivp
 import time
 import math
 from scipy import optimize
- 
-
-# import matplotlib as mpl
-# from matplotlib import cm
-# mpl.rcParams['mathtext.fontset'] = 'cm'
-# mpl.rcParams['mathtext.rm'] = 'serif'
 
 
 #% Begin problem specific functions
@@ -59,16 +53,8 @@ def eigvector_coupled(par):
     evaluelamb = np.sqrt(-0.5*(par[3]-par[6]-par[1]*(par[1]+par[6]) - np.sqrt(par[1]**4 + \
                                2*par[1]**3*par[6] + par[1]**2*(par[6]**2+2*par[3]-2*par[6]) + \
                                par[1]*( 2*par[6]**2 + 2*par[3]*par[6]) +(par[3]- par[6])**2)))
-#    correcx = par[6]/(-evaluelamb**2 -par[3]+par[6])
-#    correcy = 1
-    #
-    #
-    #eqPt = 1
-    #eqPt = get_eq_pts_coupled(eqNum, par)
-    #evalue, evector = np.linalg.eig(jacobian_coupled([eqPt[0],eqPt[1],0,0],par))
-    #evector = RemoveInfinitesimals(evector[:,2])
-    #correcx = (evector[0]*1j).real
-    #correcy = (evector[1]*1j).real
+
+
     correcx = (par[1]*par[6])/(-evaluelamb**2 - par[3] + par[6])
     correcy = par[1]
     
@@ -185,14 +171,8 @@ def get_coord_coupled(x, y, E, par):
     Returns the initial position of x/y-coordinate on the potential energy 
     surface(PES) for a specific energy E.
     """
-#    if model == 'uncoupled':
-#        return -0.5*par[3]*x**2+0.25*par[4]*x**4 +0.5*par[5]*y**2-V
-#    elif model =='coupled':
+
     return -0.5*par[3]*x**2+0.25*par[4]*x**4 +0.5*par[5]*y**2+0.5*par[6]*(x-y)**2 - E
-#    elif model== 'deleonberne':
-#        return par[3]*( 1 - math.e**(-par[4]*x) )**2 + 4*y**2*(y**2 - 1)*math.e**(-par[5]*par[4]*x) + par[2]-V
-#    else:
-#        print("The model you are chosen does not exist, enter the function for finding coordinates on the PES for given x or y and V")
 
 
 def diffcorr_acc_corr_coupled(coords, phi_t1, x0, par):
@@ -258,7 +238,7 @@ def configdiff_coupled(guess1, guess2, ham2dof_model, half_period_model, n_turn,
                                                                                  guess2, x_diff1, \
                                                                                  x_diff2))
     
-    return x_diff1, x_diff2 #,y_diff1, y_diff2
+    return x_diff1, x_diff2
 
 
 def guess_coords_coupled(guess1, guess2, i, n, e, get_coord_model, par):
@@ -287,7 +267,7 @@ def plot_iter_orbit_coupled(x, ax, e, par):
     final points marked 
     """
     
-#    label_fs = 10
+    label_fs = 10
     axis_fs = 15 # fontsize for publications 
     
     ax.plot(x[:,0],x[:,1],x[:,3],'-')
@@ -335,322 +315,3 @@ def half_period_coupled(t, x, par):
     
     return x[3]
 
-
-#% End problem specific functions
-
-#% Begin problem specific functions
[…]
