### 2019-11-12 16f6378
Add docstring for function
