+++ examples/coupled_quartic_hamiltonian.py
@@ -163,6 +163,11 @@ def diffcorr_setup_coupled():
 
 
 def conv_coord_coupled(x1, y1, dxdot1, dydot1):
+    """
+    Returns the variable we want to keep fixed during differential correction.
+    
+    dxdot1----fix x, dydot1----fix y.
+    """
     return dxdot1
 
 
