+++ examples/deleonberne_hamiltonian.py
@@ -159,6 +159,11 @@ def diffcorr_setup_deleonberne():
 
 
 def conv_coord_deleonberne(x1, y1, dxdot1, dydot1):
+    """
+    Returns the variable we want to keep fixed during differential correction.
+    
+    dxdot1----fix x, dydot1----fix y.
+    """
     return dydot1
 
 
