### 2019-11-12 ef7c00d
replace inside-code strings with comments
