+++ examples/turning_pt_coord_diff_coupled.py
@@ -53,11 +53,7 @@ deltaE_vals = [0.1, 1.00]
 xLeft = [0.0,0.01]
 xRight = [0.05,0.18]
 linecolor = ['b','r']
-"""
-e is the total energy
-n is the number of intervals we want to divide
-n_turn is the nth turning point we want to choose.
-"""
+
 for i in range(len(deltaE_vals)):
     
     e = deltaE_vals[i]
@@ -65,10 +61,9 @@ for i in range(len(deltaE_vals)):
     n_turn = 2
     deltaE = e - parameters[2]
     
-    """
-    Trial initial Condition s.t. one initial condition is on the LHS of the UPO and the 
-    other one is on the RHS of the UPO
-    """
+    #Trial initial Condition s.t. one initial condition is on the LHS of the UPO and the 
+    #other one is on the RHS of the UPO
+
     x = xLeft[i]
     f2 = lambda y: coupled.get_coord_coupled(x, y, e, parameters)
     yanalytic = math.sqrt(2/(parameters[1]+parameters[6]))*(-math.sqrt( e + \
