### 2019-11-12 c33e2ae
update function names following PEP 8
