+++ examples/turning_pt_coord_diff_coupled.py
@@ -85,7 +85,7 @@ for i in range(len(deltaE_vals)):
     [x0po_1, T_1,energyPO_1] = tpcd.turningPoint_configdiff(state0_2, state0_3, \
                                         coupled.get_coord_coupled, \
                                         coupled.pot_energy_coupled, \
-                                        coupled.varEqns_coupled, \
+                                        coupled.variational_eqns_coupled, \
                                         coupled.configdiff_coupled, \
                                         coupled.ham2dof_coupled, \
                                         coupled.half_period_coupled, \
@@ -132,8 +132,8 @@ for i in range(len(deltaE_vals)):
                     rtol=RelTol, atol=AbsTol)
     te = soln.t_events[0]
     tt = [0,te[2]]
-    t,x,phi_t1,PHI = tpcd.stateTransitMat(tt, x0po[:,i], parameters, \
-                    coupled.varEqns_coupled)
+    t,x,phi_t1,PHI = tpcd.state_transit_matrix(tt, x0po[:,i], parameters, \
+                    coupled.variational_eqns_coupled)
     
     
     ax.plot(x[:,0],x[:,1],x[:,3],'-', color=linecolor[i], \
