+++ examples/uncoupled_quartic_hamiltonian.py
@@ -91,7 +91,7 @@ def jacobian_uncoupled(eqPt, par):
     return Df
 
 
-def varEqns_uncoupled(t,PHI,par):
+def variational_eqns_uncoupled(t,PHI,par):
     """    
     Returns the state transition matrix , PHI(t,t0), where Df(t) is the Jacobian of the 
     Hamiltonian vector field
