+++ tests/compare_methods_uncoupled.py
@@ -57,7 +57,7 @@ def pot_energy_uncoupled(x, y, par):
     return -0.5*par[3]*x**2+0.25*par[4]*x**4 +0.5*par[5]*y**2
 
 
-def varEqns_uncoupled(t,PHI,par):
+def variational_eqns_uncoupled(t,PHI,par):
     """    
     Returns the state transition matrix , PHI(t,t0), where Df(t) is the Jacobian of the 
     Hamiltonian vector field
@@ -204,8 +204,8 @@ soln = solve_ivp(f, TSPAN, x0po_1_tpcd[-1,0:4],method='RK45',dense_output=True, 
 
 te = soln.t_events[0]
 tt = [0,te[2]]
-t,x,phi_t1,PHI = diffcorr_UPOsHam2dof.stateTransitMat(tt, x0po_1_tpcd[-1,0:4], parameters, \
-                                                      varEqns_uncoupled)
+t,x,phi_t1,PHI = diffcorr_UPOsHam2dof.state_transit_matrix(tt, x0po_1_tpcd[-1,0:4], parameters, \
+                                                      variational_eqns_uncoupled)
 
 ax = plt.gca(projection='3d')
 ax.plot(x[:,0],x[:,1],x[:,3],'-',label='$\Delta E$ = 0.1, using tpcd')
@@ -221,8 +221,8 @@ soln = solve_ivp(f, TSPAN, x0po_1_diffcorr, method='RK45', dense_output=True, \
 
 te = soln.t_events[0]
 tt = [0,te[2]]
-t,x,phi_t1,PHI = diffcorr_UPOsHam2dof.stateTransitMat(tt, x0po_1_diffcorr, parameters, \
-                                                      varEqns_uncoupled)
+t,x,phi_t1,PHI = diffcorr_UPOsHam2dof.state_transit_matrix(tt, x0po_1_diffcorr, parameters, \
+                                                      variational_eqns_uncoupled)
 
 ax = plt.gca(projection='3d')
 ax.plot(x[:,0],x[:,1],x[:,3],'--',label='$\Delta E$ = 0.1, using dcnc')
@@ -237,8 +237,8 @@ soln = solve_ivp(f, TSPAN, x0po_1_turningpoint[-1,0:4], method='RK45', dense_out
 
 te = soln.t_events[0]
 tt = [0,te[2]]
-t,x,phi_t1,PHI = diffcorr_UPOsHam2dof.stateTransitMat(tt, x0po_1_turningpoint[-1,0:4], parameters, \
-                                                      varEqns_uncoupled)
+t,x,phi_t1,PHI = diffcorr_UPOsHam2dof.state_transit_matrix(tt, x0po_1_turningpoint[-1,0:4], parameters, \
+                                                      variational_eqns_uncoupled)
 
 ax.plot(x[:,0],x[:,1],x[:,3],':',label='$\Delta E$ = 0.1, using tp')
 ax.scatter(x[0,0],x[0,1],x[0,3],s=20,marker='*')
@@ -252,8 +252,8 @@ soln = solve_ivp(f, TSPAN, x0po_1_analytic, method='RK45', dense_output=True, \
 
 te = soln.t_events[0]
 tt = [0,te[2]]
-t,x,phi_t1,PHI = diffcorr_UPOsHam2dof.stateTransitMat(tt, x0po_1_analytic, parameters, \
-                                                      varEqns_uncoupled)
+t,x,phi_t1,PHI = diffcorr_UPOsHam2dof.state_transit_matrix(tt, x0po_1_analytic, parameters, \
+                                                      variational_eqns_uncoupled)
 
[…]

[…]
