+++ tests/compare_methods_coupled.py
@@ -129,12 +129,11 @@ def half_period_coupled(t, x, par):
     yDot = x[3]
     """
     
-    terminal = True
-    # The zero can be approached from either direction
-    direction = 0 #0: all directions of crossing
-    
     return x[3]
 
+half_period_coupled.terminal = True 
+# The zero can be approached from either direction
+half_period_coupled.direction=0#0: all directions of crossing
 
 #% End problem specific functions
 
