+++ tests/compare_methods_deleonberne.py
@@ -136,12 +136,11 @@ def half_period_deleonberne(t,x,par):
     yDot = x[3]
     """
     
-    terminal = True
-    # The zero can be approached from either direction
-    direction = 0 #0: all directions of crossing
-    
     return x[2]
         
+half_period_deleonberne.terminal = True 
+# The zero can be approached from either direction
+half_period_deleonberne.direction=0#0: all directions of crossing
 
 #% End problem specific functions
 
