### 2019-11-13 2254339
minor updates
