+++ tests/compare_methods_deleonberne.py
@@ -12,8 +12,7 @@ from scipy.integrate import solve_ivp
 from mpl_toolkits.mplot3d import Axes3D
 import sys
 sys.path.append('./src/')
-import tpcd_UPOsHam2dof ### import module xxx where xxx is the name of the python file xxx.py 
-import diffcorr_UPOsHam2dof
+import differential_correction
 import matplotlib as mpl
 from matplotlib import cm
 mpl.rcParams['mathtext.fontset'] = 'cm'
@@ -156,11 +155,11 @@ ALPHA = 1.00
 LAMBDA = 1.5
 parameters = np.array([MASS_A, MASS_B, EPSILON_S, D_X, LAMBDA, ALPHA])
 eqNum = 1  
-eqPt = diffcorr_UPOsHam2dof.get_eq_pts(eqNum, init_guess_eqpt_deleonberne, \
+eqPt = differential_correction.get_eq_pts(eqNum, init_guess_eqpt_deleonberne, \
                                        grad_pot_deleonberne, parameters)
 
 #energy of the saddle eq pt
-eSaddle = diffcorr_UPOsHam2dof.get_total_energy([eqPt[0],eqPt[1],0,0], pot_energy_deleonberne, \
+eSaddle = differential_correction.get_total_energy([eqPt[0],eqPt[1],0,0], pot_energy_deleonberne, \
                                                 parameters)
 
 
@@ -203,7 +202,7 @@ soln = solve_ivp(f, TSPAN, x0po_1_tpcd[-1,0:4],method='RK45',dense_output=True, 
                  rtol=RelTol, atol=AbsTol)
 te = soln.t_events[0]
 tt = [0,te[2]]
-t,x,phi_t1,PHI = diffcorr_UPOsHam2dof.state_transit_matrix(tt, x0po_1_tpcd[-1,0:4], parameters, \
+t,x,phi_t1,PHI = differential_correction.state_transit_matrix(tt, x0po_1_tpcd[-1,0:4], parameters, \
                                                       variational_eqns_deleonberne)
 
 ax = plt.gca(projection='3d')
@@ -217,7 +216,7 @@ soln = solve_ivp(f, TSPAN, x0po_1_diffcorr,method='RK45',dense_output=True, \
                  rtol=RelTol, atol=AbsTol)
 te = soln.t_events[0]
 tt = [0,te[2]]
-t,x,phi_t1,PHI = diffcorr_UPOsHam2dof.state_transit_matrix(tt, x0po_1_diffcorr, parameters, \
+t,x,phi_t1,PHI = differential_correction.state_transit_matrix(tt, x0po_1_diffcorr, parameters, \
                                                       variational_eqns_deleonberne)
 
 
@@ -231,7 +230,7 @@ soln = solve_ivp(f, TSPAN, x0po_1_turningpoint[-1,0:4],method='RK45',dense_outpu
                  rtol=RelTol, atol=AbsTol)
 te = soln.t_events[0]
 tt = [0,te[2]]
-t,x,phi_t1,PHI = diffcorr_UPOsHam2dof.state_transit_matrix(tt, x0po_1_turningpoint[-1,0:4], parameters, \
+t,x,phi_t1,PHI = differential_correction.state_transit_matrix(tt, x0po_1_turningpoint[-1,0:4], parameters, \
                                                       variational_eqns_deleonberne)
 
 
@@ -249,8 +248,8 @@ xMat, yMat = np.meshgrid(xVec, yVec)
 #cset1 = ax.contour(xMat, yMat, uncoupled_tpcd.get_pot_surf_proj(xVec, yVec,parameters), [0.001,0.1,1,2,4],
 #                       linewidths = 1.0, cmap=cm.viridis, alpha = 0.8)
 cset2 = ax.contour(xMat, yMat, \
[…]

[…]
