+++ examples/turning_pt_coord_diff_deleonberne.py
@@ -74,18 +74,16 @@ for i in range(len(E_vals)):
     x0_3 = optimize.newton(f2,-0.15)
     state0_3 = [x0_3,-0.05,0.0,0.0]
     
-    po_fam_file = open("x0_tpcd_deltaE%s_deleonberne.dat" %(deltaE),'a+')
-    [x0po_1, T_1,energyPO_1] = tpcd.turningPoint_configdiff(
-        state0_2, state0_3, deleonberne.get_coord_deleonberne, \
-        deleonberne.pot_energy_deleonberne, deleonberne.variational_eqns_deleonberne, \
-        deleonberne.configdiff_deleonberne, \
-        deleonberne.ham2dof_deleonberne, \
-        deleonberne.half_period_deleonberne, \
-        deleonberne.guess_coords_deleonberne, \
-        deleonberne.plot_iter_orbit_deleonberne, \
-        parameters, e, n, n_turn, show_itrsteps_plots, po_fam_file) 
-
-    po_fam_file.close()
+    with open("x0_tpcd_deltaE%s_deleonberne.dat" %(deltaE),'a+') as po_fam_file:
+        [x0po_1, T_1,energyPO_1] = tpcd.turningPoint_configdiff(
+            state0_2, state0_3, deleonberne.get_coord_deleonberne, \
+            deleonberne.pot_energy_deleonberne, deleonberne.variational_eqns_deleonberne, \
+            deleonberne.configdiff_deleonberne, \
+            deleonberne.ham2dof_deleonberne, \
+            deleonberne.half_period_deleonberne, \
+            deleonberne.guess_coords_deleonberne, \
+            deleonberne.plot_iter_orbit_deleonberne, \
+            parameters, e, n, n_turn, show_itrsteps_plots, po_fam_file) 
 
 
 #%% Load periodic orbit data from ascii files
@@ -97,11 +95,10 @@ for i in range(len(E_vals)):
     e = E_vals[i]
     deltaE = e - parameters[2]
 
-    po_fam_file = open("x0_tpcd_deltaE%s_deleonberne.dat" %(deltaE),'a+')
-    print('Loading the periodic orbit family from data file',po_fam_file.name,'\n') 
-    x0podata = np.loadtxt(po_fam_file.name)
-    po_fam_file.close()
-    x0po[:,i] = x0podata[-1,0:4] 
+    with open("x0_tpcd_deltaE%s_deleonberne.dat" %(deltaE),'a+') as po_fam_file:
+        print('Loading the periodic orbit family from data file',po_fam_file.name,'\n') 
+        x0podata = np.loadtxt(po_fam_file.name)
+        x0po[:,i] = x0podata[-1,0:4] 
 
 
 #%% Plotting the Family
