+++ examples/differential_correction_coupled.py
@@ -67,26 +67,25 @@ Ax2  = 2*Ax1 # initial amplitude (2 of 2)
 t = time.time()
 
 #  get the initial conditions and periods for a family of periodic orbits
-po_fam_file = open("x0_diffcorr_fam_eqPt%s_coupled.dat" %eqNum,'a+')
-[po_x0Fam,po_tpFam] = diffcorr.get_po_fam(eqNum, Ax1, Ax2, nFam, \
-                                        po_fam_file, \
-                                        coupled.init_guess_eqpt_coupled, \
-                                        coupled.grad_pot_coupled, \
-                                        coupled.jacobian_coupled, \
-                                        coupled.guess_lin_coupled, \
-                                        coupled.diffcorr_setup_coupled, \
-                                        coupled.conv_coord_coupled, \
-                                        coupled.diffcorr_acc_corr_coupled, \
-                                        coupled.ham2dof_coupled, \
-                                        coupled.half_period_coupled, \
-                                        coupled.pot_energy_coupled, \
-                                        coupled.variational_eqns_coupled, \
-                                        coupled.plot_iter_orbit_coupled, \
-                                        parameters)  
+with open("x0_diffcorr_fam_eqPt%s_coupled.dat" %eqNum,'a+') as po_fam_file:
+    [po_x0Fam,po_tpFam] = diffcorr.get_po_fam(eqNum, Ax1, Ax2, nFam, \
+                                            po_fam_file, \
+                                            coupled.init_guess_eqpt_coupled, \
+                                            coupled.grad_pot_coupled, \
+                                            coupled.jacobian_coupled, \
+                                            coupled.guess_lin_coupled, \
+                                            coupled.diffcorr_setup_coupled, \
+                                            coupled.conv_coord_coupled, \
+                                            coupled.diffcorr_acc_corr_coupled, \
+                                            coupled.ham2dof_coupled, \
+                                            coupled.half_period_coupled, \
+                                            coupled.pot_energy_coupled, \
+                                            coupled.variational_eqns_coupled, \
+                                            coupled.plot_iter_orbit_coupled, \
+                                            parameters)  
 
 poFamRuntime = time.time() - t
 x0podata = np.concatenate((po_x0Fam, po_tpFam),axis=1)
-po_fam_file.close()
 
 
 
@@ -97,30 +96,27 @@ po_fam_file.close()
 for i in range(len(deltaE_vals)):
     deltaE = deltaE_vals[i]
     
-    po_fam_file = open("x0_diffcorr_fam_eqPt%s_coupled.dat" %eqNum ,'a+')
-    eTarget = eSaddle + deltaE 
-    print('Loading the periodic orbit family from data file',po_fam_file.name,'\n') 
-    x0podata = np.loadtxt(po_fam_file.name)
-    po_fam_file.close()
+    with open("x0_diffcorr_fam_eqPt%s_coupled.dat" %eqNum ,'a+') as po_fam_file:
+        eTarget = eSaddle + deltaE 
+        print('Loading the periodic orbit family from data file',po_fam_file.name,'\n') 
+        x0podata = np.loadtxt(po_fam_file.name)
     
     
-    po_brac_file = open("x0po_T_energyPO_eqPt%s_brac%s_coupled.dat" %(eqNum,deltaE),'a+')
-    t = time.time()
-    x0poTarget,TTarget = diffcorr.po_bracket_energy(eTarget, x0podata, po_brac_file, \
-                                                coupled.diffcorr_setup_coupled, \
-                                                coupled.conv_coord_coupled, \
-                                                coupled.diffcorr_acc_corr_coupled, \
-                                                coupled.ham2dof_coupled, \
-                                                coupled.half_period_coupled, \
-                                                coupled.pot_energy_coupled, \
[…]

[…]
