+++ tests/compare_methods_uncoupled.py
@@ -165,19 +165,19 @@ eSaddle = 0.0 # energy of the saddle
 data_path = "./data/"
 
 
-po_fam_file = "x0_tpcd_deltaE%s_uncoupled.txt" %(deltaE)
+po_fam_file = "x0_tpcd_deltaE%s_uncoupled.dat" %(deltaE)
 print('Loading the periodic orbit family from data file',po_fam_file,'\n') 
 x0podata = np.loadtxt(data_path + po_fam_file)
 x0po_1_tpcd = x0podata
 
 
-po_fam_file = "x0_turningpoint_deltaE%s_uncoupled.txt" %(deltaE)
+po_fam_file = "x0_turningpoint_deltaE%s_uncoupled.dat" %(deltaE)
 print('Loading the periodic orbit family from data file',po_fam_file,'\n') 
 x0podata = np.loadtxt(data_path + po_fam_file)
 x0po_1_turningpoint = x0podata
 
 
-po_fam_file = "x0_diffcorr_deltaE%s_uncoupled.txt" %(deltaE)
+po_fam_file = "x0_diffcorr_deltaE%s_uncoupled.dat" %(deltaE)
 print('Loading the periodic orbit family from data file',po_fam_file,'\n') 
 x0podata = np.loadtxt(data_path + po_fam_file)
 
