### 2019-11-14 #1684 @kthyng
Just another quick ping to check in on thoughts from @JoshKarpel and @dpsanders!
