### 2019-11-15 #1684 @JoshKarpel
@kthyng We're definitely making progress. The authors have made a lot of good changes, and I think this big PR finally gets us to a place where they can start satisfying the review requirements directly (things like setting up documentation and making it an installable package).
