+++ README.md
@@ -1,7 +1,7 @@
 -   [Summary](#summary)
 -   [Installation](#installation)
-    -   [Methods and applications](#methods-and-applications)
 -   [Usage](#usage)
+    -   [Comparing methods for applications](#comparing-methods-for-applications)
 -   [Contributing](#contributing)
 -   [Copyright and License](#copyright-and-license)
 -   [References](#references)
@@ -11,7 +11,7 @@
 Summary
 -------
 
-This Python package is a collection of three methods for computing unstable periodic orbits (Parker and Chua [1989](#ref-Parker_1989)) in Hamiltonian systems that model a diverse array of problems in physical sciences and engineering (Wiggins [2003](#ref-wiggins_introduction_2003)). The unstable periodic orbits exist in the bottleneck of the equipotential line *V*(*x*, *y*)=*E* and project as lines on the configuration space (*x*, *y*) (Wiggins [2016](#ref-wiggins_role_2016)). The three methods (Pollak, Child, and Pechukas [1980](#ref-Pollak_1980),Koon et al. ([2011](#ref-Koon2011)),Naik and Wiggins ([2019](#ref-naik_finding_2019b))) described below have been implemented for three Hamiltonian systems of the form kinetic plus potential energy and a brief description can be found in the [paper.pdf](https://github.com/WyLyu/UPOsHam/tree/master/paper/paper.pdf). The scripts are written as demonstration of how to modify and adapt the code for a problem of interest.
+This Python package is a collection of three methods for computing unstable periodic orbits in two degrees of freedom Hamiltonian systems that model a diverse array of problems in physical sciences and engineering (Parker and Chua [1989](#ref-Parker_1989), Wiggins ([2003](#ref-wiggins_introduction_2003))). The unstable periodic orbits exist in the bottleneck of the equipotential line *V*(*x*, *y*)=*E* and project as lines on the configuration space (*x*, *y*) for the Hamiltonian system of the form kinetic plus potential energy (Wiggins [2016](#ref-wiggins_role_2016)). The three methods implemented here (and available under [source](src/) directory) has been used in (Pollak, Child, and Pechukas [1980](#ref-Pollak_1980),De Leon and Berne ([1981](#ref-Deleon_Berne_1981)),Wang Sang Koon et al. ([2000](#ref-koon_heteroclinic_2000)),W. S. Koon et al. ([2011](#ref-Koon2011)),Naik and Ross ([2017](#ref-naik_geometry_2017)),Ross et al. ([2018](#ref-ross_experimental_2018)),Naik and Wiggins ([2019](#ref-naik_finding_2019b))) for transition dynamics in chemical reactions, celestial mechanics, and ship capsize. We have chosen three Hamiltonian systems that have two wells connected by a bottleneck where the unstable periodic orbits exist for energy above the energy of the saddle equilibrium point. A brief description of these systems can be found in the [paper](https://github.com/WyLyu/UPOsHam/tree/master/paper/paper.pdf).
 
 Installation
 ------------
@@ -26,12 +26,22 @@ $ pip install -r requirements.txt (or pip3 install -r requirements.txt)
 
 and check the modules shown in [requirements.txt](https://github.com/WyLyu/UPOsHam/tree/master/requirements.txt) are installed using conda/pip.
 
-### Methods and applications
-
-Methods for computing an unstable periodic orbit at a specified total energy is in the [src](src/) directory and the applications are in the [examples](examples/) directory.
-
 Usage
 -----
+
[…]
