+++ README.md
@@ -1,8 +1,7 @@
 -   [Summary](#summary)
 -   [Installation](#installation)
 -   [Usage](#usage)
-    -   [Comparing methods for
-        applications](#comparing-methods-for-applications)
+    -   [Comparing the methods for the coupled quartic Hamiltonian](#comparing-the-methods-for-the-coupled-quartic-hamiltonian)
 -   [Contributing](#contributing)
 -   [Copyright and License](#copyright-and-license)
 -   [References](#references)
@@ -12,24 +11,7 @@
 Summary
 -------
 
-This Python package is a collection of three methods for computing
-unstable periodic orbits in two degrees of freedom Hamiltonian systems
-that model a diverse array of problems in physical sciences and
-engineering (Parker and Chua [1989](#ref-Parker_1989),
-@wiggins\_introduction\_2003). The unstable periodic orbits exist in the
-bottleneck of the equipotential line *V*(*x*, *y*) = *E* and project as
-lines on the configuration space (*x*, *y*) for the Hamiltonian system
-of the form kinetic plus potential energy (Wiggins
-[2016](#ref-wiggins_role_2016)). The three methods implemented here (and
-available under [source](src/) directory) has been used in (Pollak,
-Child, and Pechukas
-[1980](#ref-Pollak_1980),@Deleon\_Berne\_1981,@koon\_heteroclinic\_2000,@Koon2011,@naik\_geometry\_2017,@ross\_experimental\_2018,@naik\_finding\_2019b)
-for transition dynamics in chemical reactions, celestial mechanics, and
-ship capsize. We have chosen three Hamiltonian systems that have two
-wells connected by a bottleneck where the unstable periodic orbits exist
-for energy above the energy of the saddle equilibrium point. A brief
-description of these systems can be found in the
-[paper](https://github.com/WyLyu/UPOsHam/tree/master/paper/paper.pdf).
+This Python package is a collection of three methods for computing unstable periodic orbits in two degrees of freedom Hamiltonian systems that model a diverse array of problems in physical sciences and engineering (Parker and Chua [1989](#ref-Parker_1989); Wiggins [2003](#ref-wiggins_introduction_2003)). The unstable periodic orbits exist in the bottleneck of the equipotential line *V*(*x*, *y*)=*E* and project as lines on the configuration space (*x*, *y*) for the Hamiltonian system of the form kinetic plus potential energy (Wiggins [2016](#ref-wiggins_role_2016)). The three methods implemented here (and available under [src](src/) directory) has been used in (Pollak, Child, and Pechukas [1980](#ref-Pollak_1980); De Leon and Berne [1981](#ref-Deleon_Berne_1981); Wang Sang Koon et al. [2000](#ref-koon_heteroclinic_2000); W. S. Koon et al. [2011](#ref-Koon2011); Naik and Ross [2017](#ref-naik_geometry_2017); Ross et al. [2018](#ref-ross_experimental_2018); Naik and Wiggins [2019](#ref-naik_finding_2019b)) for transition dynamics in chemical reactions, celestial mechanics, and ship capsize. We have chosen three Hamiltonian systems that have two wells connected by a bottleneck where the unstable periodic orbits exist for energy above the energy of the saddle equilibrium point. A brief description of these systems can be found in the [paper](https://github.com/WyLyu/UPOsHam/tree/master/paper/paper.pdf).
 
 Installation
 ------------
@@ -42,120 +24,72 @@ $ cd UPOsHam
[…]
