### 2019-11-27 #1684 @labarba
👋  @dpsanders — Your last comment from Oct. 17 says:
> There are also still no automated tests.

Do note that JOSS review criteria regarding [tests](https://joss.readthedocs.io/en/latest/review_criteria.html#tests) is that "Authors are strongly encouraged to include an automated test suite…" but "Documented manual steps that can be followed to objectively check the expected functionality of the software" are OK.

Given this, and that this review has been ongoing for three months already, and the authors have made many improvements, are you ready to submit a recommendation?
