### 2019-11-28 e65e592
Clean up comments and obsolete portions
