+++ examples/uncoupled_quartic_hamiltonian.py
@@ -13,11 +13,6 @@ import math
 from scipy import optimize
  
 
-# import matplotlib as mpl
-# from matplotlib import cm
-# mpl.rcParams['mathtext.fontset'] = 'cm'
-# mpl.rcParams['mathtext.rm'] = 'serif'
-
 #% Begin problem specific functions
 def init_guess_eqpt_uncoupled(eqNum, par):
     """
@@ -236,22 +231,13 @@ def configdiff_uncoupled(guess1, guess2, ham2dof_model, half_period_model, n_tur
     x_diff2 = guess2[0] - x_turn2
     y_diff2 = guess2[1] - y_turn2
     
-#    if model == 'uncoupled':
+
     print("Initial guess1%s, initial guess2%s, x_diff1 is %s, x_diff2 is%s " %(guess1, \
                                                                                guess2, \
                                                                                x_diff1, \
                                                                                x_diff2))
-#        return x_diff1, x_diff2
-
-#        return x_diff1, x_diff2
-#    elif model == 'deleonberne':
-#        print("Initial guess1%s, initial guess2%s, y_diff1 is %s, y_diff2 is%s " %(guess1, guess2, y_diff1, y_diff2))
-#        return y_diff1, y_diff2
-#    else: 
-#        print("Need to decide to use either x_diff or y_diff")
-#        print("Initial guess1%s, initial guess2%s, x_diff1 is %s, x_diff2 is%s, y_diff1 is %s, y_diff2 is%s " %(guess1,guess2,x_diff1, x_diff2,y_diff1, y_diff2))
         
-    return x_diff1, x_diff2 #,y_diff1, y_diff2
+    return x_diff1, x_diff2
 
 
 def guess_coords_uncoupled(guess1, guess2, i, n, e, get_coord_model, par):
@@ -324,319 +310,3 @@ def half_period_uncoupled(t, x, par):
 #% End problem specific functions
 
 
-#% Begin problem specific functions
-# def init_guess_eqpt_uncoupled(eqNum, par):
-#     """
-#     Returns configuration space coordinates of the equilibrium points according to the index:
-#     Saddle (EQNUM=1)
-#     Centre (EQNUM=2,3)
-#     """ 
-    
-#     if eqNum == 1:
-#         x0 = [0, 0]
-#     elif eqNum == 2:
-#         x0 = [np.sqrt(par[3]/par[4]),0] 
-#     elif eqNum == 3:
-#         x0 = [-np.sqrt(par[3]/par[4]),0]  
-    
-#     return x0
-
-
-# def grad_pot_uncoupled(x,par):
-#     """
-#     Returns the gradient of the potential energy function V(x,y)
-#     """     
-    
-#     dVdx = -par[3]*x[0]+par[4]*(x[0])**3
-#     dVdy = par[5]*x[1]
-    
-#     F = [-dVdx, -dVdy]
-    
-#     return F
-
-
-# def pot_energy_uncoupled(x, y, par):
-#     """Returns the potential energy function V(x,y)
-#     """
-    
-#     return -0.5*par[3]*x**2+0.25*par[4]*x**4 +0.5*par[5]*y**2
-
-
-
-# def varEqns_uncoupled(t,PHI,par):
-#     """    
-#     Returns the state transition matrix , PHI(t,t0), where Df(t) is the Jacobian of the 
-#     Hamiltonian vector field
-    
-#     d PHI(t, t0)
[…]
