### 2019-11-28 #1684 @Shibabrat
thank you @labarba

@dpsanders and @JoshKarpel the package documentation is under progress (https://uposham.readthedocs.io/en/latest/#) since the recent PR contains the updated docstrings and changes to calling the scripts from the command line.

Test (comparing the analytical and numerical solution) in the form of a notebook will also reflect these changes.
