+++ docs/source/index.rst
@@ -1,0 +1,21 @@
+.. UPOsHam documentation master file, created by
+   sphinx-quickstart on Thu Nov 28 12:01:02 2019.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+Welcome to UPOsHam's documentation!
+===================================
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+
+   LICENSE
+
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
