+++ docs/doc.rst
@@ -1,0 +1,20 @@
+
+Unstable Periodic Orbits in Hamiltonian systems
+===============================================
+
+
+Methods
+-------
+
+* Turning point
+* Turning point based on configuratio difference
+* Differential correction
+
+
+Problems
+--------
+
+* Uncoupled quartic Hamiltonian
+* Coupled quartic Hamiltonian
+* De Leon-Berne Hamiltonian
+
