+++ docs/index.rst
@@ -1,19 +1,18 @@
 .. UPOsHam documentation master file, created by
-   sphinx-quickstart on Mon Oct 14 15:36:09 2019.
+   sphinx-quickstart on Thu Nov 28 12:44:51 2019.
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
 
 Welcome to UPOsHam's documentation!
 ===================================
-UPOsHam is a Python package with a collection of three methods for computing unstable periodic orbits in Hamiltonian systems that model a diverse array of problems in physical sciences and engineering
-
 
 .. toctree::
    :maxdepth: 2
    :caption: Contents:
 
+   doc
 
-
+   
 Indices and tables
 ==================
 
