+++ index.rst
@@ -1,20 +1,0 @@
-.. UPOsHam documentation master file, created by
-   sphinx-quickstart on Mon Nov 18 11:47:51 2019.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
-
-Welcome to UPOsHam's documentation!
-===================================
-
-.. toctree::
-   :maxdepth: 2
-   :caption: Contents:
-
-
-
-Indices and tables
-==================
-
-* :ref:`genindex`
-* :ref:`modindex`
-* :ref:`search`
