### 2019-11-28 1d93fcd
set up autodocs
