+++ docs/conf.py
@@ -31,7 +31,9 @@ release = '0.1.0'
 # Add any Sphinx extension module names here, as strings. They can be
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
-extensions = ['recommonmark']
+extensions = ['sphinx.ext.autodoc','sphinx.ext.intersphinx','recommonmark']
+
+intersphinx_mapping = {'python': ('https://docs.python.org/3', None)}
 
 source_suffix = ['.rst']
 
