+++ docs/index.rst
@@ -7,7 +7,7 @@ Welcome to UPOsHam's documentation!
 ===================================
 
 .. toctree::
-   :maxdepth: 2
+   :maxdepth: 3
    :caption: Contents:
 
    doc
