### 2019-11-28 07f90ac
set up autodocs
