### 2019-11-28 81c296f
update contributing for docs
