+++ docs/doc.rst
@@ -4,23 +4,26 @@ Unstable Periodic Orbits in Hamiltonian systems
 
 
 Methods
--------
+=======
 
-* Turning point
+Turning point
+-------------
    .. automodule:: turning_point
       :members:
 
-* Turning point based on configuration difference
+Turning point based on configuration difference
+-----------------------------------------------
    .. automodule:: turning_point_coord_difference
       :members:
 
-* Differential correction
+Differential correction
+-----------------------
    .. automodule:: differential_correction
       :members:
 
 
 Problems
---------
+========
 
 * Uncoupled quartic Hamiltonian
 * Coupled quartic Hamiltonian
