+++ docs/examples.rst
@@ -1,0 +1,9 @@
+
+De Leon-Berne Hamiltonian
+=========================
+
+
+
+Coupled and uncouple quartic Hamiltonian
+========================================
+
