### 2019-11-29 0c494d6
clean up old conf and index files
