+++ README.Rmd
@@ -21,7 +21,7 @@ biblio-style: apalike
 bibliography: ./paper/paper.bib
 ---
 
-[![Documentation Status](https://readthedocs.org/projects/uposham/badge/?version=latest)](https://uposham.readthedocs.io/en/latest/?badge=latest)
+[![Documentation Status](https://uposham.readthedocs.io/en/latest/?badge=latest)](https://readthedocs.org/projects/uposham/badge/?version=latest)
       
 [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.3373396.svg)](https://doi.org/10.5281/zenodo.3373396)
 
