### 2019-12-04 4df8b1f
add docs passing badge
