### 2019-12-04 e1b3fea
add docs passing badge
