+++ README.Rmd
@@ -3,6 +3,8 @@ title: "UPOsHam"
 bibliography: ./paper/paper.bib
 link-citations: yes
 output:
+  html_document:
+    toc: yes
   md_document:
     fig_height: 6
     fig_width: 6
@@ -19,7 +21,6 @@ biblio-style: apalike
 [![Documentation Status](https://readthedocs.org/projects/uposham/badge/?version=latest)](https://uposham.readthedocs.io/en/latest/?badge=latest)
       
 [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.3373396.svg)](https://doi.org/10.5281/zenodo.3373396)
-
 
 ## Summary
 
@@ -82,6 +83,11 @@ Other tests can be performed by running the cells in the ./tests/tests.ipynb
 ## Contributing
 
 Guidelines on how to contribute to this package can be found  [here](https://github.com/WyLyu/UPOsHam/blob/master/docs/contributing.md) along with the code of conduct [here](https://github.com/WyLyu/UPOsHam/blob/master/CODE_OF_CONDUCT.md) for engaging with the fellow contributors.
+
+
+## Acknowledgements
+
+We acknowledge the support of EPSRC Grant No. EP/P021123/1 and Office of Naval Research (Grant No. N00014-01-1-0769). The authors would like to acknowledge the London Mathematical Society and School of Mathematics at the University of Bristol for supporting the undergraduate research bursary 2019. We acknowledge contributions from Shane Ross for writing the early MATLAB version of the differential correction and numerical continuation code.
 
 
 ## Copyright and License
