+++ README.Rmd
@@ -3,13 +3,13 @@ title: "UPOsHam"
 bibliography: ./paper/paper.bib
 link-citations: yes
 output:
-  html_document:
-    toc: yes
   md_document:
     fig_height: 6
     fig_width: 6
     toc: yes
     variant: markdown_github
+  html_document:
+    toc: yes
   pdf_document:
     fig_height: 6
     fig_width: 6
@@ -43,17 +43,18 @@ and check the modules shown in [requirements.txt](https://github.com/WyLyu/UPOsH
 
 ## Usage
 
-The example scripts that computes the unstable periodic orbits (UPOs) for a given Hamiltonian system is in the [examples](examples/) directory with names ```<method>_<system>.py```. For example, to compute the unstable periodic orbits (the default shows at 2 total energy levels) for the De Leon-Berne Hamiltonian [-@Deleon_Berne_1981] using the differential correction method can be run from the command line by calling
+The example scripts that computes the unstable periodic orbits (UPOs) for a given Hamiltonian system is in the [examples](examples/) directory with names ```<method>_<system>.py```. For example, to compute the unstable periodic orbits (the default setting computes the UPOs for 2 values of the total energy) for the De Leon-Berne Hamiltonian [-@Deleon_Berne_1981] using the differential correction method, from the root directory call the demonstration script at the command-line usingg
 
 ```
 $ python examples/differential_correction_deleonberne.py
 ```
 
-This will save data files in the current directory and save the following plot in the [plots](tests/plots/) directory.
+This will save data files in the root directory and save the following plot in the [plots](tests/plots/) directory.
 
-![](tests/plots/diff_corr_deleonberne_upos.png){width=80%}
 
-__Fig. Unstable periodic orbits for De Leon-Berne Hamiltonian__
+![](tests/plots/diff_corr_deleonberne_upos.png){width="80.00000%"}
+
+__Fig. 1. Unstable periodic orbits for De Leon-Berne Hamiltonian__
 
 
 The example scripts import the system specific functions and method specific functions to obtain the UPOs. These are written as illustration of how to apply these methods to new systems. 
