+++ docs/theory.Rmd
@@ -1,0 +1,29 @@
+---
+title: "UPOsHam"
+bibliography: ../paper/paper.bib
+link-citations: yes
+output:
+  md_document:
+    fig_height: 6
+    fig_width: 6
+    toc: yes
+    variant: markdown_github
+  html_document:
+    toc: yes
+  pdf_document:
+    fig_height: 6
+    fig_width: 6
+    highlight: zenburn
+    toc: yes
+biblio-style: apalike
+---
+
+# Introduction
+
+This Python package is a collection of three methods for computing unstable periodic orbits in two degrees of freedom Hamiltonian systems that model a diverse array of problems in physical sciences and engineering [@Parker_1989; @wiggins_introduction_2003]. The unstable periodic orbits exist in the bottleneck of the equipotential line $V(x,y) = E$ and project as lines on the configuration space $(x,y)$ for the Hamiltonian system of the form kinetic plus potential energy [@wiggins_role_2016]. The three methods implemented here (and available under [src](src/) directory) has been used in [@Pollak_1980;@Deleon_Berne_1981;@koon_heteroclinic_2000;@Koon2011;@naik_geometry_2017;@ross_experimental_2018;@naik_finding_2019b] for transition dynamics in chemical reactions, celestial mechanics, and ship capsize. We have chosen three Hamiltonian systems that have two wells connected by a bottleneck where the unstable periodic orbits exist for energy above the energy of the saddle equilibrium point. 
+
+
+
+
+
+
