### 2019-12-05 aab63d8
update docs, add theory
