### 2019-12-05 d473685
update docs: add theory
