+++ docs/conf.py
@@ -38,6 +38,7 @@ extensions = ['sphinx.ext.autodoc',
     'sphinx.ext.viewcode', \
     'sphinx.ext.napoleon', \
     'sphinx.ext.doctest', \
+    'sphinx.ext.mathjax', \
     'm2r']
 
 intersphinx_mapping = {'python': ('https://docs.python.org/3', None)}
@@ -67,3 +68,6 @@ html_theme = 'sphinx_rtd_theme'
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
 html_static_path = ['_static']
+
+
+
