+++ docs/doc.rst
@@ -2,7 +2,6 @@
 Unstable Periodic Orbits in Hamiltonian systems
 ===============================================
 
-.. image:: ../tests/plots/diff_corr_deleonberne_upos.png
 
 .. toctree::
    :maxdepth: 2
