+++ docs/theory.md
@@ -1,7 +1,6 @@
 -   [Introduction](#introduction)
-    -   [Examples](#examples)
-        -   [DeLeon-Berne Hamiltonian](#dbham)
-        -   [Coupled and uncoupled quartic Hamiltonian](#coupled-and-uncoupled-quartic-hamiltonian)
+    -   [DeLeon-Berne Hamiltonian](#dbham)
+    -   [Coupled quartic Hamiltonian](#coupled-quartic-hamiltonian)
 -   [References](#references)
 
 Introduction
@@ -9,14 +8,11 @@ Introduction
 
 Computation of unstable periodic orbits in two degrees of freedom Hamiltonian systems arises in a diverse array of problems in physical sciences and engineering (Parker and Chua [1989](#ref-Parker_1989); Wiggins [2003](#ref-wiggins_introduction_2003)). The unstable periodic orbits exist in the bottleneck of the equipotential line *V*(*x*, *y*)=*E* and project as lines on the configuration space (*x*, *y*) for the Hamiltonian system of the form kinetic plus potential energy (Wiggins [2016](#ref-wiggins_role_2016)). The three methods
 
--   Differential correction
--   Turning point
+-   Differential correction (Wang Sang Koon et al. [2000](#ref-koon_heteroclinic_2000); W. S. Koon et al. [2011](#ref-Koon2011); Naik and Ross [2017](#ref-naik_geometry_2017); Ross et al. [2018](#ref-ross_experimental_2018); Naik and Wiggins [2019](#ref-naik_finding_2019b))
+-   Turning point (Pollak, Child, and Pechukas [1980](#ref-Pollak_1980); Nelson De Leon and Berne [1981](#ref-Deleon_Berne_1981))
 -   Turning point based on configuration difference
 
-implemented has been used in (Pollak, Child, and Pechukas [1980](#ref-Pollak_1980); Nelson De Leon and Berne [1981](#ref-Deleon_Berne_1981); Wang Sang Koon et al. [2000](#ref-koon_heteroclinic_2000); W. S. Koon et al. [2011](#ref-Koon2011); Naik and Ross [2017](#ref-naik_geometry_2017); Ross et al. [2018](#ref-ross_experimental_2018); Naik and Wiggins [2019](#ref-naik_finding_2019b)) for transition dynamics in chemical reactions, celestial mechanics, and ship capsize. We have included Hamiltonian systems with two potential wells connected by a bottleneck where the unstable periodic orbits exist for energy above the energy of the saddle equilibrium point.
-
-Examples
---------
+implemented in this package has been used for transition dynamics in chemical reactions, celestial mechanics, and ship capsize. We have included Hamiltonian systems with two potential wells connected by a bottleneck where the unstable periodic orbits exist for energy above the energy of the saddle equilibrium point.
 
 Consider the following two degrees-of-freedom Hamiltonian model where *x*, *y* are configuration space coordinates and *p*<sub>*x*</sub>, *p*<sub>*y*</sub> are corresponding momenta, *V*(*x*, *y*) is the potential energy, and *T*(*x*, *y*) is the kinetic energy.
 
@@ -24,28 +20,15 @@ Consider the following two degrees-of-freedom Hamiltonian model where *x*, *y*
 
[…]
