+++ examples/differential_correction_coupled.py
@@ -219,5 +219,7 @@ if show_final_plot:
 if save_final_plot:  
     plt.savefig('./tests/plots/diff_corr_coupled_upos.pdf', format='pdf', \
                         bbox_inches='tight')
+    plt.savefig('./tests/plots/diff_corr_coupled_upos.png', dpi = 300, \
+                bbox_inches='tight')
 
 
