+++ examples/uncoupled_quartic_hamiltonian.py
@@ -1,8 +1,10 @@
 # -*- coding: utf-8 -*-
 """
 Created on Tue Jul 30 10:02:48 2019
+
 @author: Wenyang Lyu and Shibabrat Naik
-Script to define expressions for the coupled quartic Hamiltonian
+
+Script to define expressions for the uncoupled quartic Hamiltonian
 """
 
 import numpy as np
@@ -96,9 +98,7 @@ def variational_eqns_uncoupled(t,PHI,par):
     Returns the state transition matrix , PHI(t,t0), where Df(t) is the Jacobian of the 
     Hamiltonian vector field
     
-    d PHI(t, t0)
-    ------------ =  Df(t) * PHI(t, t0)
-        dt
+    d PHI(t, t0)/dt =  Df(t) * PHI(t, t0)
     
     """
     
