+++ src/turning_point.py
@@ -194,12 +194,8 @@ def state_transit_matrix(tf,x0,par,variational_eqns_model,fixed_step=0):
 
 
 #%%
-<<<<<<< HEAD
 def dotproduct(guess1, guess2,n_turn, ham2dof_model, half_period_model, \
                 varEqns_model, par):
-=======
-def dotproduct(guess1, guess2,n_turn, ham2dof_model, half_period_model, variational_eqns_model, par):
->>>>>>> master
     """
     Returns x,y coordinates of the turning points for guess initial conditions guess1, guess2 and the defined product product for the 2 turning points
     
