+++ docs/doc.rst
@@ -2,11 +2,9 @@
 Unstable Periodic Orbits in Hamiltonian systems
 ===============================================
 
-
-.. toctree::
-   :maxdepth: 2
-
 .. mdinclude:: theory.md
+
+
 
 
  
