+++ docs/index.rst
@@ -17,19 +17,19 @@ in two degrees of freedom Hamiltonian (*Ham*) systems
 
 .. toctree::
    :maxdepth: 2
-   :caption: Methods:
+   :caption: Methods
 
    methods
 
 .. toctree::
    :maxdepth: 2
-   :caption: Examples:
+   :caption: Examples
 
    examples
 
 .. toctree::
    :maxdepth: 2
-   :caption: Contributing and reporting:
+   :caption: Contributing and reporting
 
    contributing
 
