+++ docs/examples.rst
@@ -2,8 +2,20 @@
 De Leon-Berne Hamiltonian
 =========================
 
+.. automodule:: deleonberne_hamiltonian
+   :members:
 
 
-Coupled and uncouple quartic Hamiltonian
-========================================
+Coupled quartic Hamiltonian
+===========================
+
+.. automodule:: coupled_quartic_hamiltonian
+   :members:
+
+
+Uncoupled quartic Hamiltonian
+=============================
+
+.. automodule:: uncoupled_quartic_hamiltonian
+   :members:
 
