+++ tests/compare_methods_coupled.py
@@ -159,7 +159,7 @@ eSaddle = differential_correction.get_total_energy([eqPt[0],eqPt[1],0,0], pot_en
 deltaE = 0.10
 eSaddle = 0.0 # energy of the saddle
 
-data_path = "./data/"
+data_path = "../data/"
 po_fam_file = "x0_tpcd_deltaE%s_coupled.dat" %(deltaE)
 print('Loading the periodic orbit family from data file',po_fam_file,'\n') 
 x0podata = np.loadtxt(data_path + po_fam_file)
@@ -191,13 +191,7 @@ soln = solve_ivp(f, TSPAN, x0po_1_tpcd[-1,0:4],method='RK45',dense_output=True, 
                  rtol=RelTol, atol=AbsTol)
 te = soln.t_events[0]
 tt = [0,te[2]]
-t,x,phi_t1,PHI = differential_correction.state_transit_matrix(tt, x0po_1_tpcd[-1,0:4], parameters, \
-                                                      variational_eqns_coupled)
-
-ax = plt.gca(projection='3d')
-ax.plot(x[:,0],x[:,1],x[:,3],'-',label='$\Delta E$ = 0.1, using tpcd')
-ax.scatter(x[0,0],x[0,1],x[0,3],s=20,marker='*')
-ax.plot(x[:,0], x[:,1], zs=0, zdir='z')
+t,x_tpcd,phi_t1,PHI = differential_correction.state_transit_matrix(tt, x0po_1_tpcd[-1,0:4], parameters, variational_eqns_coupled)
 
 
 f = lambda t,x: ham2dof_coupled(t,x,parameters)
@@ -207,12 +201,9 @@ soln = solve_ivp(f, TSPAN, x0po_1_diffcorr,method='RK45',dense_output=True, \
 
 te = soln.t_events[0]
 tt = [0,te[2]]
-t,x,phi_t1,PHI = differential_correction.state_transit_matrix(tt, x0po_1_tpcd[-1,0:4], parameters, \
-                                                      variational_eqns_coupled)
+t,x_diffcorr,phi_t1,PHI = differential_correction.state_transit_matrix(tt, x0po_1_diffcorr, parameters, variational_eqns_coupled)
 
-ax.plot(x[:,0],x[:,1],x[:,3],':',label='$\Delta E$ = 0.1, using dcnc')
-ax.scatter(x[0,0],x[0,1],x[0,3],s=20,marker='*')
-ax.plot(x[:,0], x[:,1], zs=0, zdir='z')
+
 
 f = lambda t,x: ham2dof_coupled(t,x,parameters)
 soln = solve_ivp(f, TSPAN, x0po_1_turningpoint[-1,0:4], method='RK45',dense_output=True, \
@@ -221,16 +212,24 @@ soln = solve_ivp(f, TSPAN, x0po_1_turningpoint[-1,0:4], method='RK45',dense_outp
 
 te = soln.t_events[0]
 tt = [0,te[2]]
-t,x,phi_t1,PHI = differential_correction.state_transit_matrix(tt, x0po_1_turningpoint[-1,0:4], parameters, \
-                                                      variational_eqns_coupled)
-
-ax.plot(x[:,0],x[:,1],x[:,3],'-.',label='$\Delta E$ = 0.1, using tp')
-ax.scatter(x[0,0],x[0,1],x[0,3],s=20,marker='*')
-ax.plot(x[:,0], x[:,1], zs=0, zdir='z')
[…]
