### 2019-12-10 #1684 @Shibabrat
@kthyng @dpsanders and @JoshKarpel we have made some recent updates to the documentation and tests (this is available in the test directory as tests.ipynb) that are compatible with the module structure of the methods.

The correctness test will soon be restructured as an assert test along with the visual comparison (at the moment).

API documentation will feature example code snippets and reference to a script that solves the provided example systems using a method.

We would appreciate some feedback on these new developments.
