### 2019-12-10 WyLyu/UPOsHam#47 @JoshKarpel
Improve documentation for example Hamiltonians
Some general feedback on https://uposham.readthedocs.io/en/latest/examples.html

- It would be nice if each Hamiltonian was on its own page.
- Each Hamiltonian should have some descriptive text about what it might be used for, and how its properties are expressed by the functions you're defining.
- Docstrings like https://uposham.readthedocs.io/en/latest/examples.html#deleonberne_hamiltonian.diffcorr_setup_deleonberne are not very helpful. Document what the expected output of this kind of function is, and what each of the settings mean.
- **I think the big thing that's still missing is a plain-text description of how the solution algorithms and the Hamiltonians fit together.** As a user, I want to have some high-level understanding of how the problem definition I'm writing is going to be used by the algorithm to solve the problem (not necessarily at the level of individual functions, but as a general process).

(Part of https://github.com/openjournals/joss-reviews/issues/1684)
