### 2019-12-10 WyLyu/UPOsHam#48 @JoshKarpel
Improve documentation for methods
General feedback on https://uposham.readthedocs.io/en/latest/methods.html

- It would be nice if each method was on its own page.
- Each method should have a description of the general algorithm, describing how each function that needs to be defined fits into it.
- Make it clear that this is internal code that is useful as reference for implementing other algorithms, but that users shouldn't need to modify it to work with their Hamiltonian.

(Part of https://github.com/openjournals/joss-reviews/issues/1684)
