### 2019-12-10 #1684 @JoshKarpel
I submitted some new issues with high-level feedback on the docs.

The tests are moving in the right direction; very happy to hear that you'll be able to do an actual assertion. There should be some information in the README about how to run the tests and how to do any manual verification, but that can wait until the structure settles down.
