### 2019-12-12 43928a6
update readme about tests
