### 2019-12-12 e1fc8d3
Separate methods into different sections
