### 2019-12-12 6292929
Separate methods into different sections
