### 2019-12-12 d8fed5e
add brief description of each method
