+++ docs/deleonberne_ham.rst
@@ -1,0 +1,7 @@
+De Leon-Berne Hamiltonian
+=========================
+
+.. automodule:: deleonberne_hamiltonian
+   :members:
+
+
