### 2019-12-13 1359fa2
add details to the examples
