+++ paper/paper.bib
@@ -122,3 +122,15 @@
 	year = {1989},
 	doi = {10.1063/1.456915}}
 
+@article{marston_reactive_1989,
+	title = {Reactive islands as essential mediators of unimolecular conformational isomerization: {A} dynamical study of 3‐phospholene},
+	volume = {91},
+	doi = {10.1063/1.456914},
+	number = {6},
+	journal = {The Journal of Chemical Physics},
+	author = {Marston, C. Clay and De Leon, N.},
+	month = sep,
+	year = {1989},
+	pages = {3392--3404}
+}
+
