### 2019-12-13 2521dbe
minor changes in docs: clean up and cross-ref
