+++ examples/differential_correction_deleonberne.py
@@ -1,12 +1,12 @@
 # -*- coding: utf-8 -*-
-"""
-Created on Tue Jul 30 10:02:48 2019
+# """
+# Created on Tue Jul 30 10:02:48 2019
 
-@author: Wenyang Lyu and Shibabrat Naik
+# @author: Wenyang Lyu and Shibabrat Naik
 
-Script to compute unstable periodic orbits at specified energies for the coupled quartic 
-Hamiltonian using differential correction
-"""
+# Script to compute unstable periodic orbits at specified energies for the coupled quartic 
+# Hamiltonian using differential correction
+# """
 
 import numpy as np
 import matplotlib.pyplot as plt
