+++ src/turning_point.py
@@ -1,9 +1,9 @@
 # -*- coding: utf-8 -*-
-"""
-Created on Thu Sep 12 17:34:06 2019
+# """
+# Created on Thu Sep 12 17:34:06 2019
 
-@author: Wenyang Lyu and Shibabrat Naik
-"""
+# @author: Wenyang Lyu and Shibabrat Naik
+# """
 
 import numpy as np
 import matplotlib.pyplot as plt
