+++ examples/turning_pt_coord_diff_uncoupled.py
@@ -1,9 +1,9 @@
 # -*- coding: utf-8 -*-
-"""
-Created on Wed Sep 11 13:16:47 2019
+# """
+# Created on Wed Sep 11 13:16:47 2019
 
-@author: Wenyang Lyu and Shibabrat Naik
-"""
+# @author: Wenyang Lyu and Shibabrat Naik
+# """
 
 import numpy as np
 import matplotlib.pyplot as plt
