+++ docs/theory.Rmd
@@ -34,27 +34,18 @@ implemented in this package has been used for transition dynamics in chemical re
 
 In the sections below, we briefly describe the Hamiltonian systems that are used to demonstrate the methods mentioned in [Introduction](#introduction). 
 
-Consider the following two degrees-of-freedom Hamiltonian model where $x, y$ are configuration space coordinates and $p_x,p_y$ are corresponding momenta, $V(x,y)$ is the potential energy, and $T(x,y)$ is the kinetic energy.
+Consider the following two degrees-of-freedom Hamiltonian model where $x, y$ are configuration space coordinates and $p_x$ and $p_y$ are corresponding momenta, $V(x,y)$ is the potential energy, and $T(x,y)$ is the kinetic energy.
 
 ## De Leon-Berne Hamiltonian {#dbham}
 
 This Hamiltonian has been studied as a model of isomerization of a single molecule that undergoes conformational change and exhibits regular and chaotic dynamics relevant for chemical reactions [@Deleon_Berne_1981; @DeLeon_Marston_1989; @marston_reactive_1989]. 
 
 
-
-![Unstable periodic orbits](../tests/plots/diff_corr_deleonberne_upos.png)
-
-Unstable periodic orbits at two values of the total energy for the De Leon-Berne Hamiltonian computed using differential correction method.
-
-
-## Coupled quartic Hamiltonian
+## Uncoupled and coupled quartic Hamiltonian
 
 This Hamiltonian models a chemical reaction where the _molecule_ reacts by crossing the barrier separating the double-well (two molecular configurations which are denoted by the bottom of the well) potential and is coupled with the _solvent_ represented by the harmonic oscillator.
 
 
-![](../tests/plots/diff_corr_coupled_upos.png)
-
-Unstable periodic orbits at two values of the total energy for the coupled quartic Hamiltonian computed using differential correction method.
 
 
 
