### 2019-12-13 ae02bdf
minor changes to format in readme
