### 2019-12-13 7d3a003
add test for turning point
