### 2019-12-15 5d80c52
update test with turning point based on configuration difference
