+++ tests/test_upo_uncoupled_ham.py
@@ -38,58 +38,58 @@ def hausd_dist_numeric_analytic(orbit, t, deltaE_val, parameters):
 class TestUnstablePeriodicOrbit(unittest.TestCase):
     """ Test unstable periodic orbit for the uncoupled system """
 
-    # def test_differential_correction(self):
-    #     """ Test solution obtained using differential correction """
+    def test_differential_correction(self):
+        """ Test solution obtained using differential correction """
     
-    #     MASS_A = 1.0
-    #     MASS_B = 1.0
+        MASS_A = 1.0
+        MASS_B = 1.0
              
-    #     SADDLE_ENERGY = 0.0 # Energy of the saddle
-    #     ALPHA = 1.00
-    #     BETA = 1.00
-    #     OMEGA = 1
-    #     EPSILON = 0.00 # uncoupled
-    #     parameters = np.array([MASS_A, MASS_B, SADDLE_ENERGY, \
-    #                         ALPHA, BETA, OMEGA, EPSILON])
+        SADDLE_ENERGY = 0.0 # Energy of the saddle
+        ALPHA = 1.00
+        BETA = 1.00
+        OMEGA = 1
+        EPSILON = 0.00 # uncoupled
+        parameters = np.array([MASS_A, MASS_B, SADDLE_ENERGY, \
+                            ALPHA, BETA, OMEGA, EPSILON])
 
 
-    #     deltaE_vals = [0.1]
-    #     linecolors = ['b']
-    #     diff_corr_unc.upo(deltaE_vals, linecolors)
+        deltaE_vals = [0.1]
+        linecolors = ['b']
+        diff_corr_unc.upo(deltaE_vals, linecolors)
         
-    #     data_path = "./data/"
-    #     TSPAN = [0, 30] # arbitrary range, just to finish the integration
-    #     RelTol = 3.e-12
-    #     AbsTol = 1.e-12
+        data_path = "./data/"
+        TSPAN = [0, 30] # arbitrary range, just to finish the integration
+        RelTol = 3.e-12
+        AbsTol = 1.e-12
 
-    #     for deltaE_val in deltaE_vals:
-    #         po_fam_file = "x0_diffcorr_deltaE%s_uncoupled.dat" %(deltaE_val)
-    #         x0podata = np.loadtxt(data_path + po_fam_file)
-    #         x0po_diffcorr = x0podata[0:4]
+        for deltaE_val in deltaE_vals:
+            po_fam_file = "x0_diffcorr_deltaE%s_uncoupled.dat" %(deltaE_val)
+            x0podata = np.loadtxt(data_path + po_fam_file)
+            x0po_diffcorr = x0podata[0:4]
 
-    #         f = lambda t,x: uncoupled.ham2dof_uncoupled(t,x,parameters)
-    #         soln = solve_ivp(f, TSPAN, x0po_diffcorr,method='RK45',dense_output=True, \
-    #              events = lambda t,x : uncoupled.half_period_uncoupled(t,x,parameters), \
-    #              rtol=RelTol, atol=AbsTol)
+            f = lambda t,x: uncoupled.ham2dof_uncoupled(t,x,parameters)
+            soln = solve_ivp(f, TSPAN, x0po_diffcorr,method='RK45',dense_output=True, \
+                 events = lambda t,x : uncoupled.half_period_uncoupled(t,x,parameters), \
+                 rtol=RelTol, atol=AbsTol)
 
-    #         te = soln.t_events[0]
-    #         tt = [0,te[2]]
-    #         t,x_diffcorr,phi_t1,PHI = diffcorr.state_transit_matrix(
-    #             tt, x0po_diffcorr, parameters, \
-    #             uncoupled.variational_eqns_uncoupled)
+            te = soln.t_events[0]
+            tt = [0,te[2]]
[…]
