### 2019-12-15 a58421f
update test suite
