### 2019-12-16 897cfbc
add methods info and organize
