+++ docs/theory.Rmd
@@ -20,14 +20,17 @@ biblio-style: apalike
 
 # Introduction {#introduction}
 
-Computation of unstable periodic orbits in two degrees of freedom Hamiltonian systems arises in a diverse array of problems in physical sciences and engineering [@Parker_1989; @wiggins_introduction_2003]. The unstable periodic orbits exist in the bottleneck of the equipotential line $V(x,y) = E$ and project as lines on the configuration space $(x,y)$ for the Hamiltonian system of the form kinetic plus potential energy [@wiggins_role_2016]. The three methods are
+Computation of unstable periodic orbits in two degrees of freedom Hamiltonian systems arises in a diverse array of problems in physical sciences and engineering [@Parker_1989; @wiggins_introduction_2003]. The unstable periodic orbits exist in the bottleneck of the equipotential line $V(x,y) = E$ and project as lines on the configuration space $(x,y)$ for the Hamiltonian system of the form kinetic plus potential energy [@wiggins_role_2016]. The methods implemented in this package has been used for transition dynamics in chemical reactions, celestial mechanics, and ship capsize. We have included Hamiltonian systems with two potential wells connected by a bottleneck where the unstable periodic orbits exist for energy above the energy of the saddle equilibrium point [@wiggins_introduction_2003]. 
+
+
+# Available methods {#methods}
+
+This section gives a broad overview of the methods as a reference for implementing new methods and as how to guide to solve new systems. The methods are implemented as modules and are part of the internal code that do not require modification. More details can be found in the papers listed below:
 
 - Differential correction [@koon_heteroclinic_2000;@Koon2011;@naik_geometry_2017;@ross_experimental_2018;@naik_finding_2019b]
 - Turning point [@Pollak_1980;@Deleon_Berne_1981]
-- Turning point based on configuration difference
-
-implemented in this package has been used for transition dynamics in chemical reactions, celestial mechanics, and ship capsize. We have included Hamiltonian systems with two potential wells connected by a bottleneck where the unstable periodic orbits exist for energy above the energy of the saddle equilibrium point [@wiggins_introduction_2003]. 
-
+- Turning point based on configuration difference 
+  * this is a modification of the turning point method and does not rely on dot product computation.
 
 
 # Examples 
