### 2019-12-16 3b5b61d
Structure docs and add description of examples
