+++ docs/index.rst
@@ -6,13 +6,12 @@
 Homepage of UPOsHam’s documentation
 ===================================
 
-Welcome to the documentation for computing Unstable Periodic Orbits (*UPOs*) 
-in two degrees of freedom Hamiltonian (*Ham*) systems
+Welcome to the documentation for computing Unstable Periodic Orbits (*UPOs*) in two degrees of freedom Hamiltonian (*Ham*) systems
 
-.. _sect-theory:
+
 .. toctree::
    :maxdepth: 2
-   :caption: Theory
+   :caption: Introduction
 
    doc
 
