+++ README.md
@@ -1,79 +1,205 @@
--   [Summary](#summary)
--   [Installation](#installation)
-    -   [Methods and applications](#methods-and-applications)
--   [Usage](#usage)
--   [Contributing](#contributing)
--   [Copyright and License](#copyright-and-license)
--   [References](#references)
+-   [UPOsHam](#uposham)
+    -   [Summary](#summary)
+    -   [Installation](#installation)
+    -   [Usage](#usage)
+    -   [Tests](#tests)
+    -   [Contributing](#contributing)
+    -   [Acknowledgements](#acknowledgements)
+    -   [Copyright and License](#copyright-and-license)
+    -   [References](#references)
 
-[![Documentation Status](https://readthedocs.org/projects/uposham/badge/?version=latest)](https://uposham.readthedocs.io/en/latest/?badge=latest)
+UPOsHam
+=======
 
+[![Documentation
+Status](https://readthedocs.org/projects/uposham/badge/?version=latest)](https://uposham.readthedocs.io/en/latest/?badge=latest)
 [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.3373396.svg)](https://doi.org/10.5281/zenodo.3373396)
 
 Summary
 -------
 
-This Python package is a collection of three methods for computing unstable periodic orbits in Hamiltonian systems that model a diverse array of problems in physical sciences and engineering (Wiggins 2003,Parker and Chua (1989)). The unstable periodic orbits exist in the bottleneck of the equipotential line *V*(*x*, *y*)=*E* and project as lines on the configuration space (*x*, *y*) (Wiggins 2016). The three methods described below have been implemented for three Hamiltonian systems of the form kinetic plus potential energy (Pollak, Child, and Pechukas 1980,Koon et al. (2011),Naik and Wiggins (2019)). A brief description of the methods and the applications can be found in the [paper.pdf](https://github.com/WyLyu/UPOsHam/tree/master/paper/paper.pdf). The scripts are written as demonstration of how to modify and adapt the code for a problem of interest.
+This Python package is a collection of three methods for computing
+unstable periodic orbits in two degrees of freedom Hamiltonian systems
+that model a diverse array of problems in physical sciences and
+engineering (Parker and Chua [1989](#ref-Parker_1989); Wiggins
+[2003](#ref-wiggins_introduction_2003)). The unstable periodic orbits
+exist in the bottleneck of the equipotential line *V*(*x*, *y*) = *E*
+and project as lines on the configuration space (*x*, *y*) for the
+Hamiltonian system of the form kinetic plus potential energy (Wiggins
+[2016](#ref-wiggins_role_2016)). The three methods implemented here (and
+available under [src](src/) directory) has been used in (Pollak, Child,
+and Pechukas [1980](#ref-Pollak_1980); De Leon and Berne
+[1981](#ref-Deleon_Berne_1981); <span class="citeproc-not-found"
+data-reference-id="koon_heteroclinic_2000">**???**</span>; <span
+class="citeproc-not-found"
+data-reference-id="naik_geometry_2017">**???**</span>; <span
+class="citeproc-not-found"
+data-reference-id="ross_experimental_2018">**???**</span>; Koon et al.
+[2011](#ref-Koon2011); Naik and Wiggins [2019](#ref-naik_finding_2019b))
+for transition dynamics in chemical reactions, celestial mechanics, and
+ship capsize. We have chosen three Hamiltonian systems that have two
+wells connected by a bottleneck where the unstable periodic orbits exist
+for energy above the energy of the saddle equilibrium point. A brief
[…]
