+++ README.Rmd
@@ -44,10 +44,10 @@ and check the modules shown in [requirements.txt](https://github.com/WyLyu/UPOsH
 
 ## Usage
 
-To obtain the unstable periodic orbits for a two degree-of-freedom Hamiltonian system, one needs the expressions described in the examples section of the ![documentation](https://readthedocs.org/projects/uposham/badge/?version=latest). These expression can then be implemented as shown in the [script](examples/deleonberne_hamiltonian.py) and the function names of the expressions can then be passed to a method's function as in the [script](examples/differential_correction_deleonberne.py)
+To obtain the unstable periodic orbits for a two degree-of-freedom Hamiltonian system, one needs the expressions described in the examples section of the ![documentation](https://uposham.readthedocs.io/en/latest/?badge=latest). These expression can then be implemented as shown in the [script](examples/deleonberne_hamiltonian.py) and the function names of the expressions can then be passed to a method's function as in the [script](examples/differential_correction_deleonberne.py)
 
 
-## Tests
+__Solving the De Leon-Berne Hamiltonian using differential correction method__
 
 The example scripts that computes the unstable periodic orbits (UPOs) for a given Hamiltonian system is in the [examples](examples/) directory with names ```<method>_<system>.py```. For example, to test the computation of unstable periodic orbits (the default setting computes the UPOs for 2 values of the total energy) for the De Leon-Berne Hamiltonian [-@Deleon_Berne_1981] using the differential correction method, from the root directory call the demonstration script at the command-line using
 
@@ -63,16 +63,22 @@ This will save data files in the root directory and save the plot in the [./test
 __Fig. 1. Unstable periodic orbits for De Leon-Berne Hamiltonian__
 
 
-The example scripts import the system specific functions and method specific functions to obtain the UPOs. These are written as illustration of how to apply these methods to new systems. 
+__Solving the coupled quartic Hamiltonian using differential correction__
 
-To obtain the unstable periodic orbits for a specific model Hamiltonian using a specific method, one uses
+The example scripts show how to implement the Hamiltonian system specific functions and pass them to the method specific functions. These scripts are written as demonstration of how to apply these methods to a new system using the available methods. 
+
+For example, to obtain the unstable periodic orbits for a specific model Hamiltonian using a specific method, one uses
 
 ```bash
 $ ipython
->>> run ./examples/diffcorr_UPOs_coupled.py
+>>> run ./examples/differential_correction_coupled.py
 ```
 
+## Tests
+
 The uncoupled quartic Hamiltonian is an example of a system where the unstable periodic orbits can be obtained in explicit analytical form. This system is used for testing the numerical solutions obtained using the methods implemented in this package. 
+
+__Unit test__
 
 Passing a test constitutes the following:
 
@@ -81,6 +87,7 @@ Passing a test constitutes the following:
 2. The element wise comparison test of the orbits agree upto 6 decimal places. This is performed using Numpy's testing module. 
 
 Other nonlinear examples systems can not be solved in this explicit analytical form and these are tested by performing plotting comparison of the orbits obtained using the methods. These are available in the [Jupyter](./tests/tests.ipynb) notebook.
+
 
 __Comparing the methods for the coupled quartic Hamiltonian__
 
