### 2019-12-17 c0ef071
update readme with usage details
