### 2019-12-20 fc86da6
update docstrings for coupled quartic Hamiltonian
