+++ examples/coupled_quartic_hamiltonian.py
@@ -16,95 +16,208 @@ from scipy import optimize
 
 
 #% Begin problem specific functions
-def init_guess_eqpt_coupled(eqNum, par):
+def init_guess_eqpt_coupled(eqNum, parameters):
     """
-    Returns configuration space coordinates of the equilibrium points according to the index:
-    Saddle (EQNUM=1)
-    Centre (EQNUM=2,3)
+    Returns guess for solving configuration space coordinates of the equilibrium points.  
+
+    For numerical solution of the equilibrium points, this function returns the guess that be inferred from the potential energy surface. 
+
+    Parameters
+    ----------
+    eqNum : int
+        = 1 for saddle and = 2,3 for centre equilibrium points
+
+    parameters : float (list)
+        model parameters
+
+    Returns
+    -------
+    x0 : float (list of size 2)
+        configuration space coordinates of the guess: [x, y] 
+
     """
     
     if eqNum == 1:
         x0 = [0, 0]
     elif eqNum == 2:
-        x0 = [np.sqrt(par[3]-par[6]/par[4]),0] 
+        x0 = [np.sqrt(parameters[3]-parameters[6]/parameters[4]),0] 
     elif eqNum == 3:
-        x0 = [-np.sqrt(par[3]-par[6]/par[4]),0] 
+        x0 = [-np.sqrt(parameters[3]-parameters[6]/parameters[4]),0] 
     
     return x0
 
-def grad_pot_coupled(x, par):
-    """ Returns the gradient of the potential energy function V(x,y) """
+def grad_pot_coupled(x, parameters):
+    """ Returns the negative of the gradient of the potential energy function 
+    
+    Parameters
+    ----------
+    x : float (list of size 2) 
+        configuration space coordinates: [x, y]
+
+    parameters : float (list)
+        model parameters
+
+    Returns
+    -------
+    F : float (list of size 2)
+        configuration space coordinates of the guess: [x, y]
+
+    """
      
-    dVdx = (-par[3]+par[6])*x[0]+par[4]*(x[0])**3-par[6]*x[1]
-    dVdy = (par[5]+par[6])*x[1]-par[6]*x[0]
+    dVdx = (-parameters[3]+parameters[6])*x[0]+parameters[4]*(x[0])**3-parameters[6]*x[1]
+    dVdy = (parameters[5]+parameters[6])*x[1]-parameters[6]*x[0]
     
     F = [-dVdx, -dVdy]
     
     return F
 
-def pot_energy_coupled(x, y, par):
-    """ Returns the potential energy function V(x,y) """
+def pot_energy_coupled(x, y, parameters):
+    """ Returns the potential energy at the configuration space coordinates 
     
-    return -0.5*par[3]*x**2+0.25*par[4]*x**4 +0.5*par[5]*y**2+0.5*par[6]*(x-y)**2
 
+    Parameters
+    ----------
+    x : float
+        configuration space coordinate
 
-#%
-def eigvector_coupled(par):
-    """ Returns the correction factor to the eigenvectors for the linear guess """
+    y : float
+        configuration space coordinate
+
+    parameters : float (list)
+        model parameters
+
+    Returns
+    -------
+    float 
+        potential energy of the configuration
     
-    evaluelamb = np.sqrt(-0.5*(par[3]-par[6]-par[1]*(par[1]+par[6]) - np.sqrt(par[1]**4 + \
-                               2*par[1]**3*par[6] + par[1]**2*(par[6]**2+2*par[3]-2*par[6]) + \
-                               par[1]*( 2*par[6]**2 + 2*par[3]*par[6]) +(par[3]- par[6])**2)))
[…]
