+++ examples/deleonberne_hamiltonian.py
@@ -17,11 +17,25 @@ from scipy import optimize
 
 
 #% Begin problem specific functions
-def init_guess_eqpt_deleonberne(eqNum, par):
+def init_guess_eqpt_deleonberne(eqNum, parameters):
     """
-    Returns configuration space coordinates of the equilibrium points according to the index:
-    Saddle (EQNUM=1)
-    Centre (EQNUM=2,3)
+    Returns guess for solving configuration space coordinates of the equilibrium points.  
+
+    For numerical solution of the equilibrium points, this function returns the guess that be inferred from the potential energy surface. 
+
+    Parameters
+    ----------
+    eqNum : int
+        = 1 for saddle and = 2,3 for centre equilibrium points
+
+    parameters : float (list)
+        model parameters
+
+    Returns
+    -------
+    x0 : float (list of size 2)
+        configuration space coordinates of the guess: [x, y] 
+
     """
     
     if eqNum == 1:
@@ -34,72 +48,170 @@ def init_guess_eqpt_deleonberne(eqNum, par):
     return x0
 
 
-def grad_pot_deleonberne(x, par):
-    """ Returns the gradient of the potential energy function V(x,y) """
+def grad_pot_deleonberne(x, parameters):
+    """ Returns the negative of the gradient of the potential energy function 
+    
+    Parameters
+    ----------
+    x : float (list of size 2) 
+        configuration space coordinates: [x, y]
+
+    parameters : float (list)
+        model parameters
+
+    Returns
+    -------
+    F : float (list of size 2)
+        configuration space coordinates of the guess: [x, y]
+
+    """
      
-    dVdx = -2*par[3]*par[4]*np.exp(-par[4]*x[0])*(np.exp(-par[4]*x[0]) - 1) - \
-        4*par[5]*par[4]*x[1]**2*(x[1]**2 - 1)*np.exp(-par[5]*par[4]*x[0])
-    dVdy = 8*x[1]*(2*x[1]**2 - 1)*np.exp(-par[5]*par[4]*x[0])
+    dVdx = -2*parameters[3]*parameters[4]*np.exp(-parameters[4]*x[0])*(np.exp(-parameters[4]*x[0]) - 1) - \
+        4*parameters[5]*parameters[4]*x[1]**2*(x[1]**2 - 1)*np.exp(-parameters[5]*parameters[4]*x[0])
+    dVdy = 8*x[1]*(2*x[1]**2 - 1)*np.exp(-parameters[5]*parameters[4]*x[0])
     
     F = [-dVdx, -dVdy]
     
     return F
 
-def pot_energy_deleonberne(x, y, par):
-    """ Returns the potential energy function V(x,y) """
+def pot_energy_deleonberne(x, y, parameters):
+    """ Returns the potential energy at the configuration space coordinates 
     
-    return par[3]*( 1 - np.exp(-par[4]*x) )**2 + 4*y**2*(y**2 - 1)*np.exp(-par[5]*par[4]*x) + par[2]
+
+    Parameters
+    ----------
+    x : float
+        configuration space coordinate
+
+    y : float
+        configuration space coordinate
+
+    parameters : float (list)
+        model parameters
+
+    Returns
+    -------
+    float 
+        potential energy of the configuration
+    
+    """
+    
+    return parameters[3]*( 1 - np.exp(-parameters[4]*x) )**2 + 4*y**2*(y**2 - 1)*np.exp(-parameters[5]*parameters[4]*x) + parameters[2]
 
 
-def eigvector_deleonberne(par):
-    """ Returns the correction factor to the eigenvectors for the linear guess """
+def eigvector_deleonberne(parameters):
[…]
