+++ README.Rmd
@@ -13,7 +13,7 @@ output:
   pdf_document:
     fig_height: 6
     fig_width: 6
-    highlight: zenburn
+    highlight: zenburn 
     toc: yes
 biblio-style: apalike
 ---
