### 2019-12-20 fa3070d
update docstrings for uncoupled quartic Hamiltonian
