+++ examples/uncoupled_quartic_hamiltonian.py
@@ -20,6 +20,23 @@ def upo_analytical(total_energy, t, par):
     """
     Returns the analytical solution of the unstable periodic orbit at 
     total energy and discretized at the time points, t.
+
+    Parameters
+    ----------
+    total_energy : float
+        Total energy of the unstable periodic orbit
+
+    t : 1d numpy array
+        vector of time points at which analytical solution is to be evaluated
+
+    parameters : float (list)
+        model parameters
+
+    Returns
+    -------
+        2d numpy array
+        analytical solution evaluated at the time points
+
     """
     
     OMEGA = par[5]
@@ -36,9 +53,23 @@ def upo_analytical(total_energy, t, par):
 
 def init_guess_eqpt_uncoupled(eqNum, par):
     """
-    Returns configuration space coordinates of the equilibrium points according to the index:
-    Saddle (EQNUM=1)
-    Centre (EQNUM=2,3)
+    Returns guess for solving configuration space coordinates of the equilibrium points.  
+
+    For numerical solution of the equilibrium points, this function returns the guess that be inferred from the potential energy surface. 
+
+    Parameters
+    ----------
+    eqNum : int
+        = 1 for saddle and = 2,3 for centre equilibrium points
+
+    parameters : float (list)
+        model parameters
+
+    Returns
+    -------
+    x0 : float (list of size 2)
+        configuration space coordinates of the guess: [x, y] 
+
     """
     
     if eqNum == 1:
@@ -52,7 +83,22 @@ def init_guess_eqpt_uncoupled(eqNum, par):
 
 
 def grad_pot_uncoupled(x,par):
-    """ Returns the gradient of the potential energy function V(x,y) """ 
+    """ Returns the negative of the gradient of the potential energy function 
+    
+    Parameters
+    ----------
+    x : float (list of size 2) 
+        configuration space coordinates: [x, y]
+
+    parameters : float (list)
+        model parameters
+
+    Returns
+    -------
+    F : float (list of size 2)
+        configuration space coordinates of the guess: [x, y]
+
+    """ 
     
     dVdx = -par[3]*x[0]+par[4]*(x[0])**3
     dVdy = par[5]*x[1]
@@ -63,13 +109,47 @@ def grad_pot_uncoupled(x,par):
 
 
 def pot_energy_uncoupled(x, y, par):
-    """ Returns the potential energy function V(x,y) """
+    """ Returns the potential energy at the configuration space coordinates 
+    
+
+    Parameters
+    ----------
+    x : float
+        configuration space coordinate
+
+    y : float
+        configuration space coordinate
+
+    parameters : float (list)
+        model parameters
+
+    Returns
+    -------
+    float 
+        potential energy of the configuration
+    
+    """
     
     return -0.5*par[3]*x**2+0.25*par[4]*x**4 +0.5*par[5]*y**2
 
 
 def eigvector_uncoupled(par):
-    """ Returns the correction factor to the eigenvectors for the linear guess """
+    """ Returns the flag for the correction factor to the eigenvectors for the linear guess of the unstable periodic orbit.
+    
+    Parameters
+    ----------
+    parameters : float (list)
+        model parameters
+
+    Returns
+    -------
+    correcx : 1 or 0 
+        flag to set the x-component of the eigenvector
+
+    correcy : 1 or 0
+        flag to use the y-component of the eigenvector 
+    
+    """
 
     correcx = 1
     correcy = 1
@@ -78,7 +158,27 @@ def eigvector_uncoupled(par):
 
 
 def guess_lin_uncoupled(eqPt, Ax, par):
-    """ Returns an initial guess for the unstable periodic orbit """ 
+    """ Returns an initial guess as list of coordinates in the phase space 
+    
[…]
