### 2019-12-21 #1684 @Shibabrat
@dpsanders @JoshKarpel Can you take a look at the following PRs?

We've updated the documentation and started a PR: https://github.com/WyLyu/UPOsHam/pull/50. This also includes a plain text explanation of how the methods and the systems fit together.

The test branch and it's PR: https://github.com/WyLyu/UPOsHam/pull/49. This includes the unit test for all the methods.

Both these should address most of the pending issues and requirements upon merging. We look forward to your feedback.
