### 2019-12-24 #1684 @Shibabrat
@kthyng We have now updated the documentation (https://uposham.readthedocs.io/en/latest/) and the new tests have also passed (https://github.com/WyLyu/UPOsHam/pull/49). How do we proceed now?
