+++ paper/paper.bib
@@ -38,7 +38,7 @@
 	number = {2},
 	journal = {Chaos: An Interdisciplinary Journal of Nonlinear Science},
 	author = {Koon, Wang Sang and Lo, Martin W. and Marsden, Jerrold E. and Ross, Shane D.},
-	month = jun,
+	month = {jun},
 	year = {2000},
 	pages = {427--469}
 }
@@ -49,7 +49,7 @@
 	number = {5},
 	journal = {Phys. Rev. E},
 	author = {Ross, Shane D. and BozorgMagham, Amir E. and Naik, Shibabrat and Virgin, Lawrence N.},
-	month = nov,
+	month = {nov},
 	year = {2018},
 	pages = {052214}
 }
@@ -60,7 +60,7 @@
 	volume = {47},
 	journal = {Communications in Nonlinear Science and Numerical Simulation},
 	author = {Naik, Shibabrat and Ross, Shane D.},
-	month = jun,
+	month = {jun},
 	year = {2017},
 	pages = {48--70}
 }
