### 2019-12-28 WyLyu/UPOsHam#54 @JoshKarpel
Running setup.py should install dependencies
The dependencies listed in `requirements.txt` should be included in `setup.py` as well, so that running `python setup.py` also installs the dependencies.

You can do this by either manually duplicating them, or by reading in the dependencies from `requirements.txt` (like you do with the license, for example).

(Part of https://github.com/openjournals/joss-reviews/issues/1684)
