### 2019-12-28 WyLyu/UPOsHam#55 @JoshKarpel
Relative paths in compare_* broken
I was testing out #14, but ran into a lot of problems with how relative paths are used in the `compare_*` scripts under `tests/`.

When I run from inside the `tests/` directory, I get
```
$ python compare_methods_coupled.py
Traceback (most recent call last):
  File "compare_methods_coupled.py", line 14, in <module>
    import differential_correction
ModuleNotFoundError: No module named 'differential_correction'
```

When I run from the repository root, I get
```
$ python tests/compare_methods_coupled.py
Loading the periodic orbit family from data file x0_tpcd_deltaE0.1_coupled.dat

Traceback (most recent call last):
  File "tests/compare_methods_coupled.py", line 165, in <module>
    x0podata = np.loadtxt(data_path + po_fam_file)
  File "/home/jtk/.python/envs/uposham/lib/python3.7/site-packages/numpy/lib/npyio.py", line 968, in loadtxt
    fh = np.lib._datasource.open(fname, 'rt', encoding=encoding)
  File "/home/jtk/.python/envs/uposham/lib/python3.7/site-packages/numpy/lib/_datasource.py", line 269, in open
    return ds.open(path, mode, encoding=encoding, newline=newline)
  File "/home/jtk/.python/envs/uposham/lib/python3.7/site-packages/numpy/lib/_datasource.py", line 621, in open
    encoding=encoding, newline=newline)
FileNotFoundError: [Errno 2] No such file or directory: '../data/x0_tpcd_deltaE0.1_coupled.dat'
```

The problem is that `sys.path.append('./src/') ` and `data_path = "../data/"` don't agree.

The correct solution here is two-fold:
- You shouldn't need to do the `sys.path` hack, if you make the package properly installable. i.e., your setup.py should define a module named `uposham`, so that once you've run `setup.py install` you can `import uposham.differential_correction` from anywhere.
- All other paths, everywhere in the repository, should be constructed based on the magic `__file__` variable (see https://stackoverflow.com/questions/9271464/what-does-the-file-variable-mean-do).

I imagine this will impact many of the scripts under `examples/` as well.

(Part of https://github.com/openjournals/joss-reviews/issues/1684)
