### 2019-12-28 WyLyu/UPOsHam#56 @JoshKarpel
Build/dev artifact cleanup
I spotted a few more build/dev artifacts lying around the source tree that I think should be removed and put in `.gitignore` so they don't get put back in.

- `.Rhistory`
- `UPOsHam.egg-info`
- `dist/`
- `tests/plots`

(Part of https://github.com/openjournals/joss-reviews/issues/1684)
