### 2019-12-29 79bc493
update setup for installing uposham
