### 2019-12-31 #1684 @kthyng
@Shibabrat Looks like [#49](https://github.com/WyLyu/UPOsHam/pull/49) is good to go from the reviewers and maybe we can hear on [#50](https://github.com/WyLyu/UPOsHam/pull/50) from @dpsanders.

I see there are a few more small tasks to clean up.  Otherwise, @dpsanders and @JoshKarpel, could you give an overview of what's left to do at this point? I see unchecked boxes but perhaps some can be checked off now? Thanks very much.
