### 2020-01-02 2579310
update paths to files and scripts
