+++ tests/compare_methods_deleonberne.py
@@ -7,151 +7,31 @@ Created on Wed Jul 31 16:53:45 2019
 """
 
 import numpy as np
-import matplotlib.pyplot as plt
+
 from scipy.integrate import solve_ivp
+
+import matplotlib.pyplot as plt
 from mpl_toolkits.mplot3d import Axes3D
-import sys
-sys.path.append('./src/')
-import differential_correction
 import matplotlib as mpl
 from matplotlib import cm
 mpl.rcParams['mathtext.fontset'] = 'cm'
 mpl.rcParams['mathtext.rm'] = 'serif'
 
 
-#% Begin problem specific functions
-def init_guess_eqpt_deleonberne(eqNum, par):
-    """
-    Returns configuration space coordinates of the equilibrium points according to the index:
-    Saddle (EQNUM=1)
-    Centre (EQNUM=2,3)
-    """
-    
-    if eqNum == 1:
-        x0 = [0, 0]
-    elif eqNum == 2:
-        x0 = [0, 1/np.sqrt(2)]  # EQNUM = 2, center-center
-    elif eqNum == 3:
-        x0 = [0, -1/np.sqrt(2)] # EQNUM = 3, center-center
-    
-    return x0
+# This needs testing for installation 
+import uposham.differential_correction as diffcorr
+import sys
+sys.path.insert(0, './examples/')
+import deleonberne_hamiltonian as deleonberne
+# This needs testing for installation
 
-
-def grad_pot_deleonberne(x, par):
-    """Returns the gradient of the potential energy function V(x,y)
-    """ 
-    
-    dVdx = -2*par[3]*par[4]*np.exp(-par[4]*x[0])*(np.exp(-par[4]*x[0]) - 1) - \
-        4*par[5]*par[4]*x[1]**2*(x[1]**2 - 1)*np.exp(-par[5]*par[4]*x[0])
-    dVdy = 8*x[1]*(2*x[1]**2 - 1)*np.exp(-par[5]*par[4]*x[0])
-    
-    F = [-dVdx, -dVdy]
-    
-    return F
-
-def pot_energy_deleonberne(x, y, par):
-    """Returns the potential energy function V(x,y)
-    """
-    
-    return par[3]*( 1 - np.exp(-par[4]*x) )**2 + 4*y**2*(y**2 \
-        - 1)*np.exp(-par[5]*par[4]*x) + par[2]
-
-
-def variational_eqns_deleonberne(t,PHI,par):
-    """    
-    Returns the state transition matrix , PHI(t,t0), where Df(t) is the Jacobian of the 
-    Hamiltonian vector field
-    
-    d PHI(t, t0)
-    ------------ =  Df(t) * PHI(t, t0)
-        dt
-    
-    """
-    
-    phi = PHI[0:16]
-    phimatrix  = np.reshape(PHI[0:16],(4,4))
-    x,y,px,py = PHI[16:20]
-    
-    
-    # The first order derivative of the potential energy.
-    dVdx = - 2*par[3]*par[4]*np.exp(-par[4]*x)*(np.exp(-par[4]*x) - 1) \
-        - 4*par[5]*par[4]*y**2*(y**2 - 1)*np.exp(-par[5]*par[4]*x) 
-    dVdy = 8*y*(2*y**2 - 1)*np.exp(-par[5]*par[4]*x)
-
-    # The second order derivative of the potential energy. 
-    d2Vdx2 = - ( 2*par[3]*par[4]**2*( np.exp(-par[4]*x) \
-                - 2.0*np.exp(-2*par[4]*x) ) \
-                - 4*(par[5]*par[4])**2*x**2*(y**2 \
[…]
