+++ tests/compare_methods_coupled.py
@@ -6,152 +6,49 @@ Created on Wed Jul 31 16:45:58 2019
 """
 
 import numpy as np
-import matplotlib.pyplot as plt
+
 from scipy.integrate import solve_ivp
+
+import matplotlib.pyplot as plt
 from mpl_toolkits.mplot3d import Axes3D
-import sys
-sys.path.append('./src/') 
-import differential_correction
 import matplotlib as mpl
 from matplotlib import cm
 mpl.rcParams['mathtext.fontset'] = 'cm'
 mpl.rcParams['mathtext.rm'] = 'serif'
 
 
-#% Begin problem specific functions
-def init_guess_eqpt_coupled(eqNum, par):
-    """
-    Returns configuration space coordinates of the equilibrium points according to the index:
-    Saddle (EQNUM=1)
-    Centre (EQNUM=2,3)
-    """
-    
-    if eqNum == 1:
-        x0 = [0, 0]
-    elif eqNum == 2:
-        x0 = [np.sqrt(par[3]-par[6]/par[4]),0] 
-    elif eqNum == 3:
-        x0 = [-np.sqrt(par[3]-par[6]/par[4]),0] 
-    
-    return x0
+# This needs testing for installation 
+import uposham.differential_correction as diffcorr
+import sys
+sys.path.insert(0, './examples/')
+import coupled_quartic_hamiltonian as coupled
+# This needs testing for installation
 
-def grad_pot_coupled(x, par):
-    """Returns the gradient of the potential energy function V(x,y)
-    """ 
-    
-    dVdx = (-par[3]+par[6])*x[0]+par[4]*(x[0])**3-par[6]*x[1]
-    dVdy = (par[5]+par[6])*x[1]-par[6]*x[0]
-    
-    F = [-dVdx, -dVdy]
-    
-    return F
-
-def pot_energy_coupled(x, y, par):
-    """Returns the potential energy function V(x,y)
-    """
-    
-    return -0.5*par[3]*x**2+0.25*par[4]*x**4 +0.5*par[5]*y**2+0.5*par[6]*(x-y)**2
-
-
-def variational_eqns_coupled(t,PHI,par):
-    """    
-    Returns the state transition matrix , PHI(t,t0), where Df(t) is the Jacobian of the 
-    Hamiltonian vector field
-    
-    d PHI(t, t0)
-    ------------ =  Df(t) * PHI(t, t0)
-        dt
-    
-    """
-    
-    phi = PHI[0:16]
-    phimatrix  = np.reshape(PHI[0:16],(4,4))
-    x,y,px,py = PHI[16:20]
-    
-    
-    # The first order derivative of the potential energy.
-    dVdx = (-par[3]+par[6])*x+par[4]*x**3-par[6]*y
-    dVdy = (par[5]+par[6])*y-par[6]*x
-
-    # The second order derivative of the potential energy. 
-    d2Vdx2 = -par[3]+par[6]+par[4]*3*x**2
-        
-    d2Vdy2 = par[5]+par[6]
-
-    d2Vdydx = -par[6]
-
-    
-    d2Vdxdy = d2Vdydx    
-
-    Df    = np.array([[  0,     0,    par[0],    0],
-              [0,     0,    0,    par[1]],
-              [-d2Vdx2,  -d2Vdydx,   0,    0],
-              [-d2Vdxdy, -d2Vdy2,    0,    0]])
-
-    
-    phidot = np.matmul(Df, phimatrix) # variational equation
-
-    PHIdot        = np.zeros(20)
-    PHIdot[0:16]  = np.reshape(phidot,(1,16)) 
-    PHIdot[16]    = px/par[0]
-    PHIdot[17]    = py/par[1]
-    PHIdot[18]    = -dVdx 
-    PHIdot[19]    = -dVdy
-    
[…]
