### 2020-01-03 WyLyu/UPOsHam#55 @Shibabrat
I see the problem and here is what I understand.

Using the ```setup.py install```, we can make the methods importable which is congruent with python style. However, testing the methods requires importing the system-specific script (this has the definition of the vector field) from the examples directory and the initial condition in the data directory. So setting the relative paths with respect to the root directory of the package is where the ```__file__``` variable comes into play. I'm taking your suggested approach along with the advice here: https://docs.python-guide.org/writing/structure/#test-suite to fix this.
