### 2020-01-03 #1684 @JoshKarpel
The last major fix that I see is https://github.com/WyLyu/UPOsHam/issues/55. I think I'll be ready to approve once the fix for that is merged.
