### 2020-01-04 5f9dcb8
check tests using installed package and relative paths
