+++ build/lib/src/turning_point_coord_difference.py
@@ -1,393 +1,0 @@
-# -*- coding: utf-8 -*-
-# """
-# Created on Mon Sep  9 17:49:31 2019
-
-# @author: Wenyang Lyu and Shibabrat Naik
-# """
-
-import numpy as np
-import matplotlib.pyplot as plt
-from scipy.integrate import solve_ivp
-import math
-from mpl_toolkits.mplot3d import Axes3D
-from scipy.optimize import fsolve
-from scipy import optimize
-import matplotlib as mpl
-mpl.rcParams['mathtext.fontset'] = 'cm'
-mpl.rcParams['mathtext.rm'] = 'serif'
-
-    
-def get_eq_pts(eqNum, init_guess_eqpt_model, grad_pot_model, par):
-    """
-    Returns configuration space coordinates of the equilibrium points.
-
-    get_eq_pts(eqNum, init_guess_eqpt_model, grad_pot_model, par) solves the
-    coordinates of the equilibrium point for a Hamiltonian of the form kinetic
-    energy (KE) + potential energy (PE).
-
-    Parameters
-    ----------
-    eqNum : int
-        1 is the saddle-center equilibrium point
-        2 or 3 is the center-center equilibrium point
-
-    init_guess_eqpt_model : function name
-        function that returns the initial guess for the equilibrium point
-
-    grad_pot_model : function name
-        function that defines the vector of potential gradient
-
-    par : float (list)
-        model parameters
-
-    Returns
-    -------
-    float (list)
-        configuration space coordinates
-
-    """
-    #fix the equilibrium point numbering convention here and make a
-    #starting guess at the solution
-    x0 = init_guess_eqpt_model(eqNum, par)
-    
-    # F(xEq) = 0 at the equilibrium point, solve using in-built function
-    F = lambda x: grad_pot_model(x, par)
-    
-    eqPt = fsolve(F, x0, fprime = None) # Call solver
-    
-    return eqPt
-
-
-
-
-#%
-def get_total_energy(orbit, pot_energy_model, parameters):
-    """
-    Returns total energy (value of Hamiltonian) of a phase space point on an orbit
-
-    get_total_energy(orbit, pot_energy_model, parameters) returns the total energy for a
-    Hamiltonian of the form KE + PE.
-
-    Parameters
-    ----------
-    orbit : float (list)
-        phase space coordinates (x,y,px,py) of a point on an orbit
-
-    pot_energy_model : function name
-        function that returns the potential energy of Hamiltonian
-
-    parameters : float (list)
-        model parameters
-
-    Returns
-    -------
-    scalar
-        total energy (value of Hamiltonian)
-
-    """
-    x  = orbit[0]
-    y  = orbit[1]
-    px = orbit[2]
-    py = orbit[3]
-    
-    return (1.0/(2*parameters[0]))*(px**2.0) + (1.0/(2*parameters[1]))*(py**2.0) + \
-            pot_energy_model(x, y, parameters)   
-
-
-#%%
-def get_pot_surf_proj(xVec, yVec, pot_energy_model, par):            
-    """
-    Returns projection of the potential energy (PE) surface on the configuration space
-
-    Parameters
-    ----------
-    xVec, yVec : 1d numpy arrays
-        x,y-coordinates that discretizes the x, y domain of the configuration space
-
-    pot_energy_model : function name
-        function that returns the potential energy of Hamiltonian
-
-    parameters : float (list)
-        model parameters
-
-    Returns
-    -------
-    2d numpy array
-        values of the PE
-
-    """
-    
-    resX = np.size(xVec)
-    resY = np.size(xVec)
-    surfProj = np.zeros([resX, resY])
[…]
