+++ tests/compare_methods_deleonberne.py
@@ -23,11 +23,12 @@ import uposham.deleonberne_hamiltonian as deleonberne
 
 import os
 path_to_data = os.path.join(os.path.dirname(__file__), '../data/')
+path_to_saveplot = os.path.join(os.path.dirname(__file__), 'plots/')
 
 #%% Setting up parameters and global variables
 
 save_final_plot = True
-show_final_plot = True
+show_final_plot = False
 
 N = 4 # dimension of phase space
 MASS_A = 8.0 
@@ -156,10 +157,10 @@ if show_final_plot:
     plt.show()
 
 if save_final_plot:  
-    # plt.savefig('./tests/plots/comparison_deleonberne.pdf', format='pdf', \
-    #             bbox_inches='tight')
-    plt.savefig('./tests/plots/comparison_deleonberne.png', dpi = 300, \
-                bbox_inches='tight')
+    # plt.savefig( path_to_saveplot + 'comparison_deleonberne.pdf', \
+    #             format='pdf', bbox_inches='tight')
+    plt.savefig( path_to_saveplot + 'comparison_deleonberne.png', \
+                dpi = 300, bbox_inches='tight')
 
 
 
