### 2020-01-05 e086370
change scripts to load uposham module and test
