+++ tests/test_upo_uncoupled_ham.py
@@ -8,13 +8,24 @@ from scipy.integrate import solve_ivp
 
 import unittest
 
-import sys
-sys.path.insert(0, './src/')
-import differential_correction as diffcorr
-sys.path.insert(0, './examples/')
-import uncoupled_quartic_hamiltonian as uncoupled
-import differential_correction_uncoupled as diff_corr_unc
-import turning_pt_uncoupled
+# import sys
+# sys.path.insert(0, './src/')
+# import differential_correction as diffcorr
+# sys.path.insert(0, './examples/')
+# import uncoupled_quartic_hamiltonian as uncoupled
+# import differential_correction_uncoupled as diff_corr_unc
+# import turning_pt_uncoupled
+
+# Make sure uposham is installed for the following imports to work
+import uposham.differential_correction as diffcorr
+import uposham.uncoupled_quartic_hamiltonian as uncoupled
+import uposham.differential_correction_uncoupled as diff_corr_unc
+import uposham.turning_pt_uncoupled as turning_pt_uncoupled 
+
+import os
+path_to_data = os.path.join(os.path.dirname(__file__), '../data/')
+path_to_saveplot = os.path.join(os.path.dirname(__file__), 'plots/')
+
 
 from scipy.spatial.distance import directed_hausdorff
 
@@ -57,14 +68,13 @@ class TestUnstablePeriodicOrbit(unittest.TestCase):
         linecolors = ['b']
         diff_corr_unc.upo(deltaE_vals, linecolors)
         
-        data_path = "./data/"
         TSPAN = [0, 30] # arbitrary range, just to finish the integration
         RelTol = 3.e-12
         AbsTol = 1.e-12
 
         for deltaE_val in deltaE_vals:
             po_fam_file = "x0_diffcorr_deltaE%s_uncoupled.dat" %(deltaE_val)
-            x0podata = np.loadtxt(data_path + po_fam_file)
+            x0podata = np.loadtxt(path_to_data + po_fam_file)
             x0po_diffcorr = x0podata[0:4]
 
             f = lambda t,x: uncoupled.ham2dof_uncoupled(t,x,parameters)
@@ -92,112 +102,112 @@ class TestUnstablePeriodicOrbit(unittest.TestCase):
             self.assertLessEqual(hausd_dist, 1e-8)
 
 
-    def test_turning_point(self):
-        """Test the solution obtained using turning point"""
+    # def test_turning_point(self):
+    #     """Test the solution obtained using turning point"""
     
-        MASS_A = 1.0
-        MASS_B = 1.0
+    #     MASS_A = 1.0
+    #     MASS_B = 1.0
              
-        SADDLE_ENERGY = 0.0 # Energy of the saddle
-        ALPHA = 1.00
-        BETA = 1.00
-        OMEGA = 1
-        EPSILON = 0.00 # uncoupled
-        parameters = np.array([MASS_A, MASS_B, SADDLE_ENERGY, \
-                            ALPHA, BETA, OMEGA, EPSILON])
+    #     SADDLE_ENERGY = 0.0 # Energy of the saddle
+    #     ALPHA = 1.00
+    #     BETA = 1.00
+    #     OMEGA = 1
+    #     EPSILON = 0.00 # uncoupled
+    #     parameters = np.array([MASS_A, MASS_B, SADDLE_ENERGY, \
+    #                         ALPHA, BETA, OMEGA, EPSILON])
 
 
-        deltaE_vals = [0.1]
-        linecolors = ['b']
-        turning_pt_uncoupled.upo(deltaE_vals, linecolors)
+    #     deltaE_vals = [0.1]
+    #     linecolors = ['b']
+    #     turning_pt_uncoupled.upo(deltaE_vals, linecolors)
         
-        data_path = "./data/"
[…]
