+++ examples/turning_pt_coupled.py
@@ -20,17 +20,19 @@ mpl.rcParams['mathtext.fontset'] = 'cm'
 mpl.rcParams['mathtext.rm'] = 'serif'
 
 # This needs testing for installation 
-import sys
-import importlib
-sys.path.insert(0, './src/')
+# import sys
+# import importlib
+# sys.path.insert(0, './src/')
 
-import turning_point
-importlib.reload(turning_point)
-import turning_point as tp
+# import turning_point
+# importlib.reload(turning_point)
+# import turning_point as tp
 
-import coupled_quartic_hamiltonian
-importlib.reload(coupled_quartic_hamiltonian)
-import coupled_quartic_hamiltonian as coupled
+# import coupled_quartic_hamiltonian
+# importlib.reload(coupled_quartic_hamiltonian)
+# import coupled_quartic_hamiltonian as coupled
+import uposham.turning_point as tp
+import uposham.coupled_quartic_hamiltonian as coupled
 # This needs testing for installation
 
 
