### 2020-01-05 #1684 @Shibabrat
@kthyng @JoshKarpel @dpsanders
the pending PRs https://github.com/WyLyu/UPOsHam/pull/57, https://github.com/WyLyu/UPOsHam/pull/58, and https://github.com/WyLyu/UPOsHam/pull/59 are now ready for review. The major update is the package is now installable as a module and the scripts can load the system-specific functions and method-specific functions using, for example,

```
import uposham.differential_correction as diffcorr
import uposham.deleonberne_hamiltonian as deleonberne
```

As these PRs get approved, I'll merge them and run tests again.
