### 2020-01-05 e7d6343
change scripts to load uposham module and test
