### 2020-01-05 c701fb7
run tests with updated settings for uncoupled system plots
