+++ examples/differential_correction_coupled.py
@@ -26,6 +26,10 @@ mpl.rcParams['mathtext.rm'] = 'serif'
 # import coupled_quartic_hamiltonian as coupled
 import uposham.differential_correction as diffcorr
 import uposham.coupled_quartic_hamiltonian as coupled
+
+import os
+path_to_data = os.path.join(os.path.dirname(__file__), '../data/')
+path_to_saveplot = os.path.join(os.path.dirname(__file__), '../tests/plots/')
 # This needs testing for installation 
 
 #%% Setting up parameters for the method and the Hamiltonian system
@@ -219,9 +223,9 @@ if show_final_plot:
     plt.show()
 
 if save_final_plot:  
-    plt.savefig('./tests/plots/diff_corr_coupled_upos.pdf', format='pdf', \
-                        bbox_inches='tight')
-    plt.savefig('./tests/plots/diff_corr_coupled_upos.png', dpi = 300, \
-                bbox_inches='tight')
+    # plt.savefig(path_to_saveplot + 'diff_corr_coupled_upos.pdf', \
+    #             format='pdf', bbox_inches='tight')
+    plt.savefig(path_to_saveplot + 'diff_corr_coupled_upos.png', \
+                dpi = 300, bbox_inches='tight')
 
 
