+++ tests/test_upo_uncoupled_ham.py
@@ -8,23 +8,15 @@ from scipy.integrate import solve_ivp
 
 import unittest
 
-# import sys
-# sys.path.insert(0, './src/')
-# import differential_correction as diffcorr
-# sys.path.insert(0, './examples/')
-# import uncoupled_quartic_hamiltonian as uncoupled
-# import differential_correction_uncoupled as diff_corr_unc
-# import turning_pt_uncoupled
-
-# Make sure uposham is installed for the following imports to work
-import uposham.differential_correction as diffcorr
 import uposham.uncoupled_quartic_hamiltonian as uncoupled
 import uposham.differential_correction_uncoupled as diff_corr_unc
 import uposham.turning_pt_uncoupled as turning_pt_uncoupled 
+import uposham.turning_pt_coord_diff_uncoupled as tpcd_uncoupled 
+
+import uposham.differential_correction as diffcorr
 
 import os
 path_to_data = os.path.join(os.path.dirname(__file__), '../data/')
-path_to_saveplot = os.path.join(os.path.dirname(__file__), 'plots/')
 
 
 from scipy.spatial.distance import directed_hausdorff
@@ -66,16 +58,21 @@ class TestUnstablePeriodicOrbit(unittest.TestCase):
 
         deltaE_vals = [0.1]
         linecolors = ['b']
-        diff_corr_unc.upo(deltaE_vals, linecolors)
-        
+        save_final_plot = False
+        show_final_plot = True
+        diff_corr_unc.upo(deltaE_vals, linecolors, \
+                            save_final_plot, show_final_plot)
+
         TSPAN = [0, 30] # arbitrary range, just to finish the integration
         RelTol = 3.e-12
         AbsTol = 1.e-12
 
         for deltaE_val in deltaE_vals:
-            po_fam_file = "x0_diffcorr_deltaE%s_uncoupled.dat" %(deltaE_val)
-            x0podata = np.loadtxt(path_to_data + po_fam_file)
-            x0po_diffcorr = x0podata[0:4]
+            # po_fam_file = "x0_diffcorr_deltaE%s_uncoupled.dat" %(deltaE_val)
+            with open(path_to_data + "x0_diffcorr_deltaE%s_uncoupled.dat" %(deltaE_val)) as po_fam_file: 
+                x0podata = np.loadtxt(po_fam_file.name)
+                x0po_diffcorr = x0podata[0:4]
+                # print(x0po_diffcorr)
 
             f = lambda t,x: uncoupled.ham2dof_uncoupled(t,x,parameters)
             soln = solve_ivp(f, TSPAN, x0po_diffcorr,method='RK45',dense_output=True, \
@@ -102,112 +99,117 @@ class TestUnstablePeriodicOrbit(unittest.TestCase):
             self.assertLessEqual(hausd_dist, 1e-8)
 
 
-    # def test_turning_point(self):
-    #     """Test the solution obtained using turning point"""
+    def test_turning_point(self):
+        """Test the solution obtained using turning point"""
     
-    #     MASS_A = 1.0
-    #     MASS_B = 1.0
+        MASS_A = 1.0
+        MASS_B = 1.0
              
-    #     SADDLE_ENERGY = 0.0 # Energy of the saddle
-    #     ALPHA = 1.00
-    #     BETA = 1.00
-    #     OMEGA = 1
-    #     EPSILON = 0.00 # uncoupled
-    #     parameters = np.array([MASS_A, MASS_B, SADDLE_ENERGY, \
-    #                         ALPHA, BETA, OMEGA, EPSILON])
+        SADDLE_ENERGY = 0.0 # Energy of the saddle
+        ALPHA = 1.00
+        BETA = 1.00
+        OMEGA = 1
+        EPSILON = 0.00 # uncoupled
[…]
