### 2020-01-05 16cab89
fix reading requirements file
