### 2020-01-05 5ee3b9f
update paths, remove obsolete and commented code
