+++ examples/differential_correction_deleonberne.py
@@ -13,25 +13,14 @@ import matplotlib.pyplot as plt
 from scipy.integrate import solve_ivp
 import time
 
-# This needs testing for installation 
-# import sys
-# import importlib
-# sys.path.insert(0, './src/')
-
-# import differential_correction
-# importlib.reload(differential_correction)
-# import differential_correction as diffcorr
-
-# import deleonberne_hamiltonian
-# importlib.reload(deleonberne_hamiltonian)
-# import deleonberne_hamiltonian as deleonberne
 import uposham.differential_correction as diffcorr
 import uposham.deleonberne_hamiltonian as deleonberne
 
 import os
-path_to_data = os.path.join(os.path.dirname(__file__), '../data/')
-path_to_saveplot = os.path.join(os.path.dirname(__file__), '../tests/plots/')
-# This needs testing for installation
+path_to_data = os.path.join(os.path.dirname(os.path.dirname(__file__)), \
+                            'data/')
+path_to_saveplot = os.path.join(os.path.dirname(os.path.dirname(__file__)), \
+                                'tests/plots/')
 
 import matplotlib as mpl
 from matplotlib import cm
