### 2020-01-06 64f3cc2
minor changes to readme
