### 2020-01-06 25c8120
minor changes to readme
