+++ README.Rmd
@@ -41,6 +41,12 @@ $ python setup.py install
 
 and check the modules shown in [requirements.txt](https://github.com/WyLyu/UPOsHam/tree/master/requirements.txt) are installed using conda/pip. 
 
+This makes ```uposham``` available for import along with the methods and example systems, for example
+
+```
+import uposham.differential_correction as diffcorr
+import uposham.deleonberne_hamiltonian as deleonberne
+```
 
 ## Usage
 
