### 2020-01-06 8b11e4d
add plot and minor change
