+++ README.Rmd
@@ -41,9 +41,9 @@ $ python setup.py install
 
 and check the modules shown in [requirements.txt](https://github.com/WyLyu/UPOsHam/tree/master/requirements.txt) are installed using conda/pip. 
 
-This makes ```uposham``` available for import along with the methods and example systems, for example
+Then ```uposham``` is available for import along with the methods and example systems, for example
 
-```
+```python
 import uposham.differential_correction as diffcorr
 import uposham.deleonberne_hamiltonian as deleonberne
 ```
