+++ paper/paper.bib
@@ -40,7 +40,8 @@
 	author = {Koon, Wang Sang and Lo, Martin W. and Marsden, Jerrold E. and Ross, Shane D.},
 	month = {jun},
 	year = {2000},
-	pages = {427--469}
+	pages = {427--469},
+	doi = {10.1063/1.166509}
 }
 
 @article{ross_experimental_2018,
@@ -51,7 +52,8 @@
 	author = {Ross, Shane D. and BozorgMagham, Amir E. and Naik, Shibabrat and Virgin, Lawrence N.},
 	month = {nov},
 	year = {2018},
-	pages = {052214}
+	pages = {052214},
+	doi = {10.1103/PhysRevE.98.052214}
 }
 
 
@@ -62,7 +64,8 @@
 	author = {Naik, Shibabrat and Ross, Shane D.},
 	month = {jun},
 	year = {2017},
-	pages = {48--70}
+	pages = {48--70},
+	doi = {10.1016/j.cnsns.2016.10.021}
 }
 
 @article{naik_finding_2019b,
@@ -75,8 +78,8 @@
 	numpages = {14},
 	year = {2019},
 	month = {Aug},
-	doi = {10.1103/PhysRevE.100.022204}}
-
+	doi = {10.1103/PhysRevE.100.022204}
+}
 
 @article{wiggins_role_2016,
 	title = {The role of normally hyperbolic invariant manifolds ({NHIMs}) in the context of the phase space setting for chemical reaction dynamics},
@@ -87,8 +90,8 @@
 	month = {Nov},
 	year = {2016},
 	pages = {621--638},
-	doi = {10.1134/S1560354716060034}}
-
+	doi = {10.1134/S1560354716060034}
+}
 
 @article{Pollak_1980,
 	title = {Classical transition state theory: {A} lower bound to the reaction probability},
@@ -98,8 +101,8 @@
 	author = {Pollak, Eli and Child, Mark S. and Pechukas, Philip},
 	year = {1980},
 	pages = {1669--1678},
-	doi = {10.1063/1.439276}}
-
+	doi = {10.1063/1.439276}
+}
 
 @article{Deleon_Berne_1981,
 	title = {Intramolecular rate process: Isomerization dynamics and the transition to chaos},
@@ -109,8 +112,8 @@
 	author = {De Leon, Nelson  and Berne, B. J.},
 	year = {1981},
 	pages = {3495-3510},
-	doi = {10.1063/1.442459}}
-
+	doi = {10.1063/1.442459}
+}
 
 @article{DeLeon_Marston_1989,
 	author = {De Leon, N  and Marston, C. Clay },
@@ -120,17 +123,18 @@
 	number = {6},
 	pages = {3405-3425},
 	year = {1989},
-	doi = {10.1063/1.456915}}
+	doi = {10.1063/1.456915}
+}
 
 @article{marston_reactive_1989,
 	title = {Reactive islands as essential mediators of unimolecular conformational isomerization: {A} dynamical study of 3‐phospholene},
 	volume = {91},
-	doi = {10.1063/1.456914},
 	number = {6},
 	journal = {The Journal of Chemical Physics},
 	author = {Marston, C. Clay and De Leon, N.},
 	month = sep,
 	year = {1989},
-	pages = {3392--3404}
+	pages = {3392--3404},
+	doi = {10.1063/1.456914}
 }
 
