+++ paper/paper.bib
@@ -8,7 +8,9 @@
 	number = {2},
 	publisher = {Springer},
 	author = {Wiggins, Stephen},
-	year = {2003}}
+	year = {2003},
+	url = {ISBN: 978-0-387-00177-7}
+}
 
 
 @book{Koon2011,
@@ -20,7 +22,9 @@
 	isbn = {978-0-615-24095-4},
 	keywords = {3BP,Dynamical Systems,Space Mission Design},
 	pages = {327},
-	year = {2011}}
+	year = {2011},
+	url = {ISBN: 978-0-615-24095-4}
+}
 
 
 @book{Parker_1989,
@@ -29,7 +33,9 @@
 	year = {1989},
 	isbn = {0-387-96689-7},
 	publisher = {Springer-Verlag New York, Inc.},
-	address = {New York, NY, USA}} 
+	address = {New York, NY, USA},
+	url = {ISBN: 0-387-96689-7}
+} 
 
 
 @article{koon_heteroclinic_2000,
