+++ README.md
@@ -212,7 +212,7 @@ Journal of Nonlinear Science* 10 (2): 427–69.
 
 Koon, W. S., M. W. Lo, J. E. Marsden, and S. D. Ross. 2011. *Dynamical
 systems, the three-body problem and space mission design*. Marsden
-books.
+books. [ISBN: 978-0-615-24095-4](ISBN: 978-0-615-24095-4).
 
 Naik, Shibabrat, and Shane D. Ross. 2017. “Geometry of Escaping Dynamics
 in Nonlinear Ship Motion.” *Communications in Nonlinear Science and
@@ -226,6 +226,7 @@ Hénon-Heiles-Type Potential.” *Phys. Rev. E* 100 (2): 022204.
 
 Parker, T. S., and L. O. Chua. 1989. *Practical Numerical Algorithms for
 Chaotic Systems*. New York, NY, USA: Springer-Verlag New York, Inc.
+[ISBN: 0-387-96689-7](ISBN: 0-387-96689-7).
 
 Pollak, Eli, Mark S. Child, and Philip Pechukas. 1980. “Classical
 Transition State Theory: A Lower Bound to the Reaction Probability.”
@@ -239,7 +240,7 @@ Transition Between Potential Wells.” *Phys. Rev. E* 98 (5): 052214.
 
 Wiggins, Stephen. 2003. *Introduction to Applied Nonlinear Dynamical
 Systems and Chaos*. 2nd ed. Texts in Applied Mathematics 2. New York:
-Springer.
+Springer. [ISBN: 978-0-387-00177-7](ISBN: 978-0-387-00177-7).
 
 ———. 2016. “The Role of Normally Hyperbolic Invariant Manifolds (NHIMs)
 in the Context of the Phase Space Setting for Chemical Reaction
