+++ README.Rmd
@@ -48,7 +48,7 @@ import uposham.deleonberne_hamiltonian as deleonberne
 To be sure check the modules shown in [requirements.txt](https://github.com/WyLyu/UPOsHam/tree/master/requirements.txt) are installed using
 
 ```bash
-pip install -r requirements.txt (or pip3 install -r requirements.txt)
+$ pip install -r requirements.txt (or pip3 install -r requirements.txt)
 ```
 
 
@@ -88,6 +88,7 @@ This will save data files in the root directory and save the plot in the [./test
 
 
 ![](tests/plots/diff_corr_coupled_upos.png){width="50.00000%"}
+__Fig. 2. Unstable periodic orbits for the coupled quartic Hamiltonian__
 
 
 ## Tests
