### 2020-01-06 9dd80f6
correct minor typos
