+++ examples/differential_correction_uncoupled.py
@@ -26,7 +26,7 @@ import os
 path_to_data = os.path.join(os.path.dirname(os.path.dirname(__file__)), \
                             'data/')
 path_to_saveplot = os.path.join(os.path.dirname(os.path.dirname(__file__)), \
-                                'tests/plots/'))
+                                'tests/plots/')
 
 N = 4          # dimension of phase space
 MASS_A = 1.00
