+++ tests/test_upo_uncoupled_ham.py
@@ -19,204 +19,204 @@ import os
 path_to_data = os.path.join(os.path.dirname(os.path.dirname(__file__)), \
                             'data/')
 
-# from scipy.spatial.distance import directed_hausdorff
+from scipy.spatial.distance import directed_hausdorff
 
-# def hausd_dist_numeric_analytic(orbit, t, deltaE_val, parameters):
-#     """
-#     Obtain Hausdorff distance between the unstable periodic orbits obtained
-#     using a numerical method and analytical solution for the uncoupled 
-#     quartic Hamiltonian
-#     """
+def hausd_dist_numeric_analytic(orbit, t, deltaE_val, parameters):
+    """
+    Obtain Hausdorff distance between the unstable periodic orbits obtained
+    using a numerical method and analytical solution for the uncoupled 
+    quartic Hamiltonian
+    """
 
-#     total_energy = deltaE_val + parameters[2]
-#     y, py = uncoupled.upo_analytical(total_energy, t, parameters)
+    total_energy = deltaE_val + parameters[2]
+    y, py = uncoupled.upo_analytical(total_energy, t, parameters)
 
-#     numerical_orbit = np.array([orbit[:,1], orbit[:,3]])
-#     analytical_orbit = np.array([y, py])
+    numerical_orbit = np.array([orbit[:,1], orbit[:,3]])
+    analytical_orbit = np.array([y, py])
 
-#     hausd_dist = directed_hausdorff(numerical_orbit, analytical_orbit)[0]
+    hausd_dist = directed_hausdorff(numerical_orbit, analytical_orbit)[0]
 
-#     return hausd_dist
+    return hausd_dist
 
-# class TestUnstablePeriodicOrbit(unittest.TestCase):
-#     """ Test unstable periodic orbit for the uncoupled system """
+class TestUnstablePeriodicOrbit(unittest.TestCase):
+    """ Test unstable periodic orbit for the uncoupled system """
 
-#     def test_differential_correction(self):
-#         """ Test solution obtained using differential correction """
+    def test_differential_correction(self):
+        """ Test solution obtained using differential correction """
     
-#         MASS_A = 1.0
-#         MASS_B = 1.0
+        MASS_A = 1.0
+        MASS_B = 1.0
              
-#         SADDLE_ENERGY = 0.0 # Energy of the saddle
-#         ALPHA = 1.00
-#         BETA = 1.00
-#         OMEGA = 1
-#         EPSILON = 0.00 # uncoupled
-#         parameters = np.array([MASS_A, MASS_B, SADDLE_ENERGY, \
-#                             ALPHA, BETA, OMEGA, EPSILON])
+        SADDLE_ENERGY = 0.0 # Energy of the saddle
+        ALPHA = 1.00
+        BETA = 1.00
+        OMEGA = 1
+        EPSILON = 0.00 # uncoupled
+        parameters = np.array([MASS_A, MASS_B, SADDLE_ENERGY, \
+                            ALPHA, BETA, OMEGA, EPSILON])
 
 
-#         deltaE_vals = [0.1]
-#         linecolors = ['b']
-#         save_final_plot = False
-#         show_final_plot = True
-#         diff_corr_unc.upo(deltaE_vals, linecolors, \
-#                             save_final_plot, show_final_plot)
+        deltaE_vals = [0.1]
+        linecolors = ['b']
+        save_final_plot = False
+        show_final_plot = True
+        diff_corr_unc.upo(deltaE_vals, linecolors, \
+                            save_final_plot, show_final_plot)
 
-#         TSPAN = [0, 30] # arbitrary range, just to finish the integration
-#         RelTol = 3.e-12
[…]
