### 2020-01-06 #1684 @JoshKarpel
I'm happy with the changes, have checked off all my boxes, and recommend acceptance.
