### 2020-01-06 #1684 @Shibabrat
@dpsanders I'll go through this version to check for typos and update the pdf. Do you want us to address any of the 3 remaining things in your checklist?

@JoshKarpel thanks for the approval.
