+++ paper/paper.bib
@@ -38,6 +38,18 @@
 } 
 
 
+@article{farantos_pomult_1998,
+	title={POMULT: A program for computing periodic orbits in Hamiltonian systems based on multiple shooting algorithms},
+	author={Farantos, Stavros C},
+	journal={Computer physics communications},
+	volume={108},
+	number={2-3},
+	pages={240--258},
+	year={1998},
+	publisher={Elsevier},
+	doi = {10.1016/S0010-4655(97)00131-8}
+}
+
 @article{koon_heteroclinic_2000,
 	title = {Heteroclinic connections between periodic orbits and resonance transitions in celestial mechanics},
 	volume = {10},
@@ -115,14 +127,14 @@
 	volume = {75},
 	number = {7},
 	journal = {The Journal of Chemical Physics},
-	author = {De Leon, Nelson  and Berne, B. J.},
+	author = {De Leon, N.  and Berne, B. J.},
 	year = {1981},
 	pages = {3495-3510},
 	doi = {10.1063/1.442459}
 }
 
 @article{DeLeon_Marston_1989,
-	author = {De Leon, N  and Marston, C. Clay },
+	author = {De Leon, N.  and Marston, C. C. },
 	title = {Order in chaos and the dynamics and kinetics of unimolecular conformational isomerization},
 	journal = {The Journal of Chemical Physics},
 	volume = {91},
@@ -137,7 +149,7 @@
 	volume = {91},
 	number = {6},
 	journal = {The Journal of Chemical Physics},
-	author = {Marston, C. Clay and De Leon, N.},
+	author = {Marston, C. C. and De Leon, N.},
 	month = sep,
 	year = {1989},
 	pages = {3392--3404},
