+++ paper/paper.md
@@ -30,36 +30,36 @@ tags:
 - Turning point
 affiliations:
 - index: 1
-  name: School of Mathematics, University Walk, University of Bristol, Clifton BS8 1TW, Bristol, United Kingdom
+  name: School of Mathematics, University of Bristol, Fry Building, Woodland Road, Bristol BS8 1UG, United Kingdom
 ---
 
 ## Statement of Need
 
-In Hamiltonian systems the fundamental phase space structure that partitions dynamically disparate trajectories and mediates transition between multi-stable regions is an invariant manifold. In a 2N dimensional Hamiltonian phase space, the invariant manifold has 2 less dimension than the phase space and is anchored to the normally hyperbolic invariant manifold which has 3 less dimension. This becomes an unstable periodic orbit (UPO) for 2 degrees of freedom or four dimensional phase space [@wiggins_role_2016]. Since the UPO forms the basis for partitioning trajectories, hence their computation and stability analysis is the starting point for dynamical systems analysis. UPOsHam is meant to serve this purpose by providing examples of how to implement numerical methods for computing the unstable periodic orbits (UPOs) at any specified total energy as long as their existence is guaranteed. Even though, there is no lack of numerical methods for computing UPOs, we have found that they either lack in reproducibility, or have steep learning curve for using the software, or have been written using closed source software, and at times combination of these. Our aim is to provide an open source package that implements some of the standard methods and shows the results in the context of model problems. This is meant as a starting point to integrate other numerical methods in an open source package such that UPOs computed in dynamical systems papers can be reproduced with minimal tweaking while providing an exploratory environment to learn and develop the underlying methods.  
-
+Hamiltonian systems model a diverse array of problems in physical sciences and engineering. 
+In Hamiltonian systems the fundamental phase space structure that partitions dynamically distinct trajectories and mediates transition between multi-stable regions are __stable__ and __unstable invariant manifold__. In a 2N dimensional Hamiltonian phase space, the invariant manifold has 2 less dimension than the phase space and is anchored to the __normally hyperbolic invariant manifold__ (NHIM) which has 3 less dimension. For 2 degrees of freedom or four dimensional phase space, the NHIM is an __unstable periodic orbit__ (UPO) around the index-1 saddle equilibrium point [@wiggins_role_2016]. Since in this case, the UPO anchors the invariant manifolds that partition trajectories, their computation and stability analysis forms the starting point for dynamical systems analysis. UPOsHam is meant to serve this purpose by providing a module of numerical methods, along with example Hamiltonian systems, for computing the unstable periodic orbits (UPOs) at any specified total energy as long as their existence is guaranteed. Even though, there are existing numerical methods for computing UPOs, we have found that they either lack in reproducibility, or have steep learning curve for using the software, or have been written using closed source software, and at times combination of these [@farantos_pomult_1998;@Pollak_1980]. Our aim is to provide an open source package that implements some of the standard methods and shows the results in the context of example Hamiltonian systems. This is meant as a starting point to integrate other numerical methods in an open source package such that UPOs computed in dynamical systems papers can be reproduced with minimal tweaking while providing an exploratory environment to further develop the underlying methods.  
 
 
 ## Summary
 
[…]
