+++ paper/paper.md
@@ -36,7 +36,7 @@ affiliations:
 ## Statement of Need
 
 Hamiltonian systems model a diverse array of problems in physical sciences and engineering. 
-In Hamiltonian systems the fundamental phase space structure that partitions dynamically distinct trajectories and mediates transition between multi-stable regions are __stable__ and __unstable invariant manifold__. In a 2N dimensional Hamiltonian phase space, the invariant manifold has 2 less dimension than the phase space and is anchored to the __normally hyperbolic invariant manifold__ (NHIM) which has 3 less dimension. For 2 degrees of freedom or four dimensional phase space, the NHIM is an __unstable periodic orbit__ (UPO) around the index-1 saddle equilibrium point [@wiggins_role_2016]. Since in this case, the UPO anchors the invariant manifolds that partition trajectories, their computation and stability analysis forms the starting point for dynamical systems analysis. UPOsHam is meant to serve this purpose by providing a module of numerical methods, along with example Hamiltonian systems, for computing the unstable periodic orbits (UPOs) at any specified total energy as long as their existence is guaranteed. Even though, there are existing numerical methods for computing UPOs, we have found that they either lack in reproducibility, or have steep learning curve for using the software, or have been written using closed source software, and at times combination of these [@farantos_pomult_1998;@Pollak_1980]. Our aim is to provide an open source package that implements some of the standard methods and shows the results in the context of example Hamiltonian systems. This is meant as a starting point to integrate other numerical methods in an open source package such that UPOs computed in dynamical systems papers can be reproduced with minimal tweaking while providing an exploratory environment to further develop the underlying methods.  
+In Hamiltonian systems the fundamental phase space structure that partitions dynamically distinct trajectories and mediates transition between multi-stable regions are *stable* and *unstable invariant manifold*. In a 2N dimensional Hamiltonian phase space, the invariant manifold has 2 less dimension than the phase space and is anchored to the *normally hyperbolic invariant manifold* (NHIM) which has 3 less dimension. For 2 degrees of freedom or four dimensional phase space, the NHIM is an *unstable periodic orbit* (UPO) around the index-1 saddle equilibrium point [@wiggins_role_2016]. Since in this case, the UPO anchors the invariant manifolds that partition trajectories, their computation and stability analysis forms the starting point for dynamical systems analysis. UPOsHam is meant to serve this purpose by providing a module of numerical methods, along with example Hamiltonian systems, for computing the unstable periodic orbits (UPOs) at any specified total energy as long as their existence is guaranteed. Even though, there are existing numerical methods for computing UPOs, we have found that they either lack in reproducibility, or have steep learning curve for using the software, or have been written using closed source software, and at times combination of these [@farantos_pomult_1998;@Pollak_1980]. Our aim is to provide an open source package that implements some of the standard methods and shows the results in the context of example Hamiltonian systems. This is meant as a starting point to integrate other numerical methods in an open source package such that UPOs computed in dynamical systems papers can be reproduced with minimal tweaking while providing an exploratory environment to further develop the underlying methods.  
 
 
 ## Summary
@@ -55,7 +55,7 @@ This method is based on finding the UPO by detecting trajectories initialized on
 
 __Turning point based on configuration difference__
 
[…]
