### 2020-01-07 #1684 @dpsanders
There are a few typos and English issues (mainly singular / plural) in the paper that should be checked.

I think somewhere it needs to be made clearer exactly what is shown in the figures -- presumably contours of the Hamiltonian (which is not stated anywhere as far as I can see) and the periodic orbits computed by the code.

Other than that everything looks good and I recommend acceptance.
