### 2020-01-08 #1684 @labarba
Hi @kthyng ... it looks fine to me, but perhaps a verb change would help. Also, add hyphens to the compound modifiers:

"In two-degree-of-freedom systems, the phase space is four dimensional and the dynamics takes place on the three-dimensional energy surface…"
