### 2020-01-08 #1684 @kthyng
Also, you go between using and not using your acronym UPO frequently through the paper — seems a little odd but is ok.

I've made more small grammar changes in PR [#60](https://github.com/WyLyu/UPOsHam/pull/60) that you can merge if you agree with the changes.
