### 2020-01-08 #1684 @Shibabrat
@kthyng http://doi.org/10.5281/zenodo.3601504 which I've tagged as v0.1.0 release. Please note that at the moment doi on the readme links to the JOSS submitted version.
