### 2020-01-09 d04e6ee
minor changes to paper
