### 2020-01-09 080d2c0
update DOI in the readme and build reference
