### 2020-01-09 dd3f3dc
update DOI in the readme and build reference
