+++ paper/paper.bib
@@ -23,10 +23,10 @@
 	keywords = {3BP,Dynamical Systems,Space Mission Design},
 	pages = {327},
 	year = {2011},
-	doi = {10.1142/9789812792617_0222}
+	url = {www.shaneross.com/books/}
 }
 
-
+ISBN 978-0-615-24095-4
 @book{Parker_1989,
 	author = {Parker, T. S. and Chua, L. O.},
 	title = {Practical Numerical Algorithms for Chaotic Systems},
