### 2020-01-09 389de5b
update DOI in the readme and build reference
