+++ paper/paper.bib
@@ -9,7 +9,7 @@
 	publisher = {Springer},
 	author = {Wiggins, Stephen},
 	year = {2003},
-	url = {ISBN: 978-0-387-00177-7}
+	url = {10.1007/b97481}
 }
 
 
@@ -19,14 +19,13 @@
 	publisher = {Marsden books},
 	title = {{Dynamical systems, the three-body problem and space mission 
 	design}},
-	isbn = {978-0-615-24095-4},
 	keywords = {3BP,Dynamical Systems,Space Mission Design},
 	pages = {327},
 	year = {2011},
-	url = {www.shaneross.com/books/}
+	isbn = {978-0-615-24095-4}
 }
 
-ISBN 978-0-615-24095-4
+
 @book{Parker_1989,
 	author = {Parker, T. S. and Chua, L. O.},
 	title = {Practical Numerical Algorithms for Chaotic Systems},
