### 2020-01-09 3849c01
update DOI in the readme and build reference
