+++ paper/paper.bib
@@ -9,7 +9,7 @@
 	publisher = {Springer},
 	author = {Wiggins, Stephen},
 	year = {2003},
-	url = {10.1007/b97481}
+	url = {ISBN: 978-0-387-00177-7}
 }
 
 
@@ -19,10 +19,11 @@
 	publisher = {Marsden books},
 	title = {{Dynamical systems, the three-body problem and space mission 
 	design}},
+	isbn = {978-0-615-24095-4},
 	keywords = {3BP,Dynamical Systems,Space Mission Design},
 	pages = {327},
 	year = {2011},
-	isbn = {978-0-615-24095-4}
+	doi = {10.1142/9789812792617_0222}
 }
 
 
