### 2020-01-13 #1684 @kthyng
> so should I change all occurrences of degree of freedom to degree-of-freedom?

I think so, yes.
