### 2020-01-13 #1684 @kthyng
> > so should I change all occurrences of degree of freedom to degree-of-freedom?
>
> I think so, yes.

@Shibabrat Sorry I should add that the "two-" at the beginning is important. Without that, it is two of the systems, not specifying the degree of the systems.
