### 2020-01-13 #1684 @kthyng
@Shibabrat I think something that you changed in your last commit altered the ability of the paper to be processed correctly. Maybe it's the line breaks, like in the title?
