### 2020-01-13 fb5233f
minor changes to resolve whedon processing
