### 2020-01-13 #1684 @Shibabrat
@kthyng not sure what happened but I made sure the file looks markdown and pdf file looks ok on my end. Is whedon using the zenodo link to build the pdf for the paper?
