### 2020-01-13 78c4149
minor changes to resolve whedon processing
