### 2020-01-14 #1684 @arfon
@Shibabrat the errors is right there:

```
Paper YAML header is missing expected fields: date
```
