+++ paper/paper.md
@@ -10,7 +10,7 @@ authors:
 - affiliation: 1
   name: Stephen Wiggins
   orcid: 0000-0002-5036-5863
-date: '\today'
+date: 8 January 2020
 bibliography: paper.bib
 output:
   pdf_document:
