### 2020-01-14 #1684 @Shibabrat
Thank you @JoshKarpel @dpsanders @kthyng and @labarba for the guidance, patience, and help along the way. Looking forward to contributing to the community of reproducible science and open-source software.
